### 2019-08-23 #335 @florianm
Submission: ruODK, Client for the ODK Central API
Submitting Author: Florian Mayer (@florianm)
Repository:  https://github.com/dbca-wa/ruODK
Version submitted:   0.6.6.9025 (Jun 2020) (originally 0.6.0)
Archive: 2020-02-11
Version accepted: TBD

---

```
Package: ruODK
Title: An R Client for the ODK Central API
Version: 0.6.6.9025
    c(person(given = c("Florian", "W."),
             family = "Mayer",
             email = "Florian.Mayer@dbca.wa.gov.au",
      person(given = "Maëlle",
             family = "Salmon",
             role = "rev",
             email = "maelle.salmon@yahoo.se",
      person(given = "Karissa",
             family = "Whiting",
             role = "rev",
      person(given = "Jason",
             family = "Taylor",
             role = "rev"),
      person(given = "DBCA",
             role = c("cph", "fnd")),
      person(given = "NWSFTCP",
             role = "fnd"))
Description: Utilities to access and tidy up data from ODK
    Central's API.  ODK Central is OpenDataKit's clearinghouse for
    digitally captured data <https://docs.opendatakit.org/central-intro/>.
    ODK Central's API is documented at
    <https://odkcentral.docs.apiary.io/>.
    https://github.com/dbca-wa/ruODK
    clisymbols (>= 1.2.0),
    crayon (>= 1.3.4),
    fs (>= 1.4.1),
    glue (>= 1.4.0),
    janitor (>= 2.0.1),
    lifecycle (>= 0.1.0),
    lubridate (>= 1.7.8),
    magrittr (>= 1.5),
    purrr (>= 0.3.4),
    readr (>= 1.3.1),
    rlang (>= 0.4.5),
    rlist (>= 0.4.6.1),
    stringr (>= 1.4.0),
    tibble (>= 2.1.3),
    tidyr (>= 1.0.3),
    tidyselect (>= 1.0.0),
    xml2 (>= 1.2.2)
    covr (>= 3.4.0),
    DT (>= 0.9),
    ggplot2 (>= 3.2.1),
    knitr (>= 1.26),
    leaflet (>= 2.0.3),
    listviewer (>= 3.0.0),
    rmarkdown (>= 1.17),
    roxygen2 (>= 7.1.0),
    usethis (>= 1.6.0),
    vcr (>= 0.5.4)
    knitr
RdMacros:
    lifecycle
Language: en_AU
X-schema.org-applicationCategory: Data Access
X-schema.org-keywords: database, open-data, opendatakit, odk, api, data, dataset
```

Data retrieval: ruODK retrieves data from ODK Central, a data clearinghouse containing data which have been digitally captured by the data collection app ODK Collect using Xforms.

As mentioned by @annakrystalli, the below categories are touched by ruODK, but don't apply:

Data extraction or munging: ruODK transforms and sanitises the data from ODK Central from the original format (which parses to nested lists in R) into tibbles.

Reproducibility: ruODK allows to script and repeat the data extraction step - the main use case it is being written for.

Geospatial data: while ODK allows to capture location data (points, lines, polygons), and ruODK extracts these values, ruODK is not primarily a spatial package.

(Scope and use cases are also mentioned in the [README](https://dbca-wa.github.io/ruODK/index.html).)

Any organisation collecting data with the [OpenDataKit](https://opendatakit.org/) suite will need to extract the data from the data clearinghouses, ODK Aggregate (outgoing) or ODK Central (new). Some may want to analyse the data straight out of ODK Central, some may need to transfer the data into another data warehouse for further post-processing, QA, and integration with other data sources.

For both use cases, `ruODK` bridges and simplifies the gap between the data sitting in ODK Central, and the data being a tibble in R, ready for further processing.

In a nutshell, `ruODK` aims to be to ODK Central what `ckanr` is to CKAN.

As per [release notes for ODK Central 0.6](https://forum.opendatakit.org/t/odk-central-beta-v0-6/21489), next to the other options, ruODK is now the recommended package to access ODK Central data in R.

https://github.com/ropensci/software-review/issues/328
Thanks to @annakrystalli for feedback on the pre-submission.
Changes since then: added remaining functions.

Note: I would like to submit a paper about the package in a few weeks, but haven't got the manuscript ready and approved for publication just yet.

- [X] Do you intend for this package to go on CRAN?  > if you see fit, yes please (0.6.6)

Update 31 Aug: pasted new DESCRIPTION with version 0.6.1 (addressing all comments below).
Update 14 Sept: 0.6.3 after tidyr 1.0.0 moves from dev to stable dependency
Update 18 Sept: 0.6.4 uses new example data for tests and vignettes
Update 23 Sept: 0.6.5 uses lifecycle badges
Update 25 Sept: 0.6.6 odata_submission_get() parses data, dates, and downloads attachments
Update 10-18 Oct: met ODK developers and users; met @jmt2080ad, met sckott (Scott Chamberlain), held four ruODK workshops (Perth, Seattle, Portland), got constructive feedback
Update 27 Nov: 0.6.6.900* addressing reviewer comments, preparing for ODK Central 0.7 release
Update 12 Feb 2020: reviewer comments addressed, some improvements added, ready for reviewer response
Update 15 Jun 2020: reviewer response addressed, parsers added for geotrace and geoshape
