### 2019-08-27 #335 @maelle
### Editor checks:

---

#### Editor comments

:wave: @florianm! Thanks a lot for your submission! :grin: I have a few questions/asks I'd like to see tackled before I start looking for reviewers. I am also waiting to run the local checks because I don't have credentials yet, see [my forum post](https://forum.opendatakit.org/t/need-access-to-the-odk-central-sandbox/21696).

* I see a few open issues in your repo. Could you clarify here and in their description whether you're seeking advice on the functionalities' implementation/usefulness? Why not solve some of them before the reviews of the package? It'd be better for the functionality to be implemeneted so the reviewers can review them too. It is fine if we put this issue on hold for a bit in order for you to have time to work on the enhancements if needed, and you could ask for help in rOpenSci channels.

* Reg "Note: I would like to submit a paper about the package in a few weeks, but haven't got the manuscript ready and approved for publication just yet." This is also a good reason to put the issue on hold since the reviewers would read the short JOSS paper. What do you mean by approved?

* I do not understand "Especially in these trying times, it is important to ask: “ruODK?”" and " u r ODK!" in the README, could you add footnotes or so there? Unless it is very obvious :see_no_evil:

* Likewise I do not understand the use of `%<%` when describing use cases.

* There is a link missing " [ODK forum]". How should the future reviewers best ask for credentials if they're new to the service? I wasn't too sure about my own post.

* The current error messages are not informative in the absence of credentials, i.e. I get "Error in curl::curl_fetch_memory(url, handle = handle) : <url> malformed " if I run `project_list()` now without credentials. If there's no system variables set and no credentials passed either, the functions should fail with an informative error message.

* Do not use the Author and Maintainer fields in DESCRIPTION, they are automatically generated from Authors@R.

* In the example in the README you source .Rprofile. This should not be done, in your code in at least one function I see you use Sys.getenv("NAMEOFTHEVARIABLE") to access the credentials which is the best practice.

* The reference section of the pkgdwon website could use [some grouping](https://devguide.ropensci.org/building.html#function-grouping). (ignore the CI advice, outdated, cf https://ropensci.org/technotes/2019/06/07/ropensci-docs/). I see you used "usethis goodies", 1) CONTRIBUTING.md still mentions the tidyverse team at least once, 2) For info here is [our guidance](https://devguide.ropensci.org/collaboration.html#contributing-guide) for accepted packages.

*There's info about contributing both in CONTRIBUTING.md and in the README (release steps), could you consolidate those?

The following items is rather a suggestion.

* Is there any diagram of ODK services/a workflow with ODK that you could use/adapt to explain visually where your R package fits, as a complement of the written part of the README?

---

Reviewers:
Due date:
