### 2019-08-28 dbca-wa/ruODK#13 @florianm
Clarify scope and timeline of open issues
## Feature
Source https://github.com/ropensci/software-review/issues/335 @maelle

## Task
 ```
I see a few open issues in your repo.
Could you clarify here and in their description whether you're seeking advice
on the functionalities implementation/usefulness?
Why not solve some of them before the reviews of the package?
It'd be better for the functionality to be implemented so the reviewers can review them too.
It is fine if we put this issue on hold for a bit in order for you to have time to work on
the enhancements if needed, and you could ask for help in rOpenSci channels.
```
