### 2019-08-28 dbca-wa/ruODK#13 @florianm
Updated wording on currently open issues to clarify scope and timelines.
I will add a helper for #5. The other issues are musings for future nice-to-have functionality.
Note the issues are all under https://github.com/dbca-wa/ruODK/milestone/2 which targets ODK Central 1.0, which is some rounds of development and user feedback away still.
I've added "help wanted" on a question, but that's not breaking or urgent.
