### 2019-08-28 dbca-wa/ruODK#14 @florianm
Clarify publication plans
## Feature
Clarify publication

## Task
Source https://github.com/ropensci/software-review/issues/335 @maelle
```
Reg "Note: I would like to submit a paper about the package in a few weeks,
but haven't got the manuscript ready and approved for publication just yet."
This is also a good reason to put the issue on hold since the reviewers
would read the short JOSS paper. What do you mean by approved?
```
