### 2019-08-28 dbca-wa/ruODK#14 @florianm
A rudimentary JOSS-style paper is actually [included](https://github.com/dbca-wa/ruODK/blob/master/inst/joss/paper.md).
However I'd like to check scope/fit for an MEE paper first, which will both exclude prior publication to JOSS (is that right? can we have both?) and take a while longer than I'd like to hold up the ROpenSci submission. @maelle rightly mentioned that a few diagrams would help to draw the bigger picture - already working on it ;-)

The reason for not wanting to hold up ROpenSci is an ODK gathering later this year where I'd like to prepare/solidify my end (ruODK) as far as possible. I feel that the feedback from ROpenSci reviewers will help me find my blind spots and improve ruODK efficiently.

Being a public servant, writing a paper (MEE or JOSS) involves a joyful and bureaucratic approval process to make sure I'm picking the right words and not contradicting my Department's policies/positions/strategic directions. This means further delays of getting a paper written through our internal review.

So in a nutshell, ROpenSci review will help to deliver a robust, production-worthy package while the publication at MEE or JOSS will undoubtedly come but at a slightly slower pace than the ODK timeline would like.
