### 2019-08-28 dbca-wa/ruODK#15 @florianm
Clarify lingo and assumptions to make README more inclusive
## Feature
Source https://github.com/ropensci/software-review/issues/335 @maelle

```
I do not understand "Especially in these trying times, it is important to ask: “ruODK?”"
and " u r ODK!" in the README, could you add footnotes or so there?
Unless it is very obvious see_no_evil (play on words with OK?
 to my defence and the defence of my suggestion,
my not being a native speaker might have made this less clear joy )

Likewise I do not understand the use of %<% when describing use cases.

There is a link missing " [ODK forum]". How should the future reviewers best ask
for credentials if they're new to the service? I wasn't too sure about my own post.

In the example in the README you source .Rprofile. This should not be done,
in your code in at least one function I see you use Sys.getenv("NAMEOFTHEVARIABLE")
to access the credentials which is the best practice.
```
