### 2019-08-28 dbca-wa/ruODK#15 @florianm
I'll update the README to clarify:

- ruODK is a wordplay on the mental health campaign "ruODK/are you OK", the OpenDataKit acronym "ODK" and of course, being an R package, we had to have an "r" in there somewhere. The ODK community is indeed a very caring and inclusive community with a strong focus on making users feel safe and welcome. Since ruODK aims to take the "ow" out of the "how" of handling OData (XML/JSON/nested lists in R), the caring meaning translates well to both the aim of the ODK community and ruODK.

- The use of the magrittr pipe `%>%` was meant as a light-hearted way of saying "then" between the steps of the described workflows. I'll remove them as possible source of confusion and bias towards the tidyverse - base R and datatableverse are welcome too!
