### 2019-08-28 dbca-wa/ruODK#16 @florianm
Make settings fun again - including loud but intelligible yelling if something is missing
## Feature
All functions using settings should fail with a meaningful error message.

Source https://github.com/ropensci/software-review/issues/335 @maelle

```
The current error messages are not informative in the absence of credentials,
i.e. I get Error in curl::curl_fetch_memory(url, handle = handle) : <url> malformed
if I run project_list() now without credentials.
If there's no system variables set and no credentials passed either,
the functions should fail with an informative error message.
```
