### 2019-08-28 dbca-wa/ruODK#17 @florianm
Remove Author and Maintainer from DESCRIPTION
## Problem
Source https://github.com/ropensci/software-review/issues/335 @maelle

```
Do not use the Author and Maintainer fields in DESCRIPTION, they are automatically generated from Authors@R.
```
