### 2019-08-28 #335 @florianm
Very good points, thanks! I'll handle them as issues at https://github.com/dbca-wa/ruODK/milestone/3 apologies in advance for @ bombing you, @maelle
