### 2019-08-28 #335 @maelle
No problem, I was glad to see the issues. However we strive to keep the discussion in this submission thread (similar to what happens in journals). This does not prevent opening issues (which I personally would also do for project management),  but answers should also be here. So when all items are settled it'd be nice if you can write a comment here with quotes of the items and your answers. 🙂w

I will answer about publication processes in the next few days.

Can you please explain to me how to finish my credentials setup to run the tests? Thanks.
