### 2019-08-28 e137ead
Cleanup documentation
README: explain ODK, add links
CONTRIBUTING.md: tidy up, merge with release, add test instructions
CITATION: bump version number
DESCRIPTION: drop Author, Maintainer
rebuild pkgdown site

Fix #15, fix #17
