+++ README.md
@@ -25,17 +25,22 @@ software for collecting, managing, and using data in
 resource-constrained environments.
 
 ODK consists of a range of [software packages and
-apps](https://opendatakit.org/software/). For a detailed overview, read
-the extensive [ODK documentation](https://docs.opendatakit.org/).
+apps](https://opendatakit.org/software/). `ruODK` assumes some
+familiarity of its users with the ODK ecosystem and workflows. For a
+detailed overview, read the extensive [ODK
+documentation](https://docs.opendatakit.org/) and visit the friendly
+[ODK forum](https://forum.opendatakit.org/).
 
 [ODK Central](https://docs.opendatakit.org/central-intro/) is a
 cloud-based data clearinghouse for digitally captured data, replacing
 the older software [ODK
 Aggregate](https://docs.opendatakit.org/aggregate-intro/). ODK Central
 manages user accounts and permissions, stores form definitions, and
-allows data collection clients like ODK Collect to connect to it for
-form download and submission upload.
+allows data collection clients like [ODK
+Collect](https://docs.opendatakit.org/collect-intro/) to connect to it
+for form download and submission upload.
 
+A typical [ODK workflow](https://docs.opendatakit.org/#how-is-odk-used):
 After data have been captured digitally using ODK Collect, the data are
 uploaded and stored in ODK Central. The next step from there is to
 extract the data, optionally upload it into another data warehouse, and
@@ -45,6 +50,10 @@ While data can be retrieved in bulk through the GUI, ODK Central’s API
 provides access to its data and functionality through both an OData and
 a RESTful API with a comprehensive and interactive
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
+
+`ruODK` is aimed at the technically minded researcher who wishes to
+access and use the data from ODK Central using the programming language
+R.
 
 `ruODK`’s scope:
 
@@ -58,19 +67,19 @@ a RESTful API with a comprehensive and interactive
 `ruODK`’s use cases:
 
   - Smaller projects:
-      - Data collection: ODK Collect `%>%`
-      - Data clearinghouse: ODK Central `%>%`
-      - Data analysis and reporting: `Rmd` (ruODK) `%>%`
-      - Publishing and dissemination:
+    1.  Data collection: ODK Collect
+    2.  Data clearinghouse: ODK Central
+    3.  Data analysis and reporting: `Rmd` (ruODK)
+    4.  Publishing and dissemination:
         [`ckanr`](https://docs.ropensci.org/ckanr/),
         [`CKAN`](https://ckan.org/)
   - Larger projects:
-      - Data collection: ODK Collect `%>%`
-      - Data clearinghouse: ODK Central `%>%`
-      - ETL pipeline into data warehouses: `Rmd` (ruODK) `%>%`
-      - QA: in data warehouse `%>%`
-      - Reporting: `Rmd` `%>%`
-      - Publishing and dissemination:
+    1.  Data collection: ODK Collect
+    2.  Data clearinghouse: ODK Central
+    3.  ETL pipeline into data warehouses: `Rmd` (ruODK)
+    4.  QA: in data warehouse
+    5.  Reporting: `Rmd`
+    6.  Publishing and dissemination:
[…]
