+++ README.Rmd
@@ -23,19 +23,23 @@ knitr::opts_chunk$set(
 
 Especially in these trying times, it is important to ask: "ruODK?"
 
-[OpenDataKit](https://opendatakit.org/) (ODK) is a free and open-source software for 
-collecting, managing, and using data in resource-constrained environments.
+[OpenDataKit](https://opendatakit.org/) (ODK) is a free and open-source software 
+for collecting, managing, and using data in resource-constrained environments.
 
 ODK consists of a range of [software packages and apps](https://opendatakit.org/software/).
-For a detailed overview, read the extensive [ODK documentation](https://docs.opendatakit.org/).
+`ruODK` assumes some familiarity of its users with the ODK ecosystem and workflows.
+For a detailed overview, read the extensive [ODK documentation](https://docs.opendatakit.org/)
+and visit the friendly [ODK forum](https://forum.opendatakit.org/).
 
 [ODK Central](https://docs.opendatakit.org/central-intro/) is a cloud-based
 data clearinghouse for digitally captured data, replacing the older software 
 [ODK Aggregate](https://docs.opendatakit.org/aggregate-intro/). 
 ODK Central manages user accounts and permissions, stores form definitions, 
-and allows data collection clients like ODK Collect to connect to it for 
+and allows data collection clients like 
+[ODK Collect](https://docs.opendatakit.org/collect-intro/) to connect to it for 
 form download and submission upload.
 
+A typical [ODK workflow](https://docs.opendatakit.org/#how-is-odk-used):
 After data have been captured digitally using ODK Collect, the data are uploaded
 and stored in ODK Central. The next step from there is to extract the data, 
 optionally upload it into another data warehouse, and then to analyse and 
@@ -45,6 +49,9 @@ While data can be retrieved in bulk through the GUI, ODK Central's API provides
 access to its data and functionality through both an OData and a RESTful API 
 with a comprehensive and interactive 
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
+
+`ruODK` is aimed at the technically minded researcher who wishes to access and
+use the data from ODK Central using the programming language R.
 
 `ruODK`'s scope:
 
@@ -57,25 +64,27 @@ with a comprehensive and interactive
 `ruODK`'s use cases:
 
 * Smaller projects:
-  * Data collection: ODK Collect `%>%`
-  * Data clearinghouse: ODK Central `%>%` 
-  * Data analysis and reporting: `Rmd` (ruODK) `%>%`
-  * Publishing and dissemination: [`ckanr`](https://docs.ropensci.org/ckanr/), 
-    [`CKAN`](https://ckan.org/)
+  1. Data collection: ODK Collect
+  2. Data clearinghouse: ODK Central
+  3. Data analysis and reporting: `Rmd` (ruODK)
+  4. Publishing and dissemination: [`ckanr`](https://docs.ropensci.org/ckanr/), 
+     [`CKAN`](https://ckan.org/)
+     
 * Larger projects:
-  * Data collection: ODK Collect `%>%`
-  * Data clearinghouse: ODK Central `%>%` 
-  * ETL pipeline into data warehouses: `Rmd` (ruODK) `%>%`
-  * QA: in data warehouse `%>%`
-  * Reporting: `Rmd` `%>%`
-  * Publishing and dissemination: [`ckanr`](https://docs.ropensci.org/ckanr/),
-    [`CKAN`](https://ckan.org/)
+  1. Data collection: ODK Collect
+  2. Data clearinghouse: ODK Central
+  3. ETL pipeline into data warehouses: `Rmd` (ruODK)
+  4. QA: in data warehouse
+  5. Reporting: `Rmd`
[…]
