### 2019-08-28 dbca-wa/ruODK#18 @florianm
Group function docs
## Feature
Source https://github.com/ropensci/software-review/issues/335 @maelle

```
The reference section of the pkgdown website could use some grouping.
https://devguide.ropensci.org/building.html#function-grouping
(ignore the CI advice, outdated, cf https://ropensci.org/technotes/2019/06/07/ropensci-docs/).
2) For info here is our guidance for accepted packages: https://devguide.ropensci.org/collaboration.html#contributing-guide
```
