### 2019-08-28 dbca-wa/ruODK#19 @florianm
Add diagram of ODK data flow
## Feature
Add diagram of ODK data flow to README.
Source https://github.com/ropensci/software-review/issues/335 @maelle
