### 2019-08-28 #335 @florianm
No probs, when done I'll go through comments here and address how my issues address them.
Give me a day or two to address the larger issues!

Test credentials as promised:
https://github.com/dbca-wa/ruODK/blob/master/CONTRIBUTING.md#test
https://dbca-wa.github.io/ruODK/CONTRIBUTING.html#test
