### 2019-08-28 673ca94
Group function docs, fix #18
