### 2019-08-28 #335 @maelle
Thanks! I've just realized you recommend using Rprofile but it is code run at each session opening, environment variables should be in .Renviron, I'm editing mine via `usethis::edit_r_environ()` right now.
