### 2019-08-28 #335 @maelle
cf https://whattheyforgot.org/r-startup.html (maybe useful to link from docs?)
