### 2019-08-28 #335 @maelle
I can run tests now but not code such as `project_list()` (e.g. knitting `api.Rmd` fails), how do I do that, more setup steps I guess? Can I use the same credentials again, changing their names?
