### 2019-08-28 #335 @florianm
Aw crap. The vignette might need settings for our internal ODK Central server. Will send settings in ca 3h, afk just now
