### 2019-08-28 #335 @florianm
You can always use explicit url, un, pw as per vignette Setup:

project_list(url=Sys.getenv("ODKC_TEST_URL"), un=..., pw=...)
