### 2019-08-28 #335 @florianm
The test settings (sandbox) are real forms with test data.
The tests are read-only and non destructive, so no worries there.

The server used in the vignettes uses real forms with (so far) only a few test records. Both servers run ODK Central v0.6.
