### 2019-08-28 #335 @maelle
And you don't want random contributors to have access to the real forms from the vignette, correct? (I might be missing something obvious) A discuss.ropensci.org forum post about the general issue (using interesting examples in the vignette, without giving access to contributors, and with R CMD check still passing) might make sense?
