### 2019-08-28 #335 @florianm
Not a problem at all to scale actually!
Currently I'm using the few forms/records I have so far. Once significant production data comes in, I might create a separate testing project and give both selected/invited web users (ruODK un/pw) and app users (to submit data via ODL Collect) access to just that.
I think I'll get away with a manual invitation process for prospective contributors at least for the time being.
How does that sound?

Also, invite to my server coming in an hour!
