### 2019-08-28 #335 @maelle
Cool so I'll then just have to add `ODKC_URL` / other variables to my .Renviron? Thank you!
