### 2019-08-28 dbca-wa/ruODK#20 @florianm
Make contributing great again
## Feature
Clarify which accounts and settings to request and use to run tests and build the vignettes.

Source https://github.com/ropensci/software-review/issues/335 @maelle
