### 2019-08-28 a4b9794
Include info on accounts required to test and build ruODK
Fix #20
