+++ .github/ISSUE_TEMPLATE/account_request.md
@@ -1,0 +1,18 @@
+---
+name: Account request
+about: I'd like to run tests or build the vignettes
+title: 'Account request for ODK Central instances used by ruODK'
+labels: help-wanted
+assignees: florianm
+
+---
+
+## Email
+Please send me your email via email (see package contact info).
+
+## Purpose
+
+- [ ] I want to run the tests and require access to the ODK Central Sandbox 
+  https://sandbox.central.opendatakit.org/
+- [ ] I want to build the vignettes and require access to 
+  https://odkcentral.dbca.wa.gov.au/
