### 2019-08-28 #335 @florianm
> Cool so I'll then just have to add `ODKC_URL` / other variables to my .Renviron? Thank you!

This should make way more sense now: https://dbca-wa.github.io/ruODK/CONTRIBUTING.html#test
Also check your email for working settings!
