### 2019-08-28 #335 @maelle
Awesome, thank you! Got the email, will `devtools::check()` soon. I will be ODK now 😉

I think the instructions should not contain `Sys.getenv` in order to copy paste more easily to .Renviron. 🙂
