### 2019-08-28 2a9da72
Missed a spot: suggest .Renviron in README #20
