+++ README.Rmd
@@ -135,7 +135,7 @@ Sys.setenv(ODKC_PW = ".......")
 ```
 
 For a more permanent configuration setting, paste the above lines into your 
-`~/.Rprofile` (`usethis::edit_r_profile()`) and restart R.
+`~/.Renviron` (`usethis::edit_r_environ(scope = "user")`) and restart R.
 
 For all available detailed options to configure `ruODK`, read 
 `vignette("Setup", package = "ruODK")` 
