### 2019-08-28 #335 @florianm
Glad to see that urODK now :-D

> I think the instructions should not contain Sys.getenv in order to copy paste more easily to .Renviron.
Happy to adjust, which instructions do you mean?

Re authentication and settings, should I add abstraction levels like e.g. `ckanr` does?
* `get_odkc_{test_}{url, un, pw}()` to wrap `Sys.getenv("ODKC_{TEST_}{URL, UN,PW}")` as standard args for functions, e.g. `project_list(url = get_odkc_url(), un = get_odkc_un(), ...)`
* `ruODK_setup(url="xxx", un="yyy", pw="zzz", test_url="...", test_un="...", test_pw="...")` to wrap `Sys.setenv(ODKC_{TEST_}{URL, UN, PW})`
That would be an opportunity to make ruODK yell loudly but clearly about missing credentials.
