### 2019-08-28 dbca-wa/ruODK#16 @florianm
It's probably time to bite the ckanr_setup/ckanr_settings apple I've avoided so far:

* `get_odkc_{test_}{url, un, pw}()` to wrap `Sys.getenv("ODKC_{TEST_}{URL, UN,PW}")` as standard args for functions, e.g. `project_list(url = get_odkc_url(), un = get_odkc_un(), ...)`
* `ruODK_setup(url="xxx", un="yyy", pw="zzz", test_url="...", test_un="...", test_pw="...")` to wrap `Sys.setenv(ODKC_{TEST_}{URL, UN, PW})`
That would be an opportunity to make ruODK yell loudly but clearly about missing credentials.
