### 2019-08-28 #335 @maelle
I mean in https://dbca-wa.github.io/ruODK/CONTRIBUTING.html#test

```r
ODKC_TEST_URL="https://sandbox.central.opendatakit.org"
ODKC_TEST_PID=14
ODKC_TEST_FID="build_Flora-Quadrat-0-2_1558575936"
ODKC_TEST_UN="your@email.com"
ODKC_TEST_PW="..."
ODKC_URL="https://odkcentral.dbca.wa.gov.au"
ODKC_UN="your@email.com"
ODKC_PW="..."
```

that are the lines for .Renviron.

reg your 2d question better to ping ckanr's maintainer who's an editor too :-)
