### 2019-08-28 1bef59b
Fix formatting for .Renviron content
