+++ README.Rmd
@@ -126,16 +126,13 @@ The `.odkbuild` versions can be loaded into [ODK Build](https://build.opendataki
 while the `.xml` versions can be imported into ODK Central.
 
 ## Configure ruODK
-For a quick start, create R environment variables with your settings:
+For a quick start, create R environment variables with your ODK Central credentials:
   
 ```{r rprofile, eval=F}
 Sys.setenv(ODKC_URL = "https://odkcentral.mydomain.com")
 Sys.setenv(ODKC_UN = "me@mail.com")
 Sys.setenv(ODKC_PW = ".......")
 ```
-
-For a more permanent configuration setting, paste the above lines into your 
-`~/.Renviron` (`usethis::edit_r_environ(scope = "user")`) and restart R.
 
 For all available detailed options to configure `ruODK`, read 
 `vignette("Setup", package = "ruODK")` 
