### 2019-08-28 #335 @florianm
Ah, thanks! Fixed and pushed now. I'll work on https://github.com/dbca-wa/ruODK/issues/16 and https://github.com/dbca-wa/ruODK/issues/19 tomorrow. Cheers!
