### 2019-08-28 #335 @maelle
Nothing to report from automatic tests, apart from this `goodpractice`'s compliment

`♥ Hurrah! Mathematical package! Keep up the delightful work!`

so as soon as you have tackled the remaining issues and written answers here, and as I answer regarding MEE&rOpenSci publication process, we can move forward. :-)
