+++ man-roxygen/param-auth.R
@@ -1,7 +1,11 @@
-#' @param url The OData URL, ending in .svc, no trailing slash.
-#' @param un The ODK Central username (an email address),
-#'           default: Sys.getenv("ODKC_UN").
-#'           Add to your ~/.Rprofile: Sys.setenv(ODKC_UN="...@...")
-#' @param pw The ODK Central password,,
-#'           default: Sys.getenv("ODKC_PW").
-#'           Add to your ~/.Rprofile: Sys.setenv(ODKC_PW="...")
+#' @param url The base ODK Central without a trailing slash,
+#'           e.g. "https://sandbox.central.opendatakit.org".
+#'           Default: \code{get_default_url()},
+#'           which calls \code{Sys.getenv("ODKC_URL")}.
+#' @param un The ODK Central username (an email address).
+#'           Default: \code{get_default_un()},
+#'           which calls \code{Sys.getenv("ODKC_UN")}.
+#'           See \code{vignette("Setup", package = "ruODK")}.
+#' @param pw The ODK Central password.
+#'           Default: \code{get_default_pw()},
+#'           which calls \code{Sys.getenv("ODKC_PW")}.
