### 2019-08-29 046ec09
#16 ru_settings
Add getters and setter for settings, tests, docs.
Convert project_* and related tests to use settings getters as defaults.
TODO convert remaining functions.
