+++ R/ru_setup.R
@@ -1,0 +1,256 @@
+#' Get or set \code{ruODK} settings.
+#'
+#' @export
+#' @return \code{ru_settings} prints your default ODK Central url, username,
+#' and password, and corresponding optional test server settings.
+#' \code{ru_setup} sets your production and test settings, while
+#' \code{get_(test_)*} get each of those respective settings.
+#' @seealso  \code{\link{ru_setup}},
+#' \code{\link{get_default_url}},
+#' \code{\link{get_default_un}},
+#' \code{\link{get_default_pw}},
+#' \code{\link{get_test_url}},
+#' \code{\link{get_test_un}},
+#' \code{\link{get_test_pw}},
+#' \code{\link{get_test_pid}},
+#' \code{\link{get_test_fid}}.
+#' @family ru_settings
+#' @examples
+#' ru_settings()
+ru_settings <- function() {
+  ops <- list(
+    url = Sys.getenv("ODKC_URL", ""),
+    un = Sys.getenv("ODKC_UN", ""),
+    pw = Sys.getenv("ODKC_PW", ""),
+    test_url = Sys.getenv("ODKC_TEST_URL", ""),
+    test_un = Sys.getenv("ODKC_TEST_UN", ""),
+    test_pw = Sys.getenv("ODKC_TEST_PW", ""),
+    test_pid = Sys.getenv("ODKC_TEST_PID", ""),
+    test_fid = Sys.getenv("ODKC_TEST_FID", "")
+  )
+  structure(ops, class = "ru_settings")
+}
+
+#' @export
+print.ru_settings <- function(x, ...) {
+  cat("<ruODK settings>", sep = "\n")
+  cat("  Default ODK Central URL: ", x$url, "\n")
+  cat("  Default ODK Central Username: ", x$un, "\n")
+  cat("  Default ODK Central Password: ", x$pw, "\n")
+  cat("  Test ODK Central URL:", x$test_url, "\n")
+  cat("  Test ODK Central Username:", x$test_un, "\n")
+  cat("  Test ODK Central Password:", x$test_pw, "\n")
+  cat("  Test ODK Central Project ID:", x$test_pid, "\n")
+  cat("  Test ODK Central Form ID:", x$test_fid, "\n")
+}
+
+#------------------------------------------------------------------------------#
+# Setters
+#
+#' Configure default \code{ruODK} settings.
+#'
+#' @export
+#' @param url An ODK Central URL, e.g. "https://sandbox.central.opendatakit.org".
+#' @param un An ODK Central username which is the email of a "web user" in the
+#'   specified ODK Central instance \code{url} (optional, character).
+#' @param pw The password for user \code{un} (optional, character).
+#' @param test_url (optional, character) A valid ODK Central URL for testing.
+#' @param test_un (optional, character) A valid ODK Central username (email)
+#'   privileged to view the test project(s) at \code{test_url}.
+#' @param test_pw (optional, character) The valid ODK Central password for
+#'   \code{test_un}.
+#' @param test_pid (optional, integer) The numeric ID of an existing project on
+#'   \code{test_url}.
+#' @param test_fid (optional, character) The alphanumeric ID of an existing form
+#'   in \code{test_pid}.
+#' @family ru_settings
+#' @details
+#' \code{ru_setup} sets ODK Central connection details. \code{ruODK}'s functions
+#'   default to use the default URL, username, and password unless specified
+#'   explicitly.
+#'
+#' \code{ruODK}'s automated tests require a valid ODK Central URL, and a
+#'   privileged username and password of a "web user" on that ODK Central
+#'   instance, as well as an existing project and form.
+#'
[…]
