+++ man/ru_settings.Rd
@@ -1,0 +1,60 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ru_setup.R
+\name{ru_settings}
+\alias{ru_settings}
+\alias{get_default_url}
+\alias{get_default_un}
+\alias{get_default_pw}
+\alias{get_test_url}
+\alias{get_test_un}
+\alias{get_test_pw}
+\alias{get_test_pid}
+\alias{get_test_fid}
+\title{Get or set \code{ruODK} settings.}
+\usage{
+ru_settings()
+
+get_default_url()
+
+get_default_un()
+
+get_default_pw()
+
+get_test_url()
+
+get_test_un()
+
+get_test_pw()
+
+get_test_pid()
+
+get_test_fid()
+}
+\value{
+\code{ru_settings} prints your default ODK Central url, username,
+and password, and corresponding optional test server settings.
+\code{ru_setup} sets your production and test settings, while
+\code{get_(test_)*} get each of those respective settings.
+}
+\description{
+Get or set \code{ruODK} settings.
+}
+\examples{
+ru_settings()
+}
+\seealso{
+\code{\link{ru_setup}},
+\code{\link{get_default_url}},
+\code{\link{get_default_un}},
+\code{\link{get_default_pw}},
+\code{\link{get_test_url}},
+\code{\link{get_test_un}},
+\code{\link{get_test_pw}},
+\code{\link{get_test_pid}},
+\code{\link{get_test_fid}}.
+
+Other ru_settings: \code{\link{ru_setup}},
+  \code{\link{yell_if_error}},
+  \code{\link{yell_if_missing}}
+}
+\concept{ru_settings}
