+++ man/yell_if_error.Rd
@@ -1,0 +1,35 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ru_setup.R
+\name{yell_if_error}
+\alias{yell_if_error}
+\title{Warn about failed web requests and give helpful troubleshooting tips.}
+\usage{
+yell_if_error(response, url, un, pw, pid = NULL, fid = NULL)
+}
+\arguments{
+\item{response}{A httr response object}
+
+\item{url}{A URL (character)}
+
+\item{un}{A username (character)}
+
+\item{pw}{A password (character)}
+
+\item{pid}{A project ID (numeric, optional)}
+
+\item{fid}{A form ID (character, optional)}
+}
+\value{
+The response object
+}
+\description{
+A wrapper around \code{httr::stop_for_status()} with a helpful error message.
+Examples: see tests for \code{ruODK::project_list()}.
+This function is used internally but may be useful for debugging and
+\code{ruODK} development.
+}
+\seealso{
+Other ru_settings: \code{\link{ru_settings}},
+  \code{\link{ru_setup}}, \code{\link{yell_if_missing}}
+}
+\concept{ru_settings}

[…]
