### 2019-08-29 4913e4b
Fix #16: use settings getters in functions and tests, bump minor version.
