+++ tests/testthat/test-submission_detail.R
@@ -1,19 +1,19 @@
 test_that("submission_detail works", {
   sl <- submission_list(
-    Sys.getenv("ODKC_TEST_PID"),
-    Sys.getenv("ODKC_TEST_FID"),
-    url = Sys.getenv("ODKC_TEST_URL"),
-    un = Sys.getenv("ODKC_TEST_UN"),
-    pw = Sys.getenv("ODKC_TEST_PW")
+    get_test_pid(),
+    get_test_fid(),
+    url = get_test_url(),
+    un = get_test_un(),
+    pw = get_test_pw()
   )
 
   sub <- submission_detail(
-    Sys.getenv("ODKC_TEST_PID"),
-    Sys.getenv("ODKC_TEST_FID"),
+    get_test_pid(),
+    get_test_fid(),
     sl$instance_id[[1]],
-    url = Sys.getenv("ODKC_TEST_URL"),
-    un = Sys.getenv("ODKC_TEST_UN"),
-    pw = Sys.getenv("ODKC_TEST_PW")
+    url = get_test_url(),
+    un = get_test_un(),
+    pw = get_test_pw()
   )
 
   # submission_detail returns a tibble
