+++ tests/testthat/test-submission_list.R
@@ -1,16 +1,16 @@
 test_that("submission_list works", {
   sl <- submission_list(
-    Sys.getenv("ODKC_TEST_PID"),
-    Sys.getenv("ODKC_TEST_FID"),
-    url = Sys.getenv("ODKC_TEST_URL"),
-    un = Sys.getenv("ODKC_TEST_UN"),
-    pw = Sys.getenv("ODKC_TEST_PW")
+    get_test_pid(),
+    get_test_fid(),
+    url = get_test_url(),
+    un = get_test_un(),
+    pw = get_test_pw()
   )
   fl <- form_list(
-    Sys.getenv("ODKC_TEST_PID"),
-    url = Sys.getenv("ODKC_TEST_URL"),
-    un = Sys.getenv("ODKC_TEST_UN"),
-    pw = Sys.getenv("ODKC_TEST_PW")
+    get_test_pid(),
+    url = get_test_url(),
+    un = get_test_un(),
+    pw = get_test_pw()
   )
   # submission_list returns a tibble
   testthat::expect_equal(class(sl), c("tbl_df", "tbl", "data.frame"))
@@ -23,7 +23,7 @@ test_that("submission_list works", {
 
   # Number of submissions (rows) is same as advertised in form_list
   form_list_nsub <- fl %>%
-    filter(fid == Sys.getenv("ODKC_TEST_FID")) %>%
+    filter(fid == get_test_fid()) %>%
     magrittr::extract2("submissions") %>%
     as.numeric()
   testthat::expect_equal(nrow(sl), form_list_nsub)

[…]
