### 2019-08-29 389bb20
Fix tests: add auth to get_one_attachment
