### 2019-08-29 6d2e2ca
Missed a spot: fix test-attachment_get: add auth
