+++ tests/testthat/test-attachment_get.R
@@ -56,7 +56,8 @@ test_that("get_one_attachment handles repeat download and NA filenames", {
 
   # Happy path: get one attachment should work
   testthat::expect_message(
-    get_one_attachment(pth,
+    get_one_attachment(
+      pth,
       fn,
       src,
       url = get_test_url(),
@@ -66,7 +67,8 @@ test_that("get_one_attachment handles repeat download and NA filenames", {
     ),
     glue::glue("Saved {pth}\n")
   )
-  fn_local <- get_one_attachment(pth,
+  fn_local <- get_one_attachment(
+    pth,
     fn,
     src,
     url = get_test_url(),
@@ -79,7 +81,14 @@ test_that("get_one_attachment handles repeat download and NA filenames", {
 
   # Happy, but faster: keep existing download
   testthat::expect_message(
-    get_one_attachment(pth, fn, src, verbose = TRUE),
+    get_one_attachment(
+      pth,
+      fn,
+      src,
+      url = get_test_url(),
+      un = get_test_un(),
+      pw = get_test_pw(),
+      verbose = TRUE),
     glue::glue("Keeping {pth}\n")
   )
 
@@ -96,26 +105,28 @@ test_that("get_one_attachment handles repeat download and NA filenames", {
     glue::glue("Keeping {pth}\n")
   )
   testthat::expect_equal(
-    get_one_attachment(pth,
-                       NA,
-                       src,
-                       url = get_test_url(),
-                       un = get_test_un(),
-                       pw = get_test_pw(),
-                       verbose = TRUE),
+    get_one_attachment(
+      pth,
+      NA,
+      src,
+      url = get_test_url(),
+      un = get_test_un(),
+      pw = get_test_pw(),
+      verbose = TRUE),
     as.character(pth)
   )
 
   # Now make sure pth doesn't exist
   pth <- fs::path(tempdir(), NA) %>% as.character()
   testthat::expect_message(
-    get_one_attachment(pth,
-                       NA,
-                       src,
-                       url = get_test_url(),
-                       un = get_test_un(),
-                       pw = get_test_pw(),
-                       verbose = TRUE),
+    get_one_attachment(
+      pth,
+      NA,
+      src,
+      url = get_test_url(),
+      un = get_test_un(),
+      pw = get_test_pw(),
+      verbose = TRUE),
     "Filename is NA, skipping download."
   )
   testthat::expect_true(
