### 2019-08-30 dbca-wa/ruODK#14 @maelle
Thanks for the explanation! @sckott please correct me if I'm wrong.

Indeed I think you'd need to choose between JOSS  and MEE.

For MEE, the paper doesn't have to be ready but it's better if it's more or less ready. We at rOpenSci have no influence on their deeming a paper in scope/fitting.

For JOSS sorry I realize I need to ask too, will write again here (and then we can summarize in the main thread 😉).

Awesome about the diagrams! As an outsider I am looking forward to seeing them!
