### 2019-08-30 dbca-wa/ruODK#14 @sckott
I don't know about JOSS vs. MEE. I'd suggest emailing MEE and asking them. Maybe they'd consider JOSS similar enough to a pre-print, but I don't know.
