### 2019-08-30 dbca-wa/ruODK#14 @maelle
Reg JOSS it'd be fine to add the short paper.md at the very end after the reviews, in which case I as editor would read it before you submit it for the expedited process at JOSS.
