### 2019-08-31 #335 @florianm
I think I've addressed all issues now, will write up summary here as soon as I get a moment.

Last question, is there a best practice on importing packages?
in `ruODK` I managed to keep `goodpractice()` and `check()` happy by:
* DESCRIPTION `imports` all packages I use in code
* DESCRIPTION `suggests` extra packages used in vignettes
* My code prefixes imported functions e.g. dplyr::mutate_at
* My Roxygen docs `@importFrom` imported functions e.g. `@importFrom dplyr mutate_at`

What's the ROpenSci way?
