### 2019-08-31 #335 @maelle
You could drop the `@importFrom` https://r-pkgs.org/namespace.html#imports but you don't have to.🙂
