### 2019-09-06 7f48288
Add new convenience helpers
