+++ R/form_schema.R
@@ -22,8 +22,6 @@
 #'   fields are captured (e.g. "instanceId", form start datetimes etc.).
 #' @seealso \url{https://odkcentral.docs.apiary.io/#reference/forms-and-submissions/'-individual-form/retrieving-form-schema-json}
 #' @family restful-api
-#' @importFrom httr add_headers authenticate content GET
-#' @importFrom glue glue
 #' @export
 #' @examples
 #' \dontrun{

[…]
