### 2019-09-09 342a32a
prep for 0.6.2: close #19, refactor setup
