+++ man-roxygen/param-fid.R
@@ -1,1 +1,4 @@
 #' @param fid The alphanumeric form ID, e.g. "build_Spotlighting-0-8_1559885147".
+#'   Default: \code{get_default_fid()}.
+#'   Set default \code{fid} through \code{ru_setup(fid="...")}.
+#'   See \code{vignette("Setup", package = "ruODK")}.
