+++ man-roxygen/param-pid.R
@@ -1,1 +1,4 @@
 #' @param pid The numeric ID of the project, e.g.: 2.
+#'   Default: \code{get_default_pid()}.
+#'   Set default \code{pid} through \code{ru_setup(pid="...")}.
+#'   See \code{vignette("Setup", package = "ruODK")}.

[…]
