### 2019-09-09 4218fb1
Fix typos in tests
