### 2019-09-09 b6d0329
Fix tests: do not delete tempdir
