### 2019-09-09 ff7fa0e
Fix #22: share vignette attachment files
