+++ R/attachment_link.R
@@ -25,7 +25,7 @@ prefix_fn <- function(fn, prefix) {
 #' @param data_tbl The downloaded submissions from
 #'   \code{`ruODK::submission_export()`} read into a `tibble` by
 #'   \code{`readr::read_csv()`}.
-#' @param att_path A local path, defaule: "media" (as per .csv.zip export).
+#' @param att_path A local path, default: "media" (as per .csv.zip export).
 #'   Selected columns of the dataframe (containing attchment filenames) are
 #'   prefixed with `att_path`, thus turning them into relative paths.
 #' @param att_contains A shared part of attachment fieldnames, default: "photo".
