### 2019-09-09 074cf7f
Drop unused vignette, fix typo, rebuild site.
