### 2019-09-09 81f9c20
Fix check note on line length in attachment_link
