### 2019-09-10 5caeb79
Provide background info on env vars and .Renviron
Add explanation of env vars at R startup from whattheyforgot.com to vignette setup. Thanks for the suggestion and link @maelle
