### 2019-09-10 78fbd37
Include non-technical contribution instructions
