### 2019-09-10 dbca-wa/ruODK#23 @florianm
Speed up tests: use smaller form for submission_export
## Feature
Current tests download a 50MB zip file repeatedly. We should use a form with one (even in original file size very small) attachment and one submission for those tests.

Process:

* Use newer example form (flora quadrat 0.4)
* Set data collection device's ODK Collect to take smallest possible photos
* Capture a few test records (30 min)
* Update README, tests, vignettes to use new form
* If the submissions ZIP export is then small enough (say <10MB), tests will run way faster. This will fix #21 for now.
