### 2019-09-10 #335 @florianm
Hi @maelle, thanks for the editor comments and further questions. I'll address them all below.

## Editor comments

>  I see a few open issues in your repo. Could you clarify here and in their description whether you're seeking advice on the functionalities implementation/usefulness? Why not solve some of them before the reviews of the package? It'd be better for the functionality to be implemented so the reviewers can review them too. It is fine if we put this issue on hold for a bit in order for you to have time to work on the enhancements if needed, and you could ask for help in rOpenSci channels.

* Issues grouped into milestones [ROpenSci](https://github.com/dbca-wa/ruODK/milestone/3) (all addressed apart from publication) and [ODK Central Release 1.0](https://github.com/dbca-wa/ruODK/milestone/2) which is a collection of ideas rather than bugs or missing features.

>  Reg "Note: I would like to submit a paper about the package in a few weeks, but haven't got the manuscript ready and approved for publication just yet." This is also a good reason to put the issue on hold since the reviewers would read the short JOSS paper. What do you mean by approved?

* https://github.com/dbca-wa/ruODK/issues/8
* JOSS paper is included [here](https://github.com/dbca-wa/ruODK/blob/master/inst/joss/paper.md) as the short form of README and vignettes. I went with the length and scope of other JOSS papers but my JOSS manuscript feels a bit thin.
* Reviewers will probably get the best picture from the README and the vignettes.
* I would like to write a paper for MEE but have no bandwidth to do so before December.
* Writing a paper for JOSS might be quicker but unfortunately excludes the MEE option.
* I would like ruODK to be part of ROpenSci - if possible - before the ODK convening in Seattle Oct 11-13 and my presentation of ODK and ruODK to RUser PDX Oct 15. (ping @sckott)

>     I do not understand "Especially in these trying times, it is important to ask: “ruODK?”" and " u r ODK!" in the README, could you add footnotes or so there? Unless it is very obvious see_no_evil (play on words with OK? to my defence and the defence of my suggestion, my not being a native speaker might have made this less clear joy )

* Fulfills minimum requirement of at least one bad programmer-dad-joke in package name (@jennybc  https://twitter.com/JennyBryan/status/1169365834239631360 )
* Emphasises that we care (cf. ruOK?) about R users of ODK
* Clarified a bit more in README
* ruODK comes out first on [Google](https://www.google.com/search?q=ruODK). second on [duckduckgo](https://duckduckgo.com/?q=ruodk), fourth on [bing](https://www.bing.com/search?q=ruodk) (tested in private mode which should prevent search bubble).

>    Likewise I do not understand the use of %<% when describing use cases.

* Dropped the tongue-in cheek "and then" `magrittr` pipe connector between stages of workflow in README in favour of numbering

>     There is a link missing " [ODK forum]". How should the future reviewers best ask for credentials if they're new to the service? I wasn't too sure about my own post.

* Added more links and background to README. Thanks for the "non-ODK user" perspective!

>     The current error messages are not informative in the absence of credentials, i.e. I get Error in curl::curl_fetch_memory(url, handle = handle) : <url> malformed if I run project_list() now without credentials. If there's no system variables set and no credentials passed either, the functions should fail with an informative error message.

* Really good point! Addressed in https://github.com/dbca-wa/ruODK/issues/16 - added loud and clear error messages.
* Happy to take feedback on [wording](https://github.com/dbca-wa/ruODK/blob/master/R/ru_setup.R#L340).
* https://dbca-wa.github.io/ruODK/reference/yell_if_error.html
* https://dbca-wa.github.io/ruODK/reference/yell_if_missing.html
* See also complete refactor of settings (same issue), based on best practice from `ckanr`

>     Do not use the Author and Maintainer fields in DESCRIPTION, they are automatically generated from Authors@R.

* Done https://github.com/dbca-wa/ruODK/issues/17

>     In the example in the README you source .Rprofile. This should not be done, in your code in at least one function I see you use Sys.getenv("NAMEOFTHEVARIABLE") to access the credentials which is the best practice.

* Thanks, learned something new here! .Renviron instead of .Rprofile
* See also complete refactor of setup (issue 16)
* Refactored vignette [setup](https://dbca-wa.github.io/ruODK/articles/setup.html), [README](https://dbca-wa.github.io/ruODK/index.html), other vignettes, every function, every example, every test, contributing.md to use kwargs `(pid = get_default_pid(), ...)` set by `ru_setup`. Took a while, apologies!

>     The reference section of the pkgdwon website could use some grouping. (ignore the CI advice, outdated, cf https://ropensci.org/technotes/2019/06/07/ropensci-docs/). I see you used "usethis goodies", 1) CONTRIBUTING.md still mentions the tidyverse team at least once, 2) For info here is our guidance for accepted packages.

* Cleaned out the `usethis` boilerplate
* Figured out [function grouping](https://github.com/dbca-wa/ruODK/blob/master/_pkgdown.yml)
* https://github.com/dbca-wa/ruODK/issues/18

>  There's info about contributing both in CONTRIBUTING.md and in the README (release steps), could you consolidate those?
>  Can you please explain to me how to finish my credentials setup to run the tests? Thanks.

* Rewrote and consolidated contributor's in https://github.com/dbca-wa/ruODK/issues/20
* Added non-technical contributor instructions to CONTRIBUTING.md

The following items is rather a suggestion.

>     Is there any diagram of ODK services/a workflow with ODK that you could use/adapt to explain visually where your R package fits, as a complement of the written part of the README?

* There is now! Thanks for the suggestion, makes the overview much more palatable. Also included in JOSS paper. Mildly increases installed package size. https://github.com/dbca-wa/ruODK/issues/19

## Editor's further requests in thread
> You could drop the @importFrom https://r-pkgs.org/namespace.html#imports but you don't have to.

* Dropped `@importFrom` in favour of properly namespacing imported functions.

> And you don't want random contributors to have access to the real forms from the vignette, correct? (I might be missing something obvious) A discuss.ropensci.org forum post about the general issue (using interesting examples in the vignette, without giving access to contributors, and with R CMD check still passing) might make sense?

* Added an [issue template for an account request](https://github.com/dbca-wa/ruODK/issues/new/choose). I (and all other ODK Central Sandbox users) can create new web user accounts to be used for testing ruODK. I will make sure to stay in line with ODK's acceptable use of the sandbox.

## Other improvements
* See [changelog](https://dbca-wa.github.io/ruODK/news/index.html)
* Use janitor to clean names
* Add convenience helpers to [link attachments](https://dbca-wa.github.io/ruODK/reference/attachment_link.html) (by partial match of column name) and [parse dates](https://dbca-wa.github.io/ruODK/reference/parse_datetime.html).
* Refactored OData [attachment download](https://dbca-wa.github.io/ruODK/reference/attachment_get.html) to share same attachments between OData, CSV export, and REST methods. As we demonstrate all three methods, this leads to a smaller installed package size.
* All tests and vignettes use our public test forms on the ODK Central Sandbox. They contain only dummy data but are (earlier versions of) forms we use in anger (or will do soon after some further improvements).
* All tests run (against our own data) in the ODK Central Sandbox, which is kept up to date with the latest master branch. This way we make sure ruODK tests the latest & greatest of the real deal. Once ODK Central solidifies around v1.0, we can probably turn to using webmockr & vcr for caching test results. @sckott might have a qualified opinion on this.
* Tests require only five env vars in .Renviron (or CI settings) as per [testing instructions](https://dbca-wa.github.io/ruODK/CONTRIBUTING.html#test)

Output from goodpractice:
> ✖ fix this R CMD check NOTE: installed size is 5.0Mb sub-directories of 1Mb or more: doc 4.4Mb

* Chunky /doc is due to included ODK diagram and attachment files of example data used in the vignettes (2 submissions with a couple of images, resized to 300x200px post download, and a repeating group).
* This is roughly the minimum complexity to demonstrate the data wrangling.
* If need be (maybe for CRAN/ODK Central 1.0), we could out an ruODK data package with the vignettes, and keep ruODK small. However unless that's absolutely necessary, keeping things together reduces maintenance complexity.

I think that's it! New things I've learned:
* .Renviron for env vars
* Grouping functions by `@family` in `pkgdown` using `has_concept()`
* Drop `usethis` boilerplate Author, Maintainer, and only use Authors@R in Description
* Drop `@importFrom` and only use DESCRIPTION Imports/Suggests etc and `proper::namespacing()`
* Clearer idea of setting up secrets for testing (see contributing - test)
