### 2019-09-10 #335 @maelle
Thanks a lot @florianm! I'll look for reviewers now. Note that once reviewers are recruited, we give them 3 weeks to complete their review, then there's time for your changes, then for their final response (and a continuing discussion) so I think your package will not be approved before the mentioned meetings, however you already get to display an rOpenSci "in review" badge, orange for now:

I'll hide part of our discussion to make the issue thread easier to catch up with for reviewers.
