### 2019-09-10 #335 @maelle
@florianm A further request actually, sorry, I had not googled "Open Data Kit R" before now because of the existence of the pre-submission inquiry. Doing so I found [`odkr`](https://github.com/validmeasures/odkr), could you please compare the functionalities of your and that package? Feel free to also include the other packages that are somehow related to ODK. An ideal comparison would include prose and a table. Example of comparison tables for [`rtweet`](https://rtweet.info/#package-functionality) and [`CoordinateCleaner`](https://ropensci.github.io/CoordinateCleaner/articles/Background_comparison_other_software.html). I'd like to know how `ruODK` is "best in class", and how the other packages either overlap or complement its functionality. Thank you!
