### 2019-09-10 dbca-wa/ruODK#24 @florianm
Format mentions of packages with curly braces
## Feature
`{ruODK}` instead of `ruODK'.

Goal: signify package names in prose.

https://github.com/ropensci/software-review/issues/335
