### 2019-09-10 dbca-wa/ruODK#25 @florianm
Compare ruODK with other R packages related to ODK or OData
## Feature
Source https://github.com/ropensci/software-review/issues/335#issuecomment-529793170
@maelle

### Examples

* https://ropensci.github.io/CoordinateCleaner/articles/Background_comparison_other_software.html
* https://rtweet.info/#package-functionality

### Candidates

#### ODK

* https://github.com/validmeasures/odkr
* https://cran.r-project.org/web/packages/odk/index.html
* https://github.com/nap2000/odkmeta and fork https://github.com/LloydOBanwart/odkmeta

#### OData

* https://github.com/HanOostdijk/odataR
* https://cran.r-project.org/web/packages/cbsodataR/
* https://cran.r-project.org/web/packages/OData/index.html
* https://www.cdata.com/kb/tech/odata-jdbc-r.rst

Note to self: chuck in a reference to https://www.odata.org/
