### 2019-09-10 19b0ced
Fix #25: add comparison to other R/ODK/OData packages to README
