+++ README.Rmd
@@ -243,8 +243,8 @@ There are several other R packages interacting with the ODK ecosystem, and/or
 | Elevator pitch               | "[ckanr](https://github.com/ropensci/ckanr) for ODK Central"  | "Drive ODK Briefcase through R" | "Export ODK Aggregate to SPSS" | "Export ODK Aggregate to STATA" | 
 | Last commit                  | Sep 2019      |  Sep 2019     |    Nov 2017    |  Jun 2014    |
 | Website                      | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) |    [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/)   | [rdrr.io](https://rdrr.io/cran/odk/) |    [GH](https://github.com/nap2000/odkmeta) |
-| Test coverage [%]            | 100 | 68 | 0 | 0 |
-| Working examples             | README, vignettes, pkgdown | README, pkgdown | CRAN PDF | README |
+| Test coverage [%]            | [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master) | [![codecov](https://codecov.io/gh/validmeasures/odkr/branch/master/graph/badge.svg)](https://codecov.io/gh/validmeasures/odkr) | ❌ | ❌ |
+| Working examples             | README, 3 vignettes, pkgdown | README, pkgdown | CRAN PDF | README |
 | Available on CRAN            | ❌ |  ✅ |  ✅ | ❌ |
 | Language/dialect             | tidyverse R | Base R | Base R | Stata |
 | External dependencies        | None | Java, ODK Briefcase | SPSS | Stata |
@@ -262,12 +262,13 @@ There are several other R packages interacting with the ODK ecosystem, and/or
 |------------------------------|---------------|---------------|---------------|---------------|--------------|
 | Elevator pitch               | "[ckanr](https://github.com/ropensci/ckanr) for ODK Central"  | "OData client for https://opendata.cbs.nl (and similar)" | "OData client for https://www.cbs.nl" | "Minimal OData example" | "Minimal RJDBC example" | 
 | Last commit                  | Sep 2019 | Jan 2018 | Aug 2019 | Dec 2016 | ❓ |
-| Website | GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) | [GH](https://github.com/HanOostdijk/odataR) | [CRAN](https://cran.r-project.org/web/packages/cbsodataR/) [GH](https://github.com/edwindj/cbsodataR) [pkgdown](https://edwindj.github.io/cbsodataR/) | OData | odatajdbcr |
-| Test coverage [%]            | 100 | 68 | 0 | 0 | 0 |
+| Website | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) | [GH](https://github.com/HanOostdijk/odataR) | [CRAN](https://cran.r-project.org/web/packages/cbsodataR/) [GH](https://github.com/edwindj/cbsodataR) [pkgdown](https://edwindj.github.io/cbsodataR/) | [CRAN](https://cran.r-project.org/web/packages/OData/index.html) | [cdata.com](https://www.cdata.com/kb/tech/odata-jdbc-r.rst) |
[…]
