+++ README.md
@@ -398,35 +398,36 @@ and/or
 
 ### Comparison of ODK related software packages (non-ODK core)
 
-| Package                         | [ruODK](https://dbca-wa.github.io/ruODK/)                                                                                    | [odkr](https://validmeasures.org/odkr/)                                            | [odk](https://cran.r-project.org/package=odk)                                      | [odkmeta](https://github.com/nap2000/odkmeta)                     |
-| ------------------------------- | ---------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------- | ----------------------------------------------------------------- |
-| Elevator pitch                  | “[ckanr](https://github.com/ropensci/ckanr) for ODK Central”                                                                 | “Drive ODK Briefcase through R”                                                    | “Export ODK Aggregate to SPSS”                                                     | “Export ODK Aggregate to STATA”                                   |
-| Last commit                     | Sep 2019                                                                                                                     | Sep 2019                                                                           | Nov 2017                                                                           | Jun 2014                                                          |
-| Website                         | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/)                                           | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) | [rdrr.io](https://rdrr.io/cran/odk/)                                               | [GH](https://github.com/nap2000/odkmeta)                          |
-| Test coverage \[%\]             | 100                                                                                                                          | 68                                                                                 | 0                                                                                  | 0                                                                 |
-| Working examples                | README, vignettes, pkgdown                                                                                                   | README, pkgdown                                                                    | CRAN PDF                                                                           | README                                                            |
-| Available on CRAN               | ❌                                                                                                                            | ✅                                                                                  | ✅                                                                                  | ❌                                                                 |
-| Language/dialect                | tidyverse R                                                                                                                  | Base R                                                                             | Base R                                                                             | Stata                                                             |
-| External dependencies           | None                                                                                                                         | Java, ODK Briefcase                                                                | SPSS                                                                               | Stata                                                             |
-| Affiliation                     | [DBCA WA](https://flatbacks.dbca.wa.gov.au/) / [ROpenSci prospectus](https://github.com/ropensci/software-review/issues/335) | [Valid Measues](https://github.com/validmeasures)                                  | [Muntashir-Al-Arefin](https://stackoverflow.com/users/8875690/muntashir-al-arefin) | [ODK core developer Matt White](https://github.com/matthew-white) |
-| Covers ODK Central OData API    | ✅                                                                                                                            | ❌                                                                                  | ❌                                                                                  | ❌                                                                 |
-| Covers ODK Central REST API     | ✅                                                                                                                            | ❌                                                                                  | ❌                                                                                  | ❌                                                                 |
-| Covers ODK Central bulk export  | ✅                                                                                                                            | ❌                                                                                  | ❌                                                                                  | ❌                                                                 |
-| Covers ODK Central OpenRosa API | ❌                                                                                                                            | ✅ via ODK Briefcase                                                                | ❌                                                                                  | ❌                                                                 |
-| Data post-processing            | ✅                                                                                                                            | ✅                                                                                  | ❌                                                                                  | ❌                                                                 |
-| Data visualisation examples     | ✅                                                                                                                            | ❌                                                                                  | ❌                                                                                  | ❌                                                                 |
[…]
