### 2019-09-10 dbca-wa/ruODK#25 @florianm
Nice one, will include. Looks like targeted towards Kobotoolbox, which is also based on ODK.
