### 2019-09-10 #335 @sckott
> Once ODK Central solidifies around v1.0, we can probably turn to using webmockr & vcr for caching test results.

i've written a section in the http testing book trying to help folks think through when to use webmockr vs. vcr vs. fakes vs. real requests https://ropenscilabs.github.io/http-testing-book/testing-considerations.html
