### 2019-09-11 #335 @florianm
@sckott cheers, I'll read those again to figure out whether/how I can prevent testing against possibly outdated API responses in case ODK Central changes.
