### 2019-09-11 #335 @florianm
> In addition to intermittent server side issues, your tests may be performing queries with cached (vcr) or mocked (webmockr) responses that are no longer valid with the current state of the remote service. It may be harmless, for example, the response to some query now returns no data because the data for that entity was removed. But it could be more serious in that the remove service changed their API such that an API route is no longer available or the route name has changed, or similar.

The ODK Central API will change, and could bring breaking changes, until ca release 1.0. Therefore I want tests to run against the real deal (the test server is ODK's own public sandbox, which is always latest master), so that ruODK has to keep up with ODK Central.

> Testing real HTTP interactions should be the slowest option, but has the benefit of not adding any (permanent) files to your package. Mocking tests can be very lite weight, though you can include very heavy responses.

The one thing making my tests slow is repeated download of a ca 50MB zip file, both for testing the insides (which requires some form complexity as discussed above), which could change, and testing handling repeated dl/skip logic. The above reads as that response cached in vcr would be added to the installed package which will probably make CMD check very unhappy.
The best approach for ruODK might be to add a second test project with a smaller export size (maybe a few records and only one (small) file attachment) for those scenarios purely to speed up tests. ETA for that is early next week. Hope this does not hold up the submission process.
