### 2019-09-11 fed8331
#25 add koboloadeR to comparison, add evaluation summaries
