+++ README.Rmd
@@ -242,10 +242,10 @@ There are several other R packages interacting with the ODK ecosystem, and/or
 |------------------------------|---------------|---------------|---------------|---------------|---------------|
 | Elevator pitch               | "[ckanr](https://github.com/ropensci/ckanr) for ODK Central"  | "Drive ODK Briefcase through R" | "Export ODK Aggregate to SPSS" | "Export ODK Aggregate to STATA" | "Metapackage for extended ODK ecosystem" |
 | Last commit                  | Sep 2019      |  Sep 2019     |    Nov 2017    |  Jun 2014    | Jul 2019 |
-| Website                      | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) |    [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/)   | [rdrr.io](https://rdrr.io/cran/odk/) |    [GH](https://github.com/nap2000/odkmeta) | [GH](https://github.com/unhcr/koboloadeR) [pkgdown](https://unhcr.github.io/koboloadeR/docs/index.html) |
-| Test coverage             | [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master) | [![codecov](https://codecov.io/gh/validmeasures/odkr/branch/master/graph/badge.svg)](https://codecov.io/gh/validmeasures/odkr) | ❌ | ❌ | [![Travis build status](https://travis-ci.org/unhcr/koboloadeR.svg?branch=gh-pages)](https://travis-ci.org/unhcr/koboloadeR) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/unhcr/koboloadeR?branch=gh-pages&svg=true)](https://ci.appveyor.com/project/unhcr/koboloadeR) |
-| Working examples             | README, 3 vignettes, pkgdown | README, pkgdown | CRAN PDF | README README, 9 vignettes, shiny apps, pkgdown |
-| Available on CRAN            | ❌ |  ✅ |  ✅ | ❌ | ❌  |
+| Website                      | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/)   | [rdrr.io](https://rdrr.io/cran/odk/) |    [GH](https://github.com/nap2000/odkmeta) | [GH](https://github.com/unhcr/koboloadeR) [pkgdown](https://unhcr.github.io/koboloadeR/docs/index.html) |
[…]
