### 2019-09-11 #335 @sckott
> Therefore I want tests to run against the real deal

makes sense

>

I don't think the large size of the cassettes (cached http responses) affects r check, e.g., the cassettes directory in this package https://github.com/ropensci/rgbif/ is 32 MB, but running check on it throws no notes/warnings/etc - But i could be wrong -

if size of cassettes is a concern, one option is to skip tests on CRAN for which cassettes are too large to include in the pkg, and those cassette files added to .Rbuildignore
