### 2019-09-14 dbca-wa/ruODK#27 @florianm
Use stable tidyr 1.0.0
## Feature
Tidyr turned [1.0.0](https://www.tidyverse.org/articles/2019/09/tidyr-1-0-0/)!
This means that `{ruODK}` can depend on stable CRAN tidyr>=1.0.0 instead of dev version 0.8.3.9000 to have access to `tidyr::unnest_wider()`.

This will make installation of latest master of tidyr and its dependencies less frequent and save reviewers of https://github.com/ropensci/software-review/issues/335 some time.
