### 2019-09-14 #335 @florianm
Absolutely! The Rmd templates are only additional nice-to-haves, and mentioning the Pentaho Kettle tutorial will only impact the README.
