### 2019-09-14 c7b4543
Fix #24, fix #27, fix #28: tidyr 1.0.0, Kettle tutorial, formatting
