+++ README.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![ROpenSci submission status](https://badges.ropensci.org/335_status.svg)](https://github.com/ropensci/software-review/issues/335)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![GitHub issues](https://img.shields.io/github/issues/dbca-wa/ruodk.svg?style=popout)](https://github.com/dbca-wa/ruODK/issues)
+[![GitHub issues](https://img.shields.io/github/issues/dbca-wa/ruodk.svg?style=popout)](https://github.com/dbca-wa/ruODK/issues/)
 [![Travis build status](https://travis-ci.org/dbca-wa/ruODK.svg?branch=master)](https://travis-ci.org/dbca-wa/ruODK)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/dbca-wa/ruODK?branch=master&svg=true)](https://ci.appveyor.com/project/dbca-wa/ruODK)
 [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master)
@@ -28,7 +28,7 @@ Especially in these trying times, it is important to ask: "ruODK?"
 for collecting, managing, and using data in resource-constrained environments.
 
 ODK consists of a range of [software packages and apps](https://opendatakit.org/software/).
-`ruODK` assumes some familiarity of its users with the ODK ecosystem and workflows.
+`{ruODK}` assumes some familiarity of its users with the ODK ecosystem and workflows.
 For a detailed overview, read the extensive [ODK documentation](https://docs.opendatakit.org/)
 and visit the friendly [ODK forum](https://forum.opendatakit.org/).
 
@@ -58,7 +58,7 @@ access to its data and functionality through both an OData and a RESTful API
 with a comprehensive and interactive 
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
 
-`ruODK` is aimed at the technically minded researcher who wishes to access and
+`{ruODK}` is aimed at the technically minded researcher who wishes to access and
 use the data from ODK Central using the programming language R.
 
 Benefits of using the R ecosystem in combination with ODK:
@@ -74,7 +74,7 @@ Benefits of using the R ecosystem in combination with ODK:
   workflow scaling ([drake](https://docs.ropensci.org/drake/)) and a range of
   integrations with Docker (links coming soon).
 
-`ruODK`'s scope:
+`{ruODK}`'s scope:
 
 * To wrap all ODK Central API endpoints with a focus on **data access**.
 * To provide working examples of interacting with the ODK Central API.
@@ -82,7 +82,7 @@ Benefits of using the R ecosystem in combination with ODK:
   ODK Central data in R: **data munging** the ODK Central API output into tidy 
   R formats.
   
-`ruODK`'s use cases:
+`{ruODK}`'s use cases:
 
 * Smaller projects:
   1. Data collection: ODK Collect
@@ -124,7 +124,7 @@ remotes::install_github("dbca-wa/ruODK", dependencies = TRUE)
 ### Access to an ODK Central instance
 First, we need an ODK Central instance and some data to play with!
 
-Either ask in the [ODK forum] for an account in the 
[…]
