+++ README.md
@@ -11,7 +11,7 @@ status](https://badges.ropensci.org/335_status.svg)](https://github.com/ropensci
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![GitHub
-issues](https://img.shields.io/github/issues/dbca-wa/ruodk.svg?style=popout)](https://github.com/dbca-wa/ruODK/issues)
+issues](https://img.shields.io/github/issues/dbca-wa/ruodk.svg?style=popout)](https://github.com/dbca-wa/ruODK/issues/)
 [![Travis build
 status](https://travis-ci.org/dbca-wa/ruODK.svg?branch=master)](https://travis-ci.org/dbca-wa/ruODK)
 [![AppVeyor build
@@ -27,7 +27,7 @@ software for collecting, managing, and using data in
 resource-constrained environments.
 
 ODK consists of a range of [software packages and
-apps](https://opendatakit.org/software/). `ruODK` assumes some
+apps](https://opendatakit.org/software/). `{ruODK}` assumes some
 familiarity of its users with the ODK ecosystem and workflows. For a
 detailed overview, read the extensive [ODK
 documentation](https://docs.opendatakit.org/) and visit the friendly
@@ -61,7 +61,7 @@ provides access to its data and functionality through both an OData and
 a RESTful API with a comprehensive and interactive
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
 
-`ruODK` is aimed at the technically minded researcher who wishes to
+`{ruODK}` is aimed at the technically minded researcher who wishes to
 access and use the data from ODK Central using the programming language
 R.
 
@@ -80,7 +80,7 @@ Benefits of using the R ecosystem in combination with ODK:
     ([drake](https://docs.ropensci.org/drake/)) and a range of
     integrations with Docker (links coming soon).
 
-`ruODK`’s scope:
+`{ruODK}`’s scope:
 
   - To wrap all ODK Central API endpoints with a focus on **data
     access**.
@@ -89,7 +89,7 @@ Benefits of using the R ecosystem in combination with ODK:
     with ODK Central data in R: **data munging** the ODK Central API
     output into tidy R formats.
 
-`ruODK`’s use cases:
+`{ruODK}`’s use cases:
 
   - Smaller projects:
     1.  Data collection: ODK Collect
@@ -136,7 +136,8 @@ remotes::install_github("dbca-wa/ruODK", dependencies = TRUE)
 
 First, we need an ODK Central instance and some data to play with\!
 
-Either ask in the \[ODK forum\] for an account in the [ODK Central
+Either ask in the [ODK forum](https://forum.opendatakit.org/) for an
+account in the [ODK Central
 Sandbox](https://sandbox.central.opendatakit.org/), or follow the [setup
 instructions](https://docs.opendatakit.org/central-intro/) to build and
 deploy your very own ODK Central instance.
@@ -165,9 +166,9 @@ forms](https://github.com/dbca-wa/ruODK/tree/master/inst/extdata): The
 Build](https://build.opendatakit.org/), while the `.xml` versions can be
 imported into ODK Central.
 
-## Configure `ruODK`
+## Configure `{ruODK}`
 
-Set up `ruODK` with an OData Service URL and credentials of a
+Set up `{ruODK}` with an OData Service URL and credentials of a
 read-permitted ODK Central web user.
 
 ``` r
@@ -178,7 +179,7 @@ ruODK::ru_setup(
 )
 ```
 
-For all available detailed options to configure `ruODK`, read
+For all available detailed options to configure `{ruODK}`, read
 `vignette("Setup", package = "ruODK")` (online
[…]
