### 2019-09-14 d9ccc28
Fix typos, add account request to contributing guide
