### 2019-09-16 c204426
Fix #26 add Rmd template
