+++ inst/rmarkdown/templates/odata/skeleton/skeleton.Rmd
@@ -1,0 +1,211 @@
+---
+title: "Untitled"
+output:
+  html_document: default
+  pdf_document:
+    latex_engine: xelatex
+  word_document: default
+---
+
+<!--  
+This Rmd template provides an example workflow to access submissions and 
+attachments of a given ODK Central form using its OData API.
+
+
+You will need:
+* An ODK Central form's OData Service URL (ending in .svc), accessed from the 
+  "Analyse using OData" button on the form submission tab.
+* Username and password of an ODK Central webuser authorised to view this form.
+-->
+
+<!--
+Step 1: Setup ruODK
+ODK Central webuser username and password are set as environment variables to
+keep them outside of this document.
+
+Read vignette "setup" for further information.
+-->
+<!-- Open the .Renviron file  -->
+```{r edit_renviron, echo=FALSE, eval=FALSE}
+usethis::edit_r_environ()
+```
+
+<!-- and paste the following with the webuser's credentials  -->
+```{r paste_env_vars, echo=FALSE, eval=FALSE}
+ODKC_UN="..."
+ODKC_PW="..."
+```
+<!-- and restart the session to load these environment variables. -->
+
+<!-- 
+* Configure ODK Central base URL, project ID, and form ID through OData Service URL.
+  Paste the OData Service URL over `SVC_URL`.
+* Paste your local timezone over `Australia/Perth`.
+* Adjust `loc` to a local subfolder for downloaded attachments.
+-->
+```{r setup, message=FALSE, warning=FALSE}
+library(knitr)
+library(dplyr)
+library(ckanr)
+library(ruODK)
+
+ruODK::ru_setup(svc="SVC_URL")
+tz <- "Australia/Perth"
+loc <- fs::path("attachments", "media")
+fs::dir_create(loc)
+```
+
+# Download data
+<!-- 
+Get table names. 
+
+form_tables will contain one row per separate table. 
+-->
+```{r svc_get}
+form_tables <- ruODK::odata_service_get()
+form_tables %>% knitr::kable(.)
+```
+
+<!-- 
+Get and parse submissions.
+
+* Rename any point coordinates which come out as `xn`, `xn+1`, `xn+2` into 
+  longitude, latitude, altitude.
+* Repeat parse_datetime for each form field name that symbolises a datetime.
+  Default assumes "time", example shows an additional step to also parse any
+  form field with "date" in its name.
+
+
+Repeat for each table name in form_tables$url. 
+-->
+```{r dl_submissions}
+data_main_raw <- ruODK::odata_submission_get()
+data_sub1_raw <- ruODK::odata_submission_get(table = "Submissions.nested_table_name1")
+
+data <- data_main_raw %>% 
+  ruODK::odata_submission_parse(verbose = TRUE) %>%
+  ruODK::parse_datetime(tz = tz) %>% 
+  ruODK::parse_datetime(tz = tz, col_contains = "date") %>%
+  dplyr::rename(
+    # longitude = x13,
+    # latitude = x14,
+    # altitude = x15
+  ) %>%
+  dplyr::mutate(
+    my_photo = attachment_get(id, my_photo, local_dir = loc),
+    another_photo = attachment_get(id, another_photo, local_dir = loc)
+    # add for each form field of type media attachment
+  ) %>%
+  invisible()
+
+data_sub1 <- data_sub1_raw %>% 
+  ruODK::odata_submission_parse(verbose = TRUE) %>%
+  ruODK::parse_datetime(tz = tz) %>% 
+  ruODK::parse_datetime(tz = tz, col_contains = "date") %>%
+  dplyr::mutate(
+    some_photo = attachment_get(id, some_photo, local_dir = loc),
[…]
