+++ inst/rmarkdown/templates/odata/template.yaml
@@ -1,0 +1,3 @@
+name: odata
+description: >
+ Accessing ODK Central through the OData API

[…]
