### 2019-09-17 #335 @florianm
@maelle @karissawhiting also soliciting feedback from the ODK community in the ODK Forum [here](https://forum.opendatakit.org/t/seeking-your-feedback-ruodk/22160).
