### 2019-09-17 40fc5b1
add install.R for binder
