+++ README.Rmd
@@ -20,6 +20,7 @@ knitr::opts_chunk$set(
 [![Travis build status](https://travis-ci.org/dbca-wa/ruODK.svg?branch=master)](https://travis-ci.org/dbca-wa/ruODK)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/dbca-wa/ruODK?branch=master&svg=true)](https://ci.appveyor.com/project/dbca-wa/ruODK)
 [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/ruODK/master?urlpath=rstudio)
 <!-- badges: end -->
 
 Especially in these trying times, it is important to ask: "ruODK?"
@@ -252,7 +253,7 @@ There are several other R packages interacting with the ODK ecosystem, and/or
 | Covers ODK Central OData API        | ✅  |  ❌ |  ❌  | ❌  | ❌ | ✅  |
 | Covers ODK Central REST API         | ✅  |  ❌ |  ❌  | ❌  | ❌ | ❌  |
 | Covers ODK Central bulk export      | ✅  |  ❌ |  ❌  | ❌  | ❌ | ✅  |  
-| Covers ODK Central OpenRosa API     | ❌  |  ✅  via ODK Briefcase |  ❌  | ❌  | ✅  via ODK Briefcase  | ✅  |  
+| Covers ODK Central OpenRosa API     | ❌  no need, gets all data through OData/REST API |  ✅  via ODK Briefcase |  ❌  | ❌  | ✅  via ODK Briefcase  | ✅  |  
 | Data post-processing                | ✅  |  ✅ |  ❌  | ❌  | ✅ | ✅  |  
 | Data visualisation examples         | ✅  |  ❌ | ❌  | ❌  | ✅  | ❌ |  
  
