### 2019-09-17 3b1e08a
Add code quality badge, fix minor formatting suggestions
