+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
-# Non-technical contributions to ruODK
+## Non-technical contributions to ruODK
 Feel free to [report issues](https://github.com/dbca-wa/ruODK/issues):
 
 * Bug reports are for unplanned malfunctions.
@@ -6,25 +6,25 @@ Feel free to [report issues](https://github.com/dbca-wa/ruODK/issues):
 * Account requests are for getting access to the ODK Central instances run by DBCA
   (DBCA business only) or the public demo server (contributors, to run tests).
 
-# Technical contributions to ruODK
+## Technical contributions to ruODK
 
 If you would like to contribute to the code base, follow the process below.
 Note, this process is adjusted from the `usethis` tidyverse boilerplate.
 
--   [Prerequisites](#prerequisites)
--   [PR Process](#pr-process)
-    -   [Fork, clone, branch](#fork-clone-branch)
-    -   [Check](#check)
-    -   [Style](#style)
-    -   [Document](#document)
-    -   [Test](#test)
-    -   [NEWS](#news)
-    -   [Re-check](#re-check)
-    -   [Commit](#commit)
-    -   [Push and pull](#push-and-pull)
-    -   [Review, revise, repeat](#review-revise-repeat)
--   [Resources](#resources)
--   [Code of Conduct](#code-of-conduct)
+*  [Prerequisites](#prerequisites)
+*  [PR Process](#pr-process)
+  *  [Fork, clone, branch](#fork-clone-branch)
+  *  [Check](#check)
+  *  [Style](#style)
+  *  [Document](#document)
+  *  [Test](#test)
+  *  [NEWS](#news)
+  *  [Re-check](#re-check)
+  *  [Commit](#commit)
+  *  [Push and pull](#push-and-pull)
+  *  [Review, revise, repeat](#review-revise-repeat)
+*   [Resources](#resources)
+*   [Code of Conduct](#code-of-conduct)
 
 This explains how to propose a change to ruODK via a pull request using
 Git and GitHub. 
@@ -32,22 +32,22 @@ Git and GitHub.
 For more general info about contributing to `ruODK`, see the 
 [Resources](#resources) at the end of this document.
 
-## Prerequisites
+### Prerequisites
 
 Before you do a pull request, you should always file an issue and make sure
 someone from the tidyverse team agrees that it’s a problem, and is happy with
 your basic proposal for fixing it. If you’ve found a bug, first create a minimal
 [reprex](https://www.tidyverse.org/help/#reprex).
 
-## PR process
+### PR process
 
-### Fork, clone, branch
+#### Fork, clone, branch
 
 The first thing you'll need to do is to [fork](https://help.github.com/articles/fork-a-repo/) 
 the [ruODK GitHub repo](https://github.com/dbca-wa/ruODK), and 
 then clone it locally. We recommend that you create a branch for each PR.
 
-### Check
+#### Check
 
 Before changing anything, make sure the package still passes `R CMD check`
 locally for you. When in doubt, compare your `R CMD check` results with current
@@ -60,7 +60,7 @@ pinpoint any problems introduced by your changes.
 devtools::check()
 ```
 
-### Style
+#### Style
 
 Match the existing code style. This means you should follow the tidyverse 
 [style guide](http://style.tidyverse.org). Use the 
@@ -75,7 +75,7 @@ better to file an issue or a separate PR to fix that first.
 styler::style_pkg()
 ```
 
-### Document
+#### Document
 
 We use [roxygen2](https://cran.r-project.org/package=roxygen2), specifically with the 
 [Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/markdown.html),
@@ -93,7 +93,7 @@ spelling::update_wordlist()
[…]
