+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -13,10 +13,8 @@ assignees: florianm
 ## Reproducible example
 <!-- If this issue refers to a bug or unexpected behaviour of ruODK, please include a minimal reproducible example ([reprex](https://reprex.tidyverse.org/), see also <https://www.tidyverse.org/help/#reprex>). -->
 
-
 ```{r}
 # insert reprex here
-
 ```
 
 <details>
@@ -24,6 +22,5 @@ assignees: florianm
   
 ```{r}
 # utils::sessionInfo()
-
 ```
 </details>
