### 2019-09-17 9f2beeb
Add Rmd dependencies to binder install.R
