+++ README.Rmd
@@ -9,8 +9,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.path = "man/figures/",
-  out.width = "100%"
-)
+  out.width = "100%")
 ```
 # `ruODK`: An R Client for the ODK Central API <img src="man/figures/ruODK.png" align="right" alt="Are you ODK?" width="120" />
 <!-- badges: start -->
@@ -21,7 +20,6 @@ knitr::opts_chunk$set(
 [![Travis build status](https://travis-ci.org/dbca-wa/ruODK.svg?branch=master)](https://travis-ci.org/dbca-wa/ruODK)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/dbca-wa/ruODK?branch=master&svg=true)](https://ci.appveyor.com/project/dbca-wa/ruODK)
 [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master)
-[![Codacy Badge](https://api.codacy.com/project/badge/Grade/b4814112362445dfb32fc08664aadb43)](https://www.codacy.com/manual/florianm/ruODK?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=dbca-wa/ruODK&amp;utm_campaign=Badge_Grade)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/ruODK/master?urlpath=rstudio)
 <!-- badges: end -->
 
