### 2019-09-17 58f99aa
add bindr dependencies
