### 2019-09-18 4eae18b
Improve Rmd template
