+++ inst/rmarkdown/templates/odata/skeleton/skeleton.Rmd
@@ -1,29 +1,32 @@
 ---
-title: "Untitled"
+title: "Report Title"
+author: "Report Author"
+date: "`r Sys.Date()`"
 output:
-  html_document: default
+  html_document:
+    toc = TRUE
+    fig_width = 9
+    code = FALSE
   pdf_document:
     latex_engine: xelatex
   word_document: default
 ---
-
 <!--  
 This Rmd template provides an example workflow to access submissions and 
-attachments of a given ODK Central form using its OData API.
-
+attachments of a given ODK Central form using its OData API and disseminate
+data, attachments, and this report via CKAN or Google Drive.
 
 You will need:
 * An ODK Central form's OData Service URL (ending in .svc), accessed from the 
   "Analyse using OData" button on the form submission tab.
 * Username and password of an ODK Central webuser authorised to view this form.
+* Optional: base URL and write-permitted API key for a CKAN instance.
+* Optional: a Google account and the means to authenticate (e.g. phone for 2FA).
 -->
 
 <!--
-Step 1: Setup ruODK
-ODK Central webuser username and password are set as environment variables to
-keep them outside of this document.
-Optional: CKAN URL and API key.
-
+Step 1: Setup ruODK, ckanr
+Secrets are set as environment variables to keep them outside of this document. 
 Read vignette "setup" for further information.
 -->
 <!-- Open the .Renviron file  -->
@@ -39,15 +42,17 @@ ODKC_PW="..."
 CKAN_URL="https://demo.ckan.org"
 CKAN_KEY="my-ckan-api-key"
 ```
-<!-- and restart the session to load these environment variables. -->
+<!-- and restart the R session to load these environment variables. -->
 
 <!-- 
 * Configure ODK Central base URL, project ID, and form ID through OData Service URL.
-  Paste the OData Service URL over `SVC_URL`.
+  Paste the (non-sensitive) OData Service URL over `SVC_URL`.
 * Paste your local timezone over `Australia/Perth`.
 * Adjust `loc` to a local subfolder for downloaded attachments.
 -->
 ```{r setup, message=FALSE, warning=FALSE}
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
+
 library(knitr)
 library(dplyr)
 library(ckanr)
@@ -60,11 +65,7 @@ fs::dir_create(loc)
 ```
 
 # Download data
-<!-- 
-Get table names. 
-
-form_tables will contain one row per separate table. 
--->
+<!-- Get table names. `form_tables` will contain one row per separate table. -->
 ```{r svc_get}
 form_tables <- ruODK::odata_service_get()
 form_tables %>% knitr::kable(.)
@@ -83,14 +84,16 @@ Get and parse submissions.
 Repeat for each table name in form_tables$url. 
 -->
 ```{r dl_submissions}
-data_main_raw <- ruODK::odata_submission_get()
-data_sub1_raw <- ruODK::odata_submission_get(table = "Submissions.nested_table_name1")
+data_main_raw <- ruODK::odata_submission_get(table = form_tables$url[1])
+data_sub1_raw <- ruODK::odata_submission_get(table = form_tables$url[2])
+# repeat for any remaining nested tables
 
 data <- data_main_raw %>% 
   ruODK::odata_submission_parse(verbose = TRUE) %>%
   ruODK::parse_datetime(tz = tz) %>% 
   ruODK::parse_datetime(tz = tz, col_contains = "date") %>%
   dplyr::rename(
+    # # Inspect data, find colnames of coordinates (if any):
     # longitude = x13,
     # latitude = x14,
     # altitude = x15
[…]
