### 2019-09-18 #335 @florianm
Quick note on failing check: My vignette "api" fails to build on 2 out of 3 envs. Methinks the file attachments on ODK Central's sandbox might have suffered damage. This is a good moment to switch over to a similar but newer example form, and populate that form with submissions with smaller file attachments (set the data collection device camera to take 640px photos).

TLDR https://github.com/dbca-wa/ruODK/issues/23 will likely fix tests, ETA later today if things go to plan.
