### 2019-09-18 ecdcca3
Fix #23: use newer form with fresh submissions for tests, vignettes, and included data
