+++ R/attachment_link.R
@@ -39,8 +39,10 @@ prefix_fn <- function(fn, prefix) {
 #' t <- tempdir()
 #' # Set default credentials, see vignette "setup"
 #' ruODK::ru_setup(
-#'   svc = paste0("https://sandbox.central.opendatakit.org/v1/projects/14/",
-#'                "forms/build_Flora-Quadrat-0-2_1558575936.svc"),
+#'   svc = paste0(
+#'     "https://sandbox.central.opendatakit.org/v1/projects/14/",
+#'     "forms/build_Flora-Quadrat-0-2_1558575936.svc"
+#'   ),
 #'   un = "me@email.com",
 #'   pw = "..."
 #' )
