+++ man/fq_raw_strata.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{fq_raw_strata}
+\alias{fq_raw_strata}
+\title{OData submission data for a subgroup of an ODK Central form.}
+\format{A list of lists}
+\source{
+See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
+}
+\usage{
+fq_raw_strata
+}
+\description{
+The OData response for the subgroup of an ODK Central form.
+}
+\details{
+This subgroup represents vegetation strata as per the NVIS classification.
+A vegetation stratum is a layer of plants with the same height, and dominated
+by one or few plant taxa. Plant communities can be made of up to five strata,
+with two to three being most common.
+
+The data in this example is freely made up.
+}
+\seealso{
+Other included: \code{\link{fq_meta}},
+  \code{\link{fq_raw_taxa}}, \code{\link{fq_raw}},
+  \code{\link{fq_svc}}
+}
+\concept{included}
+\keyword{datasets}
