### 2019-09-18 e718fe8
Decrease required DT version from 0.9 to 0.8
