+++ DESCRIPTION
@@ -33,7 +33,7 @@ Imports:
     xml2 (>= 1.2.2)
 Suggests: 
     covr (>= 3.3.1),
-    DT (>= 0.9),
+    DT (>= 0.8),
     janitor (>= 1.2.0),
     knitr (>= 1.24),
     ggplot2 (>= 3.2.1),
