### 2019-09-18 e498de4
Fix broken test-submission_get
