+++ tests/testthat/test-submission_get.R
@@ -23,15 +23,21 @@ test_that("submission_get works", {
   # The details for one submission return exactly one row
   testthat::expect_equal(length(sub), 15)
 
-  # The columns are metadata, plus the submission data in column 'xml`
+  # The columns are form fields or groups,
+  # plus the submission data in column 'xml`
   # names(sub)
   cn <- c(
-    "meta", "encounter_start_datetime", "reporter", "device_id", "location",
-    "habitat", "vegetation_stratum", "vegetation_stratum", "perimeter",
-    "taxon_encounter", "taxon_encounter", "taxon_encounter", "taxon_encounter",
+    "meta","encounter_start_datetime", "reporter", "device_id", "location",
+    "habitat", "vegetation_stratum", "vegetation_stratum", "vegetation_stratum",
+    "vegetation_stratum", "vegetation_stratum", "perimeter", "taxon_encounter",
     "taxon_encounter", "encounter_end_datetime"
   )
   testthat::expect_equal(names(sub), cn)
+  testthat::expect_true("encounter_start_datetime" %in% names(sub))
+  testthat::expect_true("perimeter" %in% names(sub))
+  testthat::expect_true("vegetation_stratum" %in% names(sub))
+  testthat::expect_true("taxon_encounter" %in% names(sub))
+  testthat::expect_true("encounter_end_datetime" %in% names(sub))
 })
 
 
