+++ README.Rmd
@@ -29,7 +29,7 @@ Especially in these trying times, it is important to ask: "ruODK?"
 for collecting, managing, and using data in resource-constrained environments.
 
 ODK consists of a range of [software packages and apps](https://opendatakit.org/software/).
-`{ruODK}` assumes some familiarity of its users with the ODK ecosystem and workflows.
+`ruODK` assumes some familiarity of its users with the ODK ecosystem and workflows.
 For a detailed overview, read the extensive [ODK documentation](https://docs.opendatakit.org/)
 and visit the friendly [ODK forum](https://forum.opendatakit.org/).
 
@@ -59,7 +59,7 @@ access to its data and functionality through both an OData and a RESTful API
 with a comprehensive and interactive 
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
 
-`{ruODK}` is aimed at the technically minded researcher who wishes to access and
+`ruODK` is aimed at the technically minded researcher who wishes to access and
 use the data from ODK Central using the programming language R.
 
 Benefits of using the R ecosystem in combination with ODK:
@@ -75,7 +75,7 @@ Benefits of using the R ecosystem in combination with ODK:
    workflow scaling ([drake](https://docs.ropensci.org/drake/)) and a range of
    integrations with Docker (links coming soon).
 
-`{ruODK}`'s scope:
+`ruODK`'s scope:
 
 *  To wrap all ODK Central API endpoints with a focus on **data access**.
 *  To provide working examples of interacting with the ODK Central API.
@@ -83,7 +83,7 @@ Benefits of using the R ecosystem in combination with ODK:
    ODK Central data in R: **data munging** the ODK Central API output into tidy 
    R formats.
   
-`{ruODK}`'s use cases:
+`ruODK`'s use cases:
 
 *  Smaller projects:
    1. Data collection: ODK Collect
@@ -110,7 +110,7 @@ Out of scope:
    working on the "management" API endpoints of the ODK Central backend.
 *  To provide extensive data visualisation. We show only minimal examples of data 
    visualisation and presentation, mainly to illustrate the example data.
-   Once the data is in your hands as tidy tibbles... u r ODK!
+   Once the data is in your hands as tidy tibbles... urODK!
 
 ## Install
 
@@ -147,8 +147,8 @@ A note on the [included example forms](https://github.com/dbca-wa/ruODK/tree/mas
 The `.odkbuild` versions can be loaded into [ODK Build](https://build.opendatakit.org/), 
 while the `.xml` versions can be imported into ODK Central.
 
-## Configure `{ruODK}`
-Set up `{ruODK}` with an OData Service URL and credentials of a read-permitted
+## Configure `ruODK`
+Set up `ruODK` with an OData Service URL and credentials of a read-permitted
 ODK Central web user.
   
 ```{r ru_setup}
@@ -159,7 +159,7 @@ ruODK::ru_setup(
 )
 ```
 
-For all available detailed options to configure `{ruODK}`, read 
+For all available detailed options to configure `ruODK`, read 
 `vignette("Setup", package = "ruODK")` 
 (online [here](https://dbca-wa.github.io/ruODK/articles/setup.html)).
 
@@ -200,6 +200,15 @@ See also `vignette("api", package="ruODK")`
 (online [here](https://dbca-wa.github.io/ruODK/articles/api.html))
 for examples using the alternative RESTful API.
 
+## Try ruODK
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio) 
+will launch a disposable, hosted RStudio instance with `ruODK` installed and
[…]
