+++ README.md
@@ -29,7 +29,7 @@ software for collecting, managing, and using data in
 resource-constrained environments.
 
 ODK consists of a range of [software packages and
-apps](https://opendatakit.org/software/). `{ruODK}` assumes some
+apps](https://opendatakit.org/software/). `ruODK` assumes some
 familiarity of its users with the ODK ecosystem and workflows. For a
 detailed overview, read the extensive [ODK
 documentation](https://docs.opendatakit.org/) and visit the friendly
@@ -63,7 +63,7 @@ provides access to its data and functionality through both an OData and
 a RESTful API with a comprehensive and interactive
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
 
-`{ruODK}` is aimed at the technically minded researcher who wishes to
+`ruODK` is aimed at the technically minded researcher who wishes to
 access and use the data from ODK Central using the programming language
 R.
 
@@ -82,7 +82,7 @@ Benefits of using the R ecosystem in combination with ODK:
     ([drake](https://docs.ropensci.org/drake/)) and a range of
     integrations with Docker (links coming soon).
 
-`{ruODK}`’s scope:
+`ruODK`’s scope:
 
   - To wrap all ODK Central API endpoints with a focus on **data
     access**.
@@ -91,7 +91,7 @@ Benefits of using the R ecosystem in combination with ODK:
     with ODK Central data in R: **data munging** the ODK Central API
     output into tidy R formats.
 
-`{ruODK}`’s use cases:
+`ruODK`’s use cases:
 
   - Smaller projects:
     1.  Data collection: ODK Collect
@@ -121,7 +121,7 @@ Out of scope:
   - To provide extensive data visualisation. We show only minimal
     examples of data visualisation and presentation, mainly to
     illustrate the example data. Once the data is in your hands as tidy
-    tibbles… u r ODK\!
+    tibbles… urODK\!
 
 ## Install
 
@@ -168,9 +168,9 @@ forms](https://github.com/dbca-wa/ruODK/tree/master/inst/extdata): The
 Build](https://build.opendatakit.org/), while the `.xml` versions can be
 imported into ODK Central.
 
-## Configure `{ruODK}`
+## Configure `ruODK`
 
-Set up `{ruODK}` with an OData Service URL and credentials of a
+Set up `ruODK` with an OData Service URL and credentials of a
 read-permitted ODK Central web user.
 
 ``` r
@@ -181,7 +181,7 @@ ruODK::ru_setup(
 )
 ```
 
-For all available detailed options to configure `{ruODK}`, read
+For all available detailed options to configure `ruODK`, read
 `vignette("Setup", package = "ruODK")` (online
 [here](https://dbca-wa.github.io/ruODK/articles/setup.html)).
 
@@ -308,7 +308,19 @@ available in the `vignette("odata", package="ruODK")` (online
 
 See also `vignette("api", package="ruODK")` (online
 [here](https://dbca-wa.github.io/ruODK/articles/api.html)) for examples
-using the alternative RESTful API.
+using the alternative RESTful
+API.
+
+## Try ruODK
+
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
+will launch a disposable, hosted RStudio instance with `ruODK` installed
+and the companion package [`urODK`](https://github.com/dbca-wa/urODK)
+opened as starting point for a hands-on workshop or instant demo of
+`ruODK` usage.
+
+Create a new RMarkdown workbook from `ruODK` template “odata” and follow
+the instructions within.
 
 ## Contribute
 
[…]
