### 2019-09-18 5ca4bb1
README: more badges, less cruft in comparison for cleaner look
