+++ README.md
@@ -128,7 +128,7 @@ Out of scope:
 You can install ruODK from GitHub with:
 
 ``` r
-# install.packages("devtools")
+if (!requireNamespace("remotes")) {install.packages("remotes")}
 remotes::install_github("dbca-wa/ruODK", dependencies = TRUE)
 ```
 
@@ -382,32 +382,32 @@ and/or
 
 ### Comparison of ODK related software packages (non-ODK core)
 
-| Package                         | [`ruODK`](https://dbca-wa.github.io/ruODK/)                                                                                                                                                                                                                                                                                                                                                                                            | [`{odkr}`](https://validmeasures.org/odkr/)                                                                                    | [`{odk}`](https://cran.r-project.org/package=odk)                                               | [`odkmeta`](https://github.com/nap2000/odkmeta)                   | [`{koboloadeR}`](https://unhcr.github.io/koboloadeR/docs/index.html)                                                                                                                                                                                                                                                                                                                                                                 | [Pentaho Kettle tutorial](https://github.com/schemetrica/automating-data-delivery-odk-central) |
-| ------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------ | ----------------------------------------------------------------------------------------------- | ----------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ | ---------------------------------------------------------------------------------------------- |
-| Elevator pitch                  | “[ckanr](https://github.com/ropensci/ckanr) for ODK Central”                                                                                                                                                                                                                                                                                                                                                                           | “Drive ODK Briefcase through R”                                                                                                | “Export ODK Aggregate to SPSS”                                                                  | “Export ODK Aggregate to STATA”                                   | “Metapackage for extended ODK ecosystem”                                                                                                                                                                                                                                                                                                                                                                                             | “What ruODK does, but as GUI”                                                                  |
-| Last commit                     | Sep 2019                                                                                                                                                                                                                                                                                                                                                                                                                               | Sep 2019                                                                                                                       | Nov 2017                                                                                        | Jun 2014                                                          | Jul 2019                                                                                                                                                                                                                                                                                                                                                                                                                             | Sept 2019                                                                                      |
[…]
