+++ README.Rmd
@@ -1,7 +1,6 @@
 ---
 output: github_document
 ---
-
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
 ```{r setup, include = FALSE}
@@ -117,7 +116,7 @@ Out of scope:
 You can install ruODK from GitHub with:
 
 ```{r gh-installation, eval = FALSE}
-# install.packages("devtools")
+if (!requireNamespace("remotes")) {install.packages("remotes")}
 remotes::install_github("dbca-wa/ruODK", dependencies = TRUE)
 ```
 
@@ -247,15 +246,15 @@ There are several other R packages interacting with the ODK ecosystem, and/or
 
 ### Comparison of ODK related software packages (non-ODK core)
 
-| Package                         | [`ruODK`](https://dbca-wa.github.io/ruODK/) | [`{odkr}`](https://validmeasures.org/odkr/) | [`{odk}`](https://cran.r-project.org/package=odk) | [`odkmeta`](https://github.com/nap2000/odkmeta) | [`{koboloadeR}`](https://unhcr.github.io/koboloadeR/docs/index.html) | [Pentaho Kettle tutorial](https://github.com/schemetrica/automating-data-delivery-odk-central)
+| Package                         | [`ruODK`](https://dbca-wa.github.io/ruODK/) | [`odkr`](https://validmeasures.org/odkr/) | [`odk`](https://cran.r-project.org/package=odk) | [`odkmeta`](https://github.com/nap2000/odkmeta) | [`koboloadeR`](https://unhcr.github.io/koboloadeR/docs/index.html) | [Pentaho Kettle tutorial](https://github.com/schemetrica/automating-data-delivery-odk-central)
 |------------------------------|---------------|---------------|---------------|---------------|---------------|---------------|
 | Elevator pitch               | "[ckanr](https://github.com/ropensci/ckanr) for ODK Central"  | "Drive ODK Briefcase through R" | "Export ODK Aggregate to SPSS" | "Export ODK Aggregate to STATA" | "Metapackage for extended ODK ecosystem" | "What ruODK does, but as GUI" |
-| Last commit                  | Sep 2019      |  Sep 2019     |    Nov 2017    |  Jun 2014    | Jul 2019 | Sept 2019 |
-| Website                      | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/) | [GH](https://github.com/dbca-wa/ruODK) [pkgdown](https://dbca-wa.github.io/ruODK/)   | [rdrr.io](https://rdrr.io/cran/odk/) |    [GH](https://github.com/nap2000/odkmeta) | [GH](https://github.com/unhcr/koboloadeR) [pkgdown](https://unhcr.github.io/koboloadeR/docs/index.html) | [GH](https://github.com/schemetrica/automating-data-delivery-odk-central) |
[…]
