+++ inst/rmarkdown/templates/odata/skeleton/skeleton.Rmd
@@ -29,36 +29,49 @@ Step 1: Setup ruODK, ckanr
 Secrets are set as environment variables to keep them outside of this document. 
 Read vignette "setup" for further information.
 -->
-<!-- Open the .Renviron file  -->
+<!-- 1a. Open the .Renviron file:  -->
 ```{r edit_renviron, echo=FALSE, eval=FALSE}
 usethis::edit_r_environ()
 ```
 
-<!-- and paste the following with the webuser's credentials  -->
+<!-- 1b. Paste the following block usign your own, working credentials:  -->
 ```{r paste_env_vars, echo=FALSE, eval=FALSE}
-ODKC_UN="..."
-ODKC_PW="..."
+# ODK Central web user with "manager" role on project
+ODKC_UN="my@email.com"
+ODKC_PW="my-odkc-password"
 
+# CKAN user with permissions to create a dataset
 CKAN_URL="https://demo.ckan.org"
 CKAN_KEY="my-ckan-api-key"
 ```
-<!-- and restart the R session to load these environment variables. -->
+<!-- 1c: Restart the R session to load these environment variables. -->
 
 <!-- 
-* Configure ODK Central base URL, project ID, and form ID through OData Service URL.
-  Paste the (non-sensitive) OData Service URL over `SVC_URL`.
+* Configure ODK Central base URL (url), project ID (pid), and form ID (fid) 
+  in one step using the OData Service URL.
+* Paste your own Service URL over `svc="..."`.
 * Paste your local timezone over `Australia/Perth`.
 * Adjust `loc` to a local subfolder for downloaded attachments.
 -->
 ```{r setup, message=FALSE, warning=FALSE}
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
-library(knitr)
-library(dplyr)
+# All packages used in this example
 library(ckanr)
+library(dplyr)
+library(DT)
+library(fs)
+library(googledrive)
+library(knitr)
+library(leaflet)
+library(readr)
+library(skimr)
 library(ruODK)
 
-ruODK::ru_setup(svc="SVC_URL")
+# Set ruODK defaults to an ODK Central form
+ruODK::ru_setup(
+  svc=paste0("https://sandbox.central.opendatakit.org/v1/projects",
+             "/14/forms/build_Flora-Quadrat-0-4_1564384341.svc"))
 tz <- "Australia/Perth"
 loc <- fs::path("attachments", "media")
 fs::dir_create(loc)
@@ -86,6 +99,7 @@ Repeat for each table name in form_tables$url.
 ```{r dl_submissions}
 data_main_raw <- ruODK::odata_submission_get(table = form_tables$url[1])
 data_sub1_raw <- ruODK::odata_submission_get(table = form_tables$url[2])
+data_sub2_raw <- ruODK::odata_submission_get(table = form_tables$url[3])
 # repeat for any remaining nested tables
 
 data <- data_main_raw %>% 
@@ -93,27 +107,40 @@ data <- data_main_raw %>%
   ruODK::parse_datetime(tz = tz) %>% 
   ruODK::parse_datetime(tz = tz, col_contains = "date") %>%
   dplyr::rename(
-    # # Inspect data, find colnames of coordinates (if any):
-    # longitude = x13,
-    # latitude = x14,
-    # altitude = x15
+    # Adjust coordinate colnames
+    longitude = x13,
+    latitude = x14,
+    altitude = x15
   ) %>%
   dplyr::mutate(
-    my_photo = ruODK::attachment_get(id, my_photo, local_dir = loc),
-    another_photo = ruODK::attachment_get(id, another_photo, local_dir = loc)
[…]
