+++ inst/rmarkdown/templates/odata/template.yaml
@@ -1,3 +1,3 @@
-name: odata
+name: ODK Central via OData
 description: >
- Accessing ODK Central through the OData API
+  Access data captured with OpenDataKit through the ODK Central OData API
