### 2019-09-19 #335 @maelle
Second reviewer recruited, thanks @jmt2080ad! :tada:  [Reviewer guide](https://devguide.ropensci.org/reviewerguide.html)

@florianm can you point both reviewers to instructions for testing to keep things smooth? Thanks!
