### 2019-09-19 #335 @florianm
@maelle thanks heaps!
@karissawhiting @jmt2080ad could you dm me your email addresses on https://discuss.ropensci.org/ ?
