### 2019-09-19 0a3fd36
Mogrify photos harder to make package smaller
