### 2019-09-19 e7d2766
Remove documentation cruft to reduce package size
