### 2019-09-19 d364e93
update CTB with new test settings
