### 2019-09-19 c890319
Speed up tests (10m -> 1m40): use smaller forms
Repeated download (testing overwrite/keep) uses a small form without attachments.
Attachment link uses small form with only two photos in one submission.
