### 2019-09-19 d135c41
Add cran-comments to explain R CMD CHECK note
