### 2019-09-20 3f9962f
#23 speed up tests: use separate forms
Refactor ru_setup, update tests, vignette "setup", ctb guidelines.
