+++ NAMESPACE
@@ -22,6 +22,8 @@ export(get_default_un)
 export(get_default_url)
 export(get_one_attachment)
 export(get_test_fid)
+export(get_test_fid_att)
+export(get_test_fid_zip)
 export(get_test_pid)
 export(get_test_pw)
 export(get_test_un)

[…]
