### 2019-09-20 e9adc60
Update CRAN comments with test build details
