### 2019-09-20 22cb9b4
Add funding authors
