+++ inst/rmarkdown/templates/odata/skeleton/skeleton.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Report Title"
 author: "Report Author"
-date: "`r Sys.Date()`"
+date: "`r Sys.date()`"
 output:
   html_document:
     toc = TRUE
