+++ README.md
@@ -182,8 +182,8 @@ ruODK::ru_setup(
 ```
 
 For all available detailed options to configure `ruODK`, read
-`vignette("Setup", package = "ruODK")` (online
-[here](https://dbca-wa.github.io/ruODK/articles/setup.html)).
+[`vignette("setup", package =
+"ruODK")`](https://dbca-wa.github.io/ruODK/articles/setup.html).
 
 ## Use ruODK
 
@@ -303,12 +303,12 @@ data %>% knitr::kable(.)
 | uuid:529cb189-8bb2-4cf1-9041-dcde716efb4f | 2019-09-18T14:08:43.257+08:00 | Florian Mayer | f73d2e1221ceaa06 | 2019-09-18T14:18:53.583+08:00 | 2019-09-18T06:20:25.780Z | 241           | flora           | uuid:529cb189-8bb2-4cf1-9041-dcde716efb4f | Kensington Planter Box 1 | 1568786958640.jpg | Point    | 115.8843 | \-31.99615 | \-17.37241 |        4.288 | flat                | 1568787004467.jpg          | Point    | 115.8844 | \-31.99620 |  \-8.662476 |        4.288 | Point    | 115.8844 | \-31.99622 |  \-6.266144 |        4.288 | Point    | 115.8844 | \-31.99621 |  \-6.597748 |        4.288 | 1568787172983.jpg | <https://sandbox.central.opendatakit.org/v1/projects/14/forms/build_Flora-Quadrat-0-4_1564384341.svc/$metadata#Submissions> |
 
 A more detailed walk-through with some data visualisation examples is
-available in the `vignette("odata", package="ruODK")` (online
-[here](https://dbca-wa.github.io/ruODK/articles/odata.html)).
+available in the [`vignette("odata",
+package="ruODK")`](https://dbca-wa.github.io/ruODK/articles/odata.html).
 
-See also `vignette("api", package="ruODK")` (online
-[here](https://dbca-wa.github.io/ruODK/articles/api.html)) for examples
-using the alternative RESTful
+See also [`vignette("api",
+package="ruODK")`](https://dbca-wa.github.io/ruODK/articles/api.html)
+for examples using the alternative RESTful
 API.
 
 ## Try ruODK
@@ -319,8 +319,8 @@ and the companion package [`urODK`](https://github.com/dbca-wa/urODK)
 opened as starting point for a hands-on workshop or instant demo of
 `ruODK` usage.
 
-Create a new RMarkdown workbook from `ruODK` template “odata” and follow
-the instructions within.
+Create a new RMarkdown workbook from `ruODK` template “ODK Central via
+OData” and follow the instructions within.
 
 ## Contribute
 
