+++ DESCRIPTION
@@ -8,7 +8,8 @@ Authors@R:
            role = c("aut", "cre"),
            email = "Florian.Mayer@dbca.wa.gov.au",
            comment = c(ORCID = "0000-0003-4269-4242")),
-           person("DBCA", role = c("cph", "fnd")))
+           person("DBCA", role = c("cph", "fnd")),
+           person("NWSFTCP", role = "fnd"))
 Description: Utilities to access and tidy up data from ODK Central's API.
     ODK Central is OpenDataKit's clearinghouse for digitally captured data 
     <https://docs.opendatakit.org/central-intro/>. 
