+++ _pkgdown.yml
@@ -16,5 +16,8 @@ reference:
   - has_concept("included")
 authors:
   DBCA:
-    html: "<img src='man/figures/BCS.png' width='200' alt='DBCA'/>"
+    html: "<img src='man/figures/dbca_bcs.png' width='150' alt='DBCA'/>"
     href: https://www.dbca.wa.gov.au/science/
+  NWSFTCP:
+    html: "<img src='man/figures/flatbacks.svg' width='150' alt='NWSFTCP'/>"
+    href: https://flatbacks.dbca.wa.gov.au/
