### 2019-09-23 dbca-wa/ruODK#29 @florianm
Use lifecycle badges on functions
## Feature
Use lifecycle badges on each fuction.

https://lifecycle.r-lib.org/

This will assist reviewers of https://github.com/ropensci/software-review/issues/335 to see where minor changes might come through and what's already on par with ODK Central 0.6.
