### 2019-09-23 dbca-wa/ruODK#29 @florianm
Warnings reported at https://github.com/r-lib/lifecycle/issues/22
