### 2019-09-23 c95202a
v0.6.5 add lifecycle badges fix #29
