+++ data-raw/make_release.R
@@ -1,6 +1,11 @@
+# Package mainainers can use the steps below to prepare a new release.
+
+# -----------------------------------------------------------------------------#
+# Prepare package
+#
 # Tests
 devtools::test()
-
+#
 # Docs
 styler::style_pkg()
 devtools::document(roclets = c("rd", "collate", "namespace"))
@@ -9,20 +14,20 @@ spelling::spell_check_files("README.Rmd", lang = "en_AU") # How to update word l
 spelling::update_wordlist()
 codemetar::write_codemeta("ruODK")
 usethis::edit_file("inst/CITATION")
-
-# Compile README.Rmd
 if (fs::file_info("README.md")$modification_time <
-    fs::file_info("README.Rmd")$modification_time){
-  rmarkdown::render("README.Rmd",  encoding = "UTF-8", clean = TRUE)
+  fs::file_info("README.Rmd")$modification_time) {
+  rmarkdown::render("README.Rmd", encoding = "UTF-8", clean = TRUE)
   if (fs::file_exists("README.html")) fs::file_delete("README.html")
 }
-
+#
 # Checks
 goodpractice::goodpractice(quiet = FALSE)
 devtools::check(cran = TRUE, remote = TRUE, incoming = TRUE)
 
-# Release
-usethis::use_version("minor")
+# -----------------------------------------------------------------------------#
+# Release package
+#
+usethis::use_version("minor") # or hand-edit DESC, CIT
 usethis::edit_file("NEWS.md")
 pkgdown::build_site()
 
