+++ R/data.R
@@ -1,6 +1,13 @@
 #' OData service document for an ODK Central form.
 #'
+#' \lifecycle{stable}
+#'
 #' The OData response for the metadata of an ODK Central form.
+#'
+#' This data is kept up to date with the data used in vignettes and package tests.
+#' The data is comprised of test records with nonsensical data.
+#' The forms used to capture this data are development versions of real-world
+#' forms.
 #'
 #' @format A tibble with one row per submission data endpoint.
 #' @source OData service document for
@@ -10,7 +17,14 @@
 
 #' OData metadata document for an ODK Central form.
 #'
+#' \lifecycle{stable}
+#'
 #' The OData response for the metadata of an ODK Central form.
+#'
+#' This data is kept up to date with the data used in vignettes and package tests.
+#' The data is comprised of test records with nonsensical data.
+#' The forms used to capture this data are development versions of real-world
+#' forms.
 #'
 #' @format A list of lists
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
@@ -18,6 +32,8 @@
 "fq_meta"
 
 #' OData submission data for an ODK Central form.
+#'
+#' \lifecycle{stable}
 #'
 #' The OData response for the submissions of an ODK Central form.
 #' This form represents a Flora Quadrat, which is a ca 50 by 50 m quadrat of
@@ -28,7 +44,10 @@
 #' and `system.file("extdata", "FloraQuadrat04.odkbuild", package = "ruODK")`,
 #' respectively.
 #'
-#' The data in this example are made up.
+#' This data is kept up to date with the data used in vignettes and package tests.
+#' The data is comprised of test records with nonsensical data.
+#' The forms used to capture this data are development versions of real-world
+#' forms.
 #'
 #' @format A list of lists
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
@@ -37,6 +56,8 @@
 
 #' OData submission data for a subgroup of an ODK Central form.
 #'
+#' \lifecycle{stable}
+#'
 #' The OData response for the subgroup of an ODK Central form.
 #'
 #' This subgroup represents vegetation strata as per the NVIS classification.
@@ -44,7 +65,10 @@
 #' by one or few plant taxa. Plant communities can be made of up to five strata,
 #' with two to three being most common.
 #'
-#' The data in this example is freely made up.
+#' This data is kept up to date with the data used in vignettes and package tests.
+#' The data is comprised of test records with nonsensical data.
+#' The forms used to capture this data are development versions of real-world
+#' forms.
 #'
 #' @format A list of lists
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
@@ -52,6 +76,8 @@
 "fq_raw_strata"
 
 #' OData submission data for a subgroup of an ODK Central form.
+#'
+#' \lifecycle{stable}
 #'
 #' The OData response for a subgroup of an ODK Central form.
 #'
@@ -62,7 +88,10 @@
 #' The voucher specimens are later determined by taxonomic experts, who then
 #' provide the real, terminal taxonomic name for a given voucher specimen.
 #'
-#' The data in this example is freely made up. The plants shown are real.
+#' This data is kept up to date with the data used in vignettes and package tests.
+#' The data is comprised of test records with nonsensical data.
+#' The forms used to capture this data are development versions of real-world
+#' forms.
 #'
 #' @format A list of lists
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
