### 2019-09-23 #335 @florianm
@karissawhiting @jmt2080ad @maelle
Update, have added lifecycle badges to all functions. The resulting NOTE is a bug upstream of the lifecycle package, see https://github.com/r-lib/lifecycle/issues/22. I've added an explanation to my `cran-comments.md`.

Karissa, Jason, feel free to contact me for test credentials on the ROpenSci slack when you're ready to commence the review. Your involvement is very appreciated!
