### 2019-09-23 fa30bd9
Broken CI: use GH remotes for lifecycle, usethis
