+++ DESCRIPTION
@@ -28,7 +28,6 @@ Imports:
     fs (>= 1.3.1),
     glue (>= 1.3.1),
     httr (>= 1.4.1),
-    lifecycle (>= 0.1.0.9000),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5),
     purrr (>= 0.3.2),
@@ -48,8 +47,10 @@ Suggests:
     listviewer (>= 2.1.0),
     rlist (>= 0.4.6.1),
     rmarkdown (>= 1.15),
-    testthat (>= 2.2.1),
-    usethis (>= 1.5.1.9000)
+    testthat (>= 2.2.1)
+Remotes:
+    r-lib/lifecycle
+    r-lib/usethis
 VignetteBuilder: 
     knitr
 RdMacros: 
