### 2019-09-23 76e18f1
Fix CI: typo in DESC Remotes
