+++ R/utils-lifecycle.R
@@ -1,0 +1,3 @@
+# Shim to fix <https://github.com/r-lib/lifecycle/issues/22>
+# not covered by tests as it is never run
+if (FALSE) lifecycle::deprecate_soft()
