+++ man/audit_get.Rd
@@ -103,8 +103,9 @@ names(logs)
 
 Other restful-api: \code{\link{attachment_link}},
   \code{\link{attachment_list}}, \code{\link{form_detail}},
-  \code{\link{form_list}}, \code{\link{form_schema}},
-  \code{\link{form_xml}}, \code{\link{project_create}},
+  \code{\link{form_list}}, \code{\link{form_schema_parse}},
+  \code{\link{form_schema}}, \code{\link{form_xml}},
+  \code{\link{project_create}},
   \code{\link{project_detail}}, \code{\link{project_list}},
   \code{\link{submission_detail}},
   \code{\link{submission_export}},
