+++ man/project_create.Rd
@@ -62,6 +62,7 @@ names(p)
 Other restful-api: \code{\link{attachment_link}},
   \code{\link{attachment_list}}, \code{\link{audit_get}},
   \code{\link{form_detail}}, \code{\link{form_list}},
+  \code{\link{form_schema_parse}},
   \code{\link{form_schema}}, \code{\link{form_xml}},
   \code{\link{project_detail}}, \code{\link{project_list}},
   \code{\link{submission_detail}},
