### 2019-09-24 4530e0e
Drop obsolete helpers, add pre-built PDF manual
