+++ R/utils.R
@@ -1,15 +1,3 @@
-#' #' Check whether we have internet
-#' #'
-#' #' @importFrom httr GET
-#' #' @importFrom rlang warn
-#' #' @noRd
-#' has_internet <- function() {
-#'   online <- !is.null(httr::GET("google.com"))
-#'   if (online == FALSE) rlang::warn("Please check internet connectivity!")
-#'   invisible(online)
-#' }
-
-
 #' Parse an ISO8601 datetime string to a timezone aware datetime.
 #'
 #' \lifecycle{stable}
