+++ R/ruODK.R
@@ -1,0 +1,11 @@
+#' @description ruODK is an R Client for the ODK Central API.
+#'
+#'   Please see the ruODK website for full documentation:
+#'   * <https://dbca-wa.github.io/ruODK/>
+#'
+#'   ruODK is "pipe-friendly" and, in fact, re-exports `%>%`, but does not
+#'   require its use.
+#'
+#' @importFrom rlang %||%
+#' @keywords internal
+"_PACKAGE"
