+++ R/odata_submission_get.R
@@ -1,4 +1,4 @@
-#' Retrieve /Submissions from an OData URL ending in .svc as list of lists
+#' Retrieve /Submissions from an OData URL ending in .svc as list of lists.
 #'
 #' \lifecycle{stable}
 #'
