### 2019-09-24 #335 @florianm
Update, just pushed a minor clean-up following advice from Jenny B on the ROpenSci slack. Removed some now obsolete helpers (map_chr_hack and friends), rebuilt docs.
