### 2019-09-24 485afac
Update remaining issues, link from docs.
