### 2019-09-24 41d8e60
Update template: fix typo
