### 2019-09-24 4a13f80
Include manual as PDF vignette
