### 2019-09-24 6960f77
update cran-comments.md
