+++ data-raw/make_release.R
@@ -34,6 +34,7 @@ usethis::edit_file("NEWS.md")
 fs::file_delete("inst/extdoc/ruODK.pdf")
 devtools::build_manual(path = "inst/extdoc") # rm > pdf
 fs::file_move(fs::dir_ls("inst/extdoc/"), "inst/extdoc/ruODK.pdf") # rename pdf
+tools::compactPDF("inst/extdoc/ruODK.pdf", gs_quality = "ebook") # compress pdf
 usethis::edit_file("vignettes/manual.Rnw") # hit "Compile PDF"
 #
 pkgdown::build_site()
