### 2019-09-25 302f150
Bugfix: OData Rmd template header options
