### 2019-09-25 16aac06
Drop pre-built PDF manual, keep workflow.
HT @jimhester at ROpenSci Slack
