### 2019-09-25 cac7e73
Use rlang %||% in odata_submission_get
