### 2019-09-25 #335 @florianm
@karissawhiting @jmt2080ad @maelle apologies for the impending code churn - I found a way to make ruODK way easier to drive: odata_submission_parse() now automatically parses the raw submission data into tibbles, parses dates, and downloads attachments. This will make the "OData" vignette a bit more concise and address https://github.com/dbca-wa/ruODK/issues/6
