### 2019-09-25 e7f92da
Fix #6: old dog (odata_submission_get), new tricks (parse)
