+++ R/odata_submission_get.R
@@ -6,27 +6,36 @@
 #' [Parse submissions and download attachments](https://github.com/dbca-wa/ruODK/issues/6)
 #'
 #' @param table The submission EntityType, or in plain words, the table name.
-#'            Default: "Submissions" (the main table).
-#'            Change to "Submissions.GROUP_NAME" for repeating form groups.
-#'            The group name can be found in the metadata under
-#'            DataServices.Schema.[EntityType].Name
+#'   Default: "Submissions" (the main table).
+#'   Change to "Submissions.GROUP_NAME" for repeating form groups.
+#'   The group name can be found through \code{`odata_service_get()`}.
 #' @param skip The number of rows to be omitted from the results.
-#'             Example: 10, default: NA (none skipped).
+#'   Example: 10, default: NA (none skipped).
 #' @param top The number of rows to return.
-#'            Example: 100, default: NA (all returned).
+#'   Example: 100, default: NA (all returned).
 #' @param count If TRUE, an `@odata.count` property will be returned in the
-#'              response from ODK Central. Default: FALSE.
+#'   response from ODK Central. Default: FALSE.
 #' @param wkt If TRUE, geospatial data will be returned as WKT (Well Known Text)
-#'            strings. Default: FALSE, returns GeoJSON structures.
+#'   strings. Default: FALSE, returns GeoJSON structures.
+#'   Note, ODK Central currently only honours this parameter for Point geometries.
+#'   Line and Polygon geometries are returned as "ODK WKT".
+#' @param parse Whether to parse submission data based on form schema.
+#'   Dates and datetimes will be parsed into local time.
+#'   Attachments will be downloaded, and the field updated to the local file path.
+#'   Default: TRUE.
+#' @template param-verbose
+#' @param tz A timezone, e.g. "Australia/Perth" or "UTC". Default: "UTC".
+#' @param local_dir The local folder to save the downloaded files to,
+#'   default: "media".
 #' @template param-pid
 #' @template param-fid
 #' @template param-url
 #' @template param-auth
-#' @return A nested list of lists.
-#'         `value` contains the submissions, which can be "rectangled" using
-#'     `tidyr::unnest_wider("element_name")`.
-#'     `@odata.context` is the URL of the metadata.
-#'     `@odata.count` is the total number of rows in the table.
+#' @return A list of lists.
+#'
+#'   * `value` contains the submissions as list of lists.
+#'   * `@odata.context` is the URL of the metadata.
+#'   * `@odata.count` is the total number of rows in the table.
 #' @seealso \url{https://odkcentral.docs.apiary.io/#reference/odata-endpoints/odata-form-service}
 #' @seealso \url{https://odkcentral.docs.apiary.io/#reference/odata-endpoints/odata-form-service/data-document}
 #' @family odata-api
@@ -68,13 +77,21 @@
 #'
 #' listviewer::jsonedit(data)
 #'
-#' # Next: parse this nested list in to a tidy tibble with `parse_submissions`
+#' # Next:
+#' # odata_submission_parse()
+#' # parse_datetime()
+#' # attachment_get()
+#' # or: odata_submission_get(parse = TRUE)
 #' }
 odata_submission_get <- function(table = "Submissions",
                                  skip = NULL,
                                  top = NULL,
                                  count = FALSE,
                                  wkt = FALSE,
+                                 parse = TRUE,
+                                 verbose = FALSE,
+                                 tz = "UTC",
[…]
