+++ tests/testthat/test-odata_submission_get.R
@@ -1,15 +1,40 @@
 context("test-odata_submission_get.R")
 
 test_that("odata_submission_get works with one known dataset", {
+  t <- tempdir()
   fresh_raw <- odata_submission_get(
     pid = get_test_pid(),
     fid = get_test_fid(),
     url = get_test_url(),
     un = get_test_un(),
-    pw = get_test_pw()
+    pw = get_test_pw(),
+    parse = FALSE
+  )
+  fresh_raw_parsed <- odata_submission_get(
+    pid = get_test_pid(),
+    fid = get_test_fid(),
+    url = get_test_url(),
+    un = get_test_un(),
+    pw = get_test_pw(),
+    parse = TRUE,
+    verbose = TRUE,
+    local_dir = t
   )
   fresh_parsed <- fresh_raw %>% odata_submission_parse()
   testthat::expect_gte(nrow(fresh_parsed), length(fresh_raw$value))
+  testthat::expect_gte(nrow(fresh_parsed), nrow(fresh_raw_parsed))
+
+  testthat::expect_equal(
+    class(fresh_raw_parsed$encounter_start_datetime),
+    c("POSIXct", "POSIXt")
+  )
+
+  local_files <- fresh_raw_parsed %>%
+    dplyr::filter(!is.null(quadrat_photo)) %>%
+    magrittr::extract2("quadrat_photo") %>%
+    as.character()
+  purrr::map(local_files, ~ testthat::expect_true(fs::file_exists(.)))
+
 })
 
 
@@ -28,6 +53,7 @@ test_that("odata_submission_get skip omits number of results", {
     url = get_test_url(),
     un = get_test_un(),
     pw = get_test_pw(),
+    parse = FALSE,
     table = fq_svc$name[2] # brittle: depends on form used
   )
   fresh_parsed <- fresh_raw %>% odata_submission_parse()
@@ -39,6 +65,7 @@ test_that("odata_submission_get skip omits number of results", {
     url = get_test_url(),
     un = get_test_un(),
     pw = get_test_pw(),
+    parse = FALSE,
     table = fq_svc$name[2] # brittle: depends on form used
   )
   skip_parsed <- skip_raw %>% odata_submission_parse()
@@ -53,7 +80,8 @@ test_that("odata_submission_get top limits number of results", {
     fid = get_test_fid(),
     url = get_test_url(),
     un = get_test_un(),
-    pw = get_test_pw()
+    pw = get_test_pw(),
+    parse = FALSE
   )
   top_parsed <- top_raw %>% odata_submission_parse()
 
@@ -70,7 +98,8 @@ test_that("odata_submission_get count returns total number or rows", {
     fid = get_test_fid(),
     url = get_test_url(),
     un = get_test_un(),
-    pw = get_test_pw()
+    pw = get_test_pw(),
+    parse = FALSE
   )
   x_parsed <- x_raw %>% odata_submission_parse()
 

[…]
