### 2019-09-25 4347370
v0.6.6: update examples in README, Rmd template
