+++ README.md
@@ -197,14 +197,14 @@ proj <- ruODK::project_list()
 proj %>% head() %>% knitr::kable(.)
 ```
 
-| id | name          | forms | app\_users | last\_submission    | created\_at         | updated\_at         | archived |
-| -: | :------------ | ----: | ---------: | :------------------ | :------------------ | :------------------ | :------- |
-|  8 | AmenazasRD    |     0 |          0 | NA                  | 2019-04-14 06:21:00 | NA                  | FALSE    |
-|  9 | BAFCO Test    |     1 |          1 | 2019-04-19 01:26:49 | 2019-04-19 00:30:45 | NA                  | FALSE    |
-| 11 | CaVaTeCo      |     2 |          1 | 2019-04-30 08:07:55 | 2019-04-30 06:43:22 | NA                  | FALSE    |
-| 18 | CherryPatch   |     2 |          1 | 2019-05-28 17:32:15 | 2019-05-27 21:09:33 | 2019-05-29 20:08:10 | FALSE    |
-| 32 | Collect3289v2 |     1 |          2 | 2019-08-07 13:44:19 | 2019-08-07 12:52:56 | NA                  | FALSE    |
-|  4 | Curso         |     7 |          4 | 2019-03-21 22:22:57 | 2019-02-26 03:55:50 | NA                  | FALSE    |
+| id | name          | forms | app\_users | created\_at         | updated\_at | last\_submission    | archived |
+| -: | :------------ | ----: | ---------: | :------------------ | :---------- | :------------------ | :------- |
+|  8 | AmenazasRD    |     0 |          0 | 2019-04-14 06:21:00 | NA          | NA                  | FALSE    |
+|  9 | BAFCO Test    |     1 |          1 | 2019-04-19 00:30:45 | NA          | 2019-04-19 01:26:49 | FALSE    |
+| 11 | CaVaTeCo      |     2 |          1 | 2019-04-30 06:43:22 | NA          | 2019-04-30 08:07:55 | FALSE    |
+| 18 | CherryPatch   |     2 |          1 | 2019-05-27 21:09:33 | NA          | 2019-05-28 17:32:15 | FALSE    |
+| 32 | Collect3289v2 |     1 |          2 | 2019-08-07 12:52:56 | NA          | 2019-08-07 13:44:19 | FALSE    |
+|  4 | Curso         |     7 |          4 | 2019-02-26 03:55:50 | NA          | 2019-03-21 22:22:57 | FALSE    |
 
 ``` r
 
@@ -238,69 +238,34 @@ frmd %>% knitr::kable(.)
 ``` r
 
 # Form schema
-meta <- ruODK::odata_metadata_get()
-# listviewer::jsonedit(meta)
-meta$Edmx$DataServices$Schema
-#> $ComplexType
-#> $ComplexType$Property
-#> list()
-#> attr(,"Name")
-#> [1] "submissionDate"
-#> attr(,"Type")
-#> [1] "Edm.DateTimeOffset"
-#> 
-#> $ComplexType$Property
-#> list()
-#> attr(,"Name")
-#> [1] "submitterId"
-#> attr(,"Type")
-#> [1] "Edm.String"
-#> 
-#> $ComplexType$Property
-#> list()
-#> attr(,"Name")
-#> [1] "submitterName"
-#> attr(,"Type")
-#> [1] "Edm.String"
-#> 
-#> $ComplexType$Property
-#> list()
-#> attr(,"Name")
-#> [1] "status"
-#> attr(,"Type")
-#> [1] "org.opendatakit.submission.Status"
-#> 
-#> attr(,"Name")
[…]
