### 2019-09-25 e414e33
Bugfix: odata_submission_get
Pipe through auth params for attachment_get, define local variable "type"
