+++ R/odata_submission_get.R
@@ -98,6 +98,7 @@ odata_submission_get <- function(table = "Submissions",
                                  un = get_default_un(),
                                  pw = get_default_pw()) {
   . <- NULL
+  type <- NULL
   yell_if_missing(url, un, pw)
 
   # Parse params
@@ -177,7 +178,17 @@ odata_submission_get <- function(table = "Submissions",
 
   sub <- sub %>% dplyr::mutate_at(
     dplyr::vars(att_cols),
-    ~ ruODK::attachment_get(id, ., local_dir = local_dir, verbose = verbose)
+    ~ ruODK::attachment_get(
+      id,
+      .,
+      local_dir = local_dir,
+      verbose = verbose,
+      pid = pid,
+      fid = fid,
+      url = url,
+      un = un,
+      pw = pw
+    )
   )
 
   #
