### 2019-09-25 fb699c9
Bugfix: update tests, clean up code, update notes
Update test-attachment_get to new behaviour of odata_submission_get().
Update cran comments.
Shorten a long code line in ru_setup.
