+++ cran-comments.md
@@ -21,7 +21,7 @@
 0 errors | 0 warnings | 1 note
 
 * This is a new release.
-* "installed size is  5.2Mb, sub-directories of 1Mb or more: doc 4.4Mb":
+* "installed size is  5.2Mb, sub-directories of 1Mb or more: doc 4.5Mb":
   * The vignette "odata" contains a leaflet map with popups showing embedded
     photos, a highly sought after use case.
     The third party dependencies, especially leaflet-awesomemarkers (for map 
