+++ R/attachment_get.R
@@ -28,7 +28,7 @@ prepend_uuid <- function(md5hash) {
 #'
 #' \lifecycle{stable}
 #'
-#' This is a helper function used by `attachment_get`.
+#' This is a helper function used by \code{`attachment_get()`}.
 #' This function is vectorised and accepts single values or vectors for uuid and
 #' fn.
 #'
@@ -54,14 +54,15 @@ attachment_url <- function(uuid,
 #'
 #' \lifecycle{stable}
 #'
-#' This is a helper function used by `get_attachment`.
-#' This function is not vectorised, but mapped by `get_attachment` to a tibble
-#' of input parameters.
+#' This is a helper function used by \code{`get_attachment()`}.
+#' This function is not vectorised, but mapped by \code{`get_attachment()`}
+#' to a tibble of input parameters.
 #'
 #' @param pth A local file path to save the attachment to.
 #' @param fn The attachment filename, as per ODK form submission. If NA, no file
-#'           will be downloaded, but NA will be returned.
-#' @param src The attachment's download URL, generated by `attachment_url`.
+#'   will be downloaded, but NA will be returned.
+#' @param src The attachment's download URL, generated by
+#'   \code{`attachment_url()`}.
 #' @template param-url
 #' @template param-auth
 #' @param verbose Whether to display debug messages or not (default).
@@ -99,20 +100,18 @@ get_one_attachment <- function(pth,
 #'
 #' \lifecycle{stable}
 #'
-#' This function is vectorised and can handle either one or many records.
+#' @details This function is vectorised and can handle either one or many records.
 #' Parameters submission_uuid and attachment_filename accept single or exactly
 #' the same number of multiple values.
 #' The other parameters are automatically repeated.
 #'
-#' The media attachments are downloaded into a folder named by the submission's
-#' uuid (minus "uuid:" for Windows compatibility) inside the given
-#' `local_dir`. E.g.:
+#' The media attachments are downloaded into a folder given by `local_dir`:
 #'
-#' workdir/attachments/xxxxxxxxxx/filename1.jpg
+#' workdir/media/filename1.jpg
 #'
-#' workdir/attachments/xxxxxxxxxx/filename2.jpg
+#' workdir/media/filename2.jpg
 #'
-#' workdir/attachments/xxxxxxxxxx/filename3.jpg
+#' workdir/media/filename3.jpg
 #'
 #' @param sid One or many ODK submission UUIDs, an MD5 hash.
 #' @param fn One or many ODK form attachment filenames, e.g. "1558330537199.jpg".
