### 2019-09-26 b6a60c0
Move rlist to Depends, used in form_schema_parse
