+++ DESCRIPTION
@@ -34,6 +34,7 @@ Imports:
     purrr (>= 0.3.2),
     readr (>= 1.3.1),
     rlang (>= 0.4.0),
+    rlist (>= 0.4.6.1),
     stringr (>= 1.4.0),
     tibble (>= 2.1.3),
     tidyr (>= 1.0.0),
@@ -46,7 +47,6 @@ Suggests:
     leaflet (>= 2.0.2),
     lifecycle (>= 0.1.0),
     listviewer (>= 2.1.0),
-    rlist (>= 0.4.6.1),
     rmarkdown (>= 1.15),
     testthat (>= 2.2.1),
     usethis (>= 1.5.1)
