### 2019-09-26 c69a984
update make_release.R
