+++ NEWS.md
@@ -1,3 +1,5 @@
+# ruODK (development version)
+
 # ruODK 0.6.6
 * The big one has landed: `odata_submission_get()` now defaults to parse 
   submissions from nested lists into a tibble, parse dates and datetimes,
