### 2019-09-26 c156330
Add lifecycle back to Imports, shim and test
