+++ tests/testthat/test-utils-lifecycle.R
@@ -1,0 +1,3 @@
+test_that("lifecycle_shim does nothing as expected", {
+  testthat::expect_null(lifecycle_shim())
+})
