+++ man/lifecycle_shim.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils-lifecycle.R
+\name{lifecycle_shim}
+\alias{lifecycle_shim}
+\title{Shim to allow Import of lifecycle, required for building docs.}
+\usage{
+lifecycle_shim()
+}
+\description{
+HT Jim Hester, Lionel Henry, Jenny Bryan for advice
+}
+\keyword{internal}
