+++ NAMESPACE
@@ -58,6 +58,7 @@ importFrom(httr,GET)
 importFrom(httr,add_headers)
 importFrom(httr,authenticate)
 importFrom(httr,content)
+importFrom(lifecycle,deprecate_soft)
 importFrom(magrittr,"%>%")
 importFrom(rlang,"%||%")
 importFrom(rlang,":=")
