+++ DESCRIPTION
@@ -29,6 +29,7 @@ Imports:
     glue (>= 1.3.1),
     httr (>= 1.4.1),
     janitor (>= 1.2.0),
+    lifecycle (>= 0.1.0),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5),
     purrr (>= 0.3.2),
@@ -45,7 +46,6 @@ Suggests:
     ggplot2 (>= 3.2.1),
     knitr (>= 1.24),
     leaflet (>= 2.0.2),
-    lifecycle (>= 0.1.0),
     listviewer (>= 2.1.0),
     rmarkdown (>= 1.15),
     testthat (>= 2.2.1),
