+++ codecov.yml
@@ -10,3 +10,6 @@ coverage:
       default:
         target: auto
         threshold: 1%
+ignore:
+  - "R/utils-lifecycle.R"
+
