+++ README.md
@@ -79,8 +79,9 @@ Benefits of using the R ecosystem in combination with ODK:
     [Sweave](https://support.rstudio.com/hc/en-us/articles/200552056-Using-Sweave-and-knitr),
     [RMarkdown](https://rmarkdown.rstudio.com/)), interactive web apps
     ([Shiny](https://shiny.rstudio.com/)), workflow scaling
-    ([drake](https://docs.ropensci.org/drake/)) and a range of
-    integrations with Docker (links coming soon).
+    ([drake](https://docs.ropensci.org/drake/))
+  - Rstudio-as-a-Service (RaaS) at
+    [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 
 `ruODK`’s scope:
 
@@ -112,12 +113,10 @@ Benefits of using the R ecosystem in combination with ODK:
 
 Out of scope:
 
-  - To wrap “management” API endpoints. The ODK Central GUI provides a
-    comprehensive interface for the management of users, roles,
-    permissions, projects, and forms. Behind the scenes, it is a [VueJS
-    application](https://github.com/opendatakit/central-frontend/)
-    working on the “management” API endpoints of the ODK Central
-    backend.
+  - To wrap “management” API endpoints. ODK Central is a [VueJS/NodeJS
+    application](https://github.com/opendatakit/central-frontend/) which
+    provides a comprehensive graphical user interface for the management
+    of users, roles, permissions, projects, and forms.
   - To provide extensive data visualisation. We show only minimal
     examples of data visualisation and presentation, mainly to
     illustrate the example data. Once the data is in your hands as tidy
@@ -125,7 +124,7 @@ Out of scope:
 
 ## Install
 
-You can install ruODK from GitHub with:
+You can install `ruODK` from GitHub with:
 
 ``` r
 if (!requireNamespace("remotes")) {install.packages("remotes")}
@@ -154,19 +153,17 @@ descriptions of the steps below.
     account](https://docs.opendatakit.org/central-users/#creating-a-web-user)
     on an ODK Central instance. Your username will be an email address.
   - [Create a project](https://docs.opendatakit.org/central-projects/)
-    and give the web user the relevant permissions.
-  - Create an XForm, e.g. using ODK Build, or use the provided example
-    forms.
-  - [Publish the form](https://docs.opendatakit.org/central-forms/) to
+    and give the web user at least [read
+    permissions](https://docs.opendatakit.org/central-projects/#managing-project-managers).
+  - Create an XForm, e.g. using ODK Build, or use the [example
+    forms](https://github.com/dbca-wa/ruODK/tree/master/inst/extdata)
+    provided by `ruODK`. The `.odkbuild` versions can be loaded into
+    [ODK Build](https://build.opendatakit.org/), while the `.xml`
+    versions can be directly imported into ODK Central.
+  - [Publish the form](https://docs.opendatakit.org/central-forms/)to
     ODK Central.
   - Collect some data for this form on ODK Collect and let ODK Collect
     submit the finalised forms to ODK Central.
-
[…]
