+++ README.Rmd
@@ -71,8 +71,9 @@ Benefits of using the R ecosystem in combination with ODK:
    [Sweave](https://support.rstudio.com/hc/en-us/articles/200552056-Using-Sweave-and-knitr), 
    [RMarkdown](https://rmarkdown.rstudio.com/)), interactive web apps 
    ([Shiny](https://shiny.rstudio.com/)), 
-   workflow scaling ([drake](https://docs.ropensci.org/drake/)) and a range of
-   integrations with Docker (links coming soon).
+   workflow scaling ([drake](https://docs.ropensci.org/drake/))
+* Rstudio-as-a-Service (RaaS) at 
+  [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 
 `ruODK`'s scope:
 
@@ -102,18 +103,17 @@ Benefits of using the R ecosystem in combination with ODK:
   
 Out of scope:
 
-*  To wrap "management" API endpoints. The ODK Central GUI provides a comprehensive 
-   interface for the management of users, roles, permissions, projects, and forms.
-   Behind the scenes, it is a 
-   [VueJS application](https://github.com/opendatakit/central-frontend/) 
-   working on the "management" API endpoints of the ODK Central backend.
+*  To wrap "management" API endpoints. ODK Central is a
+   [VueJS/NodeJS application](https://github.com/opendatakit/central-frontend/) 
+   which provides a comprehensive graphical user interface for the management of 
+   users, roles, permissions, projects, and forms.
 *  To provide extensive data visualisation. We show only minimal examples of data 
    visualisation and presentation, mainly to illustrate the example data.
    Once the data is in your hands as tidy tibbles... urODK!
 
 ## Install
 
-You can install ruODK from GitHub with:
+You can install `ruODK` from GitHub with:
 
 ```{r gh-installation, eval = FALSE}
 if (!requireNamespace("remotes")) {install.packages("remotes")}
@@ -135,26 +135,35 @@ provides up-to-date descriptions of the steps below.
 
 *  [Create a web user account](https://docs.opendatakit.org/central-users/#creating-a-web-user) 
    on an ODK Central instance. Your username will be an email address.
-*  [Create a project](https://docs.opendatakit.org/central-projects/) and give the
-   web user the relevant permissions.
-*  Create an XForm, e.g. using ODK Build, or use the provided example forms.
-*  [Publish the form](https://docs.opendatakit.org/central-forms/) to ODK Central.
+*  [Create a project](https://docs.opendatakit.org/central-projects/) and give 
+   the web user at least
+   [read permissions](https://docs.opendatakit.org/central-projects/#managing-project-managers).
+*  Create an XForm, e.g. using ODK Build, or use the 
+   [example forms](https://github.com/dbca-wa/ruODK/tree/master/inst/extdata) 
+   provided by `ruODK`. The `.odkbuild` versions can be loaded into 
+   [ODK Build](https://build.opendatakit.org/), while the `.xml` versions can be 
+   directly imported into ODK Central.
+*  [Publish the form](https://docs.opendatakit.org/central-forms/)to ODK Central.
 *  Collect some data for this form on ODK Collect and let ODK Collect submit the
    finalised forms to ODK Central.
-
-A note on the [included example forms](https://github.com/dbca-wa/ruODK/tree/master/inst/extdata): 
-The `.odkbuild` versions can be loaded into [ODK Build](https://build.opendatakit.org/), 
-while the `.xml` versions can be imported into ODK Central.
 
 ## Configure `ruODK`
[…]
