### 2019-10-12 8c18931
remove currently broken skimr from Rmd template
