### 2019-10-12 28e65fc
Bring skimr back to Rmd template. See https://github.com/tidyverse/purrr/issues/711
