+++ R/utils.R
@@ -46,7 +46,8 @@ isodt_to_local <- function(datetime_string,
 #' }
 parse_datetime <- function(df,
                            col_contains = "time",
-                           orders = c("YmdHMS", "YmdHMSz"),
+                           orders = c("YmdHMS", "YmdHMSz",
+                                      "Ymd HMS", "Ymd HMSz"),
                            tz = "Australia/Perth") {
   df %>%
     dplyr::mutate_at(
