### 2019-10-12 f46faf1
Update docs, format code, add date formats to parst_datetime
