+++ tests/testthat/test-utils-lifecycle.R
@@ -1,3 +1,3 @@
 test_that("lifecycle_shim does nothing as expected", {
-  testthat::expect_warning(lifecycle_shim())
+  testthat::expect_warning(ruODK:::lifecycle_shim())
 })
