+++ R/ru_setup.R
@@ -103,15 +103,16 @@ odata_svc_parse <- function(svc) {
 #'   A numeric value for \code{pid} will be converted to character.
 #' @param fid (optional, character) The alphanumeric ID of an existing form
 #'   in \code{pid}. This will override the form ID from \code{svc}.
-#' @param url An ODK Central URL, e.g. "https://sandbox.central.opendatakit.org".
+#' @param url An ODK Central URL,
+#'   e.g. "https://sandbox.central.opendatakit.org".
 #'   This will override the ODK Central base URL from \code{svc}.
 #' @param un An ODK Central username which is the email of a "web user" in the
 #'   specified ODK Central instance \code{url} (optional, character).
 #' @param pw The password for user \code{un} (optional, character).
 #' @param test_svc (optional, character) The OData service URL of a test form.
-#'   This parameter will set \code{test_pid}, \code{test_fid}, and \code{test_url}.
-#'   It is sufficient to supply \code{test_svc}, \code{test_un}, and
-#'   \code{test_pw} to configure testing.
+#'   This parameter will set \code{test_pid}, \code{test_fid}, and
+#'   \code{test_url}. It is sufficient to supply \code{test_svc},
+#'   \code{test_un}, and \code{test_pw} to configure testing.
 #' @param test_pid (optional, character) The numeric ID of an existing project
 #'   on \code{test_url}. This will override the project ID from \code{test_svc}.
 #'   A numeric value for \code{test_pid} will be converted to character.
@@ -119,12 +120,14 @@ odata_svc_parse <- function(svc) {
 #'   in \code{test_pid}. This will override the form ID from \code{test_svc}.
 #'   This form is used as default form in all tests, examples, vignettes, data,
 #'   and Rmd templates.
-#' @param test_fid_zip (optional, character) The alphanumeric ID of an existing form
-#'   in \code{test_pid}. This will override the form ID from \code{test_svc}.
+#' @param test_fid_zip (optional, character) The alphanumeric ID of an existing
+#'   form in \code{test_pid}. This will override the form ID from
+#'   \code{test_svc}.
 #'   Provide the form ID of a form with few submissions and without attachments.
 #'   This form is used to test the repeated download of all form submissions.
-#' @param test_fid_att (optional, character) The alphanumeric ID of an existing form
-#'   in \code{test_pid}. This will override the form ID from \code{test_svc}.
+#' @param test_fid_att (optional, character) The alphanumeric ID of an existing
+#'   form in \code{test_pid}. This will override the form ID from
+#'   \code{test_svc}.
 #'   Provide the form ID of a form with few submissions and few attachments.
 #'   This form is used to test downloading and linking attachments.
 #' @param test_url (optional, character) A valid ODK Central URL for testing.
@@ -143,19 +146,19 @@ odata_svc_parse <- function(svc) {
 #' to set up username and password initially with
 #' \code{ru_setup(un="XXX", pw="XXX")}, and switch between forms with
 #' \code{ru_setup(svc="XXX")}, supplying the form's OData service URL.
-#' ODK Central conveniently provides the OData service URL in the form submission
-#' tab, which in turn contains base URL, project ID, and form ID.
+#' ODK Central conveniently provides the OData service URL in the form
+#' submission tab, which in turn contains base URL, project ID, and form ID.
 #'
 #' \code{ruODK}'s automated tests require a valid ODK Central URL, and a
 #'   privileged username and password of a "web user" on that ODK Central
 #'   instance, as well as an existing project and form.
 #'
 #' @examples
-#' # \code{ruODK} users only need default settings to their ODK Central instance:
[…]
