+++ man/fq_meta.Rd
@@ -17,8 +17,8 @@ fq_meta
 \details{
 The OData response for the metadata of an ODK Central form.
 
-This data is kept up to date with the data used in vignettes and package tests.
-The data is comprised of test records with nonsensical data.
+This data is kept up to date with the data used in vignettes and package
+tests. The data is comprised of test records with nonsensical data.
 The forms used to capture this data are development versions of real-world
 forms.
 }

[…]
