### 2019-10-12 6fd857d
Use dev versions of rlang, purrr, lifecycle
