+++ DESCRIPTION
@@ -29,10 +29,13 @@ Imports:
     glue (>= 1.3.1),
     httr (>= 1.4.1),
     janitor (>= 1.2.0),
+    lifecycle (>= 0.1.0.9000),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5),
     readr (>= 1.3.1),
+    rlang (>= 0.4.0.9000),
     rlist (>= 0.4.6.1),
+    purrr (>= 0.3.2.9000),
     stringr (>= 1.4.0),
     tibble (>= 2.1.3),
     tidyr (>= 1.0.0),
@@ -47,10 +50,6 @@ Suggests:
     rmarkdown (>= 1.16),
     testthat (>= 2.2.1),
     usethis (>= 1.5.1.9000)
-Remotes:
-    tidyverse/purrr,
-    r-lib/rlang,
-    r-lib/lifecycle
 VignetteBuilder: 
     knitr
 RdMacros: 
