### 2019-10-13 #335 @karissawhiting
- [x] **Vignette(s)** demonstrating major functionality that runs successfully locally – _I'm not able to build vignettes without first setting up ODK credentials_
- [ ] **Examples** for all exported functions in R Help that run successfully locally – _not all exported functions have examples, but some internal helper functions are exported that could potentially be removed from exports (e.g. `prepend_uuid()`)_

- [x] **Installation:** Installation succeeds as documented. – _`remotes::install_github()` works by default, but fails when `build_vignettes = TRUE` if you do not have your credentials stored prior to build._

---

`ruODK` provides a much-needed link between the ODK data collection/management software suite and R. The package is modular and well-structured, and the code is tidy, styled and readable. Documentation and vignettes are informative and make the project accessible through concise descriptions and links to resources as needed.

For context, this review is coming from the perspective of someone who has experience with the ODK software suite, but little experience with R API wrapper packages.

#### Build
-  I wasn't able to build without first specifying ODK credentials (because of `api.Rmd` and `odata.Rmd` vignettes). I’m not sure if there are established best practices with API packages and this issue, but @noamross pointed out this [rcites package vignette](https://github.com/ropensci/rcites/blob/master/vignettes/a_get_started.Rmd) as an example of one way this is addressed in API packages. Vignettes decide whether or not to eval code chunks based on an environment variable (in this case ` NOT_CRAN`) and both ways are tested in the [Travis-CI testing grid](https://github.com/ropensci/rcites/blob/master/.travis.yml).

- The diagram is extremely helpful!
- You could add a bit more description about ODK Central Sandbox in this doc to help with multiple points of entry. New prospective users may be looking for the quickest way to test and play around with ODK/`ruODK` to see if it fits the needs of their project, and it seems this would be via a Sandbox account. To this point, you could adjust the "ODK Central Setup" section to say something like “Create a web user account on an ODK Central instance, or request an account on Sandbox..."
- Put “(ruODK)” in backticks for consistency (in CONTRIBUTING doc as well)
- Typo in link “Publish the formto ODK Central”.

#### Tests and CONTRIBUTING
- All tests passed, and overall good test coverage
- There is a barrier to entry in contributing to the package and running tests because of the need for testing credentials. That being said, the CONTRIBUTING document and the new github form you set up to request credentials definitely does help with this. Is there a way to set up a public set of ODK credentials using a dummy account? If not, consider me a +1 vote for setting up cached responses in the future when the ODK API is more stable if it will provide a way for tests to work out of the box :)
- Title of [CONTRIBUTING](https://dbca-wa.github.io/ruODK/CONTRIBUTING.html#test) is NA.
- You still have some leftover language from the `usethis` contributing doc. For example: “Before you do a pull request, you should always file an issue and make sure someone from the tidyverse team…”
- "To run tests, you’ll need access to the ODK Central sandbox instance…" I think this should go right at the top or even in the “prerequisites” section of technical contributions.

#### NAMESPACE
- Not all exported functions have examples, but some internal helper functions are exported that may not need to be (e.g. `prepend_uuid()`, `yell_if_missing()`).
- Do you need [utils-tidy-eval.R](https://github.com/dbca-wa/ruODK/blob/a5dceb349f036caa77a712589cfbebd5e6921318/R/utils-tidy-eval.R), and these rlang functions exported?

-	I’m having some trouble distinguishing between the OData API and RESTful API functionalities. Do these workflows mirror each other? Does one workflow have functionality available that the other does not? I think some additional delineation between these could be helpful. Here are a few ideas on how to address this (suggestions only):
    - If these are parallel workflows, consider having the vignette structures mirror each other, and add the same blurb at the top of each (like “three ways to happiness") with links to the other vignette. If there are any large benefits or drawbacks of one method or the other, call them out there or in the README.
    - Does your preferred workflow consist of a mixture of both OData and RESTful calls? If so, consider creating just a typical workflow vignette that goes through a case study using both kinds of functions.
- Consider renaming vignette file names to `odata-api`, and `restful-api`

#### Notes on Code
-	I see several functions have `. <- NULL`. I’m assuming this is to avoid “no visible bindings for global variables”. One solutions is to add `utils::globalVariables(".")` in `ruODK.R` instead of specifying this in each function.
-	Do you need `yell_if_missing()` in functions that have the get functions as their default args? In those cases, it looks like if you are missing credentials the `get_default_<things>()` functions already do the yelling, but I may be missing some edge case here.
- `ru_setup()`
   - I was expecting a confirmation message of some sort. Maybe it could print `ru_settings()` when you run it by default?
    -  Is there a way to actually check the authentication of your username/project/form credentials here or in another function?
    - `ggmaps::register_google()` has an optional argument to `write = TRUE` where it will write credentials to your `.Renviron`. Just wanted to point this out as a cool enhancement option.
- @examples sections are formatted as R code so you can take out \code{ } tags (e.g. `ru_setup()`)
- `odata_submission_get()`
    - It could be helpful to separate arguments for parsing the data and getting attachments in this function.
    - It looks like `odata_submission_get()` returns all submissions, but the equivalent function for the RESTful workflow `submission_get()` returns one submission at a time. I expected these functions to mirror each other in terms of outputs. Perhaps it would make sense to turn this code chunk you have in the vignette into a function? This would mirror the `attachment_get() `/ `get_one_attachment()` functions too.

``` # Multiple submissions
submissions <- tibble::tibble(
  pid = ruODK::get_test_pid(),
  fid = ruODK::get_test_fid(),
  iid = sl$instance_id,  # this is a vector of multiple instance_ids
  url = ruODK::get_test_url(),
  un = ruODK::get_test_un(),
  pw = ruODK::get_test_pw()
) %>%
  purrr::pmap(ruODK::submission_get)
```

Thank you for sharing this great package!!
