### 2019-10-14 #335 @florianm
Thanks @karissawhiting for the review! I'll create issues from the review, and once all are closed, will summarise them here.
