### 2019-10-14 dbca-wa/ruODK#32 @florianm
Add examples to all exported functions
## Feature
Examples for all exported functions in R Help that run successfully locally – not all exported functions have examples, but some internal helper functions are exported that could potentially be removed from exports (e.g. prepend_uuid()).

Not all exported functions have examples, but some internal helper functions are exported that may not need to be (e.g. prepend_uuid(), yell_if_missing()).

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
