### 2019-10-14 dbca-wa/ruODK#33 @florianm
Use canned data in vignettes
## Problem
Karissa found:

* Installation succeeds as documented.
* `remotes::install_github()` works by default, but fails when `build_vignettes = TRUE` if you do not have your credentials stored prior to build.

Jason found same breakage with auth credentials not available in `devtools::install(build_vignettes=TRUE)`.

## Solution
* Have a squizzle at https://github.com/ropensci/rcites/blob/master/vignettes/a_get_started.Rmd for best practices: Vignettes decide whether or not to eval code chunks based on an environment variable (in this case NOT_CRAN) and both ways are tested in the Travis-CI testing grid. https://github.com/ropensci/rcites/blob/master/.travis.yml
* Include canned data for all results from all web requests sent by vignettes.
* This will enable vignettes to build from canned data without the need for ODK Central sandbox credentials.
* This will require extra tests to verify that canned data is identical to results from web requests.
* Add a clear delineation as to what can be done with and without test credentials.

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
