### 2019-10-14 dbca-wa/ruODK#34 @florianm
Improve documentation
## Feature
* Acknowledge Karissa and Jason (pending his approval) as reviewers and Maelle as editor in DESC.
* Add a bit more description about ODK Central Sandbox in this doc to help with multiple points of entry. New prospective users may be looking for the quickest way to test and play around with ODK/ruODK to see if it fits the needs of their project, and it seems this would be via a Sandbox account. To this point, you could adjust the "ODK Central Setup" section to say something like “Create a web user account on an ODK Central instance, or request an account on Sandbox..."
* Put “(ruODK)” in backticks for consistency (in CONTRIBUTING doc as well)
* Typo in link “Publish the formto ODK Central”.
* Title of CONTRIBUTING is NA.
* Delete leftover language from the usethis contributing doc. For example: “Before you do a pull request, you should always file an issue and make sure someone from the tidyverse team…”
* "To run tests, you’ll need access to the ODK Central sandbox instance…" I think this should go right at the top or even in the “prerequisites” section of technical contributions.
* Review the need for utils-tidy-eval.R, and these rlang functions exported?
* Drop \code{} from `@examples`

> I’m having some trouble distinguishing between the OData API and RESTful API functionalities. Do these workflows mirror each other? Does one workflow have functionality available that the other does not? I think some additional delineation between these could be helpful. Here are a few ideas on how to address this (suggestions only):
> If these are parallel workflows, consider having the vignette structures mirror each other, and add the same blurb at the top of each (like “three ways to happiness") with links to the other vignette. If there are any large benefits or drawbacks of one method or the other, call them out there or in the README.
> Does your preferred workflow consist of a mixture of both OData and RESTful calls? If so, consider creating just a typical workflow vignette that goes through a case study using both kinds of functions.

* Clarify purpose of different approaches:
  * REST API one by one: e.g. for incremental ETL to data warehouse. Most modular way to access and parse data.
  * REST API zip dump: for complete export. Easiest way to get all data.
  * OData: most comfortable way to access and parse data.
* Rename vignette file names to odata-api, and restful-api

> Why provide a 'hard way' to pass credentials via a function call? My
sense is that this will encourage bad practices in writing code, as
this would give users a way of hard coding passwords in workflows,
which is typically discouraged.

* Clarified and warned in vignette "setup".

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
