### 2019-10-14 dbca-wa/ruODK#35 @florianm
Define `.` as global variable
## Feature
* add `utils::globalVariables(".")` to `ruODK.R` (check other global vars too)
* drop `. <- NULL` from functions
* enjoy the contented silence of devtools::check()
* owe Karissa a beer for the great tip

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
