### 2019-10-14 dbca-wa/ruODK#36 @florianm
Drop yell_if_missing in functions using get_default_* (which already yells if missing)
## Feature

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
