### 2019-10-14 dbca-wa/ruODK#37 @florianm
Make settings great again
## Feature
* Add ru_settings(pw=FALSE) to hide or show pw
* ru_setup
  * Print ru_settings(pw=FALSE) as last step of ru_setup
  * add parameter write=FALSE to write to .Renviron (steal from ggmaps::register_google())
* add ru_soundcheck to test:
  * default settings: set? working? warn if not
  * test settings: set? working? warn if not
  * return TRUE if both work

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
