### 2019-10-14 dbca-wa/ruODK#38 @florianm
Make odata_submission_get and submission_get behave the same
## Feature
*  separate arguments for parsing the data and getting attachments in this function.
 * odata_submission_get() returns all submissions, but the equivalent function for the RESTful workflow submission_get() returns one submission at a time. Turn this code chunk you have in the vignette into one function.

* Rename submission_get into get_one_submission
* Create submission_get from below code as seen in vignette:

```
submissions <- tibble::tibble(
  pid = ruODK::get_test_pid(),
  fid = ruODK::get_test_fid(),
  iid = sl$instance_id,  # this is a vector of multiple instance_ids
  url = ruODK::get_test_url(),
  un = ruODK::get_test_un(),
  pw = ruODK::get_test_pw()
) %>%
  purrr::pmap(ruODK::submission_get)
```

source https://github.com/ropensci/software-review/issues/335#issuecomment-541438418
