### 2019-10-15 06c64ce
Update README with odkmeta details.
Thanks to odkmeta author @matthew-white for the info!
