### 2019-10-15 7292f20
Fix #35 - define `.` as global variable.
Use utils::globalVariables(".") over `. <- NULL` to silence R CMD check.

`form_schema_parse` still contains three other variables (type, name, children) which are asigned to NULL as not to pollute global env.
