+++ R/odata_metadata_get.R
@@ -31,7 +31,6 @@ odata_metadata_get <- function(pid = get_default_pid(),
                                url = get_default_url(),
                                un = get_default_un(),
                                pw = get_default_pw()) {
-  . <- NULL
   yell_if_missing(url, un, pw)
   glue::glue("{url}/v1/projects/{pid}/forms/{fid}.svc/$metadata") %>%
     httr::GET(
