+++ R/project_detail.R
@@ -40,7 +40,6 @@ project_detail <- function(pid = get_default_pid(),
                            url = get_default_url(),
                            un = get_default_un(),
                            pw = get_default_pw()) {
-  . <- NULL
   yell_if_missing(url, un, pw, pid = pid)
   glue::glue("{url}/v1/projects/{pid}") %>%
     httr::GET(
