+++ R/form_list.R
@@ -35,7 +35,6 @@ form_list <- function(pid = get_default_pid(),
                       url = get_default_url(),
                       un = get_default_un(),
                       pw = get_default_pw()) {
-  . <- NULL
   yell_if_missing(url, un, pw, pid = pid)
   glue::glue("{url}/v1/projects/{pid}/forms") %>%
     httr::GET(
