+++ R/submission_get.R
@@ -46,7 +46,6 @@ submission_get <- function(iid,
                            url = get_default_url(),
                            un = get_default_un(),
                            pw = get_default_pw()) {
-  . <- NULL
   yell_if_missing(url, un, pw, pid = pid, fid = fid)
   glue::glue("{url}/v1/projects/{pid}/forms/{fid}/submissions/{iid}.xml") %>%
     httr::GET(httr::authenticate(un, pw)) %>%
