### 2019-10-15 706dab3
#37 prevent leaking passwords in print.ru_settings
