### 2019-10-15 c81f775
#34 Drop \code{} markup from @examples
