+++ R/form_schema_parse.R
@@ -34,10 +34,10 @@
 #'
 #' fsp
 #'
-#' # Attachments: use \code{`attachment_get`} on each of
+#' # Attachments: use `attachment_get` on each of
 #' fsp %>% dplyr::filter(type == "binary")
 #'
-#' # dateYime: use \code{`parse_datetime`} on each of
+#' # dateYime: use `parse_datetime` on each of
 #' fsp %>% dplyr::filter(type == "dateTime")
 #'
 #' # Point location: will be split into lat/lon/alt/acc
