+++ R/form_schema.R
@@ -105,7 +105,7 @@
 #' # Attachments: use \code{`attachment_get`} on each of
 #' fsp %>% dplyr::filter(type == "binary")
 #'
-#' # dateYime: use \code{`parse_datetime`} on each of
+#' # dateTime: use \code{`parse_datetime`} on each of
 #' fsp %>% dplyr::filter(type == "dateTime")
 #'
 #' # Point location: will be split into lat/lon/alt/acc
