+++ man/form_schema.Rd
@@ -139,7 +139,7 @@ fsp <- form_schema(parse = TRUE)
 # Attachments: use \\code{`attachment_get`} on each of
 fsp \%>\% dplyr::filter(type == "binary")
 
-# dateYime: use \\code{`parse_datetime`} on each of
+# dateTime: use \\code{`parse_datetime`} on each of
 fsp \%>\% dplyr::filter(type == "dateTime")
 
 # Point location: will be split into lat/lon/alt/acc
