### 2019-10-15 #335 @jmt2080ad
# Package Review

   + Great package comparison in README.

   + `devtools::install(build_vignettes = TRUE)` does not install for
     me. I get permission errors that I do not get when I run the
     vingette chunks on their own.

   + I cannot build the vignettes using devtools, but...
   + I can run all the Rmd chunks used to build the vignettes. They are verbose and clear. Great documentation.

 - [X] **Community guidelines** including contribution guidelines in
      the README or CONTRIBUTING, and DESCRIPTION with `URL`,
      `BugReports` and `Maintainer` (which may be autogenerated via
      `Authors@R`).

   + Due to my credential error, I was not able to successfully run the tests.
   + I am also getting the following error:

test-attachment_get.R:80: error: get_one_attachment handles repeat
download and NA filenames
unused argument (check = TRUE)
1: .handleSimpleError(function (e)
   {
       handled <<- TRUE
       test_error <<- e
       options(expressions = expressions_opt_new)
       on.exit(options(expressions = expressions_opt), add = TRUE)
       e$expectation_calls <- frame_calls(11, 2)
       test_error <<- e
       register_expectation(e)
       e$handled <- TRUE
       test_error <<- e
   }, "unused argument (check = TRUE)", quote(tempdir(check = TRUE)))
   at /home/....../test-attachment_get.R:80
```

   + The namespace conflict with `readr` for the `parse_datetime` is discouraged by the guidelines.

---

+ Building vignettes require a properly setup account. Even with a
  setup account, `devtools::install(build_vignettes = TRUE)` fails.

+ When attempting to build vignettes via devtools fails, the user
  password is printed to the console. `ru_settings()` also prints the
  password to the console. I would feel better if that was hidden from
  the user.

### Notes on setup functions and vignette

+ Why provide a 'hard way' to pass credentials via a function call? My
  sense is that this will encourage bad practices in writing code, as
  this would give users a way of hard coding passwords in workflows,
  which is typically discouraged.

### Notes on API functions and vignette

+ `parse_datetime()` appears to parse any column name with the text
  "time" in it by default or a user defined string. This could
  potentially try to parse unintended columns containing words like
  "candidate" or "centimeter". I think this functionality might be
  best left to the user to define for themselves.

tibble::tibble(
            centimeter = 1:10,
            millimeter = 1:10,
            candidate  = sample(c("Jack", "Jill"), 10, replace = T),
            rally      = sample(c("Home", "Away"), 10, replace = T)
        ) %>%
    ruODK::parse_datetime(.) %>%
    ruODK::parse_datetime(col_contains = "date")
```

+ Also, `parse_datetime()` is a `readr` function. During my review,
  those namespaces became confused in a couple of the examples and I
  was no longer able to run `ruODK` vignette code. Would a different
  name be better suited?

### Notes on accessing odata functions and vignette

+ This error message should quote email address.

```
   Not Found (HTTP 404). Failed to get desired response from server
   https://sandbox.central.opendatakit.org/ as user
   "myemail@email.com".
```

+ In the odata vignette you have an expression `fq_svc <-
  odata_service_get()`. Everywhere else you define the namespace for
  an `ruODK` function, i.e. `ruODK::odata_service_get()`. For
  constancy I would edit this to include the namespace here and
  elsewhere where that is missing.

+ Also in the odata vignette, in the section "OData service document",
  there appears to be a network call to ODK Central to assign the
  `fc_svc` variable. The next comment however indicates that the
  canned data is used. It doesn't appear that the datasets associated
  with this package that are loaded early in this vignette are used.

+ In the odata vignette, in the section "Summarising data", the text
  is a bit odd as it implies the previous leaflet examples
  performed data transformations. Consider revising the language.

+ The rmarkdown template suggestion is great and the template would
  make a great quick guide.
