### 2019-10-15 #335 @florianm
Thanks @jmt2080ad - as we discussed during the beer review process, all of that makes wild sense, some typos and wrinkles to iron out, nothing breaking except to use canned data in vignettes.

re vignettes not building - will rewrite vignettes to 1. use canned data but also 2. show but not evaluate the real functions, as discussed with Jason. Might need some help from Scott tomorrow on that. (Great to be in PDX)

re parse_datetime - will rename to avoid namespace conflict, and while the default is to parse anyt column containing "date" (common denominator of datetime and date, thought that was clever) actually is only used explicitly on date and dateTime columns as per `form_schema`, so the "candidate" fields will never be parsed automatically.

re explicit parameters for credentials - don't think this will promote bad code, it is meant to be used with Sys.getenv() or other variables. I will amend the vignette "setup" to show this.

First few fixes are done but will not be pushed until after the PDX and Seattle workshops. I hope to get around to the full list by the end of next week (work travel with spotty internet until Thu 24 Oct).
