### 2019-10-15 c690dfc
Hide sensitive settings
