+++ tests/testthat/test-utils-lifecycle.R
@@ -1,3 +1,4 @@
 test_that("lifecycle_shim does nothing as expected", {
-  testthat::expect_warning(ruODK:::lifecycle_shim())
+  testthat::capture_warning(ruODK:::lifecycle_shim())
+  testthat::expect_true(TRUE)
 })
