+++ man-roxygen/param-url.R
@@ -1,4 +1,4 @@
-#' @param url The ODK Central base URL without trailing slash.#'
+#' @param url The ODK Central base URL without trailing slash.
 #'   Default: \code{get_default_url()}.
 #'   Set default \code{url} through \code{ru_setup(url="...")}.
 #'   See \code{vignette("Setup", package = "ruODK")}.
