+++ NEWS.md
@@ -1,7 +1,12 @@
-# ruODK (development version)
+# `ruODK` (development version)
+This version addresses ROpenSci reviewer comments from @karissawhiting.
+
 ## Major fixes
 
 ## Minor fixes
+* Drop `. <- NULL` in favour of `utils::globalVariables(".")` (#35)
+* Print settings now hides passwords but instructs how to show them. (#37)
+* `ru_setup()` now prints settings. (#37)
 
 ## Dependencies
 * Moved `rlist` to Imports, as it is now used in `odata_submission_get()` (#6)
@@ -14,12 +19,12 @@
 * Updated workshop companion package [urODK](https://github.com/dbca-wa/urODK)
 
 
-# ruODK 0.6.6
+# `ruODK` 0.6.6
 * The big one has landed: `odata_submission_get()` now defaults to parse 
   submissions from nested lists into a tibble, parse dates and datetimes,
   download and link attachments (#6)
 
-# ruODK 0.6.5
+# `ruODK` 0.6.5
 ## Documentation
 * Use lifecycle badges on functions. Add lifecycle to dependencies, version bump
   `usethis` (#29)
@@ -30,7 +35,7 @@
 * Use dummy imports to silence R CMD check NOTES as per [googledrive](https://github.com/tidyverse/googledrive/blob/050a982cba630503702bdde05a77d727baa36d48/R/googledrive-package.R)'s example (thanks to @jennybc)
 * Drop unused internal helper functions
 
-# ruODK 0.6.4
+# `ruODK` 0.6.4
 ## Data
 *  Use three new test forms to make package smaller and tests faster.
    Use the main test form for example data, including data from two nested tables.
@@ -40,7 +45,7 @@
    downloading attachments.
    The test credentials are unchanged (#23)
 
-# ruODK 0.6.3
+# `ruODK` 0.6.3
 ## Dependencies
 *  [tidyr 1.0.0](https://www.tidyverse.org/articles/2019/09/tidyr-1-0-0/) is out!
   Move `{tidyr}` dependency from GitHub master to CRAN version (#27)
@@ -55,7 +60,7 @@
 *  Add inaugural RMarkdown template "odata" (#26)
 *  Add [binder](https://mybinder.org/) launch button (one click start for #26)
 
-# ruODK 0.6.2
+# `ruODK` 0.6.2
 ## Code
 *  Simplifiy `ru_setup()` to use OData Service URL.
 *  Change all functions to default to `get_default{pid,fid,url,un,pw}()`, partly
@@ -73,7 +78,7 @@
   to vignette "setup". @maelle
 *  Add comparison of similar software to README (#25)
 
-# ruODK 0.6.1
+# `ruODK` 0.6.1
 *  ROpenSci submission review [milestone](https://github.com/dbca-wa/ruODK/milestone/3), 
   [discussion](https://github.com/ropensci/software-review/issues/335).
 *  Updates to documentation (#13 #15 #17)
@@ -90,18 +95,18 @@
 *  Since `submission_detail` is now metadata only, add `submission_get` to download
   submission data.
 
-# ruODK 0.6.0
-*  Version bump and lifecycle bump to indicate that ruODK is ready to be used
+# `ruODK` 0.6.0
+*  Version bump and lifecycle bump to indicate that `ruODK` is ready to be used
   against ODK Central 0.6.
 
-# ruODK 0.6.0.9000
+# `ruODK` 0.6.0.9000
 *  Version bump to match ODK Central's version.
 *  Updated to match ODK Central's API at 0.6 (RESTful and OData) (c.f.).
 *  Add commented out crosslinks to source code and related tests for convenience.
 *  Encryption endpoints (new in 0.6) are not yet supported.
[…]
