### 2019-10-15 #335 @florianm
For the public record:  Both reviewers have given me permission to attribute them as "rev" in the DESCRIPTION through the ROpenSci Slack chat.
