+++ DESCRIPTION
@@ -13,6 +13,10 @@ Authors@R:
              email = "maelle.salmon@yahoo.se", 
              role = "rev", 
              comment = c(ORCID = "0000-0002-2815-0399")), 
+      person(given = "Karissa",
+             family = "Whiting",
+             role = "rev",
+             comment = c(ORCID = "0000-0002-4683-1868")),
       person(given = "DBCA",
              role = c("cph", "fnd")),
       person(given = "NWSFTCP",
