### 2019-10-21 fd4e457
Add, document, and use canned data for Vignette "odata".
