+++ man/fq_fs.Rd
@@ -1,39 +1,38 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{fq_raw}
-\alias{fq_raw}
-\title{OData submission data for an ODK Central form.}
-\format{A list of lists}
+\name{fq_fs}
+\alias{fq_fs}
+\title{JSON form schema for an ODK Central form.}
+\format{The output of `ruODK::form_schema()`, a tibble with columns "type",
+  "name" and "path" and one row per form field.}
 \source{
 See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
+  and `ruODK::form_schema()`.
 }
 \usage{
-fq_raw
+fq_fs
 }
 \description{
 \lifecycle{stable}
 }
 \details{
-The OData response for the submissions of an ODK Central form.
-This form represents a Flora Quadrat, which is a ca 50 by 50 m quadrat of
-a uniform plant community.
-
-The XML and .odkbuild versions for this form are
-available as `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
-and `system.file("extdata", "FloraQuadrat04.odkbuild", package = "ruODK")`,
-respectively.
+The parsed form schema of an ODK Central form.
 
 This data is kept up to date with the data used in vignettes and package
 tests. The data is comprised of test records with nonsensical data.
 The forms used to capture this data are development versions of real-world
 forms.
+
+This data is used to build vignettes offline and without the need for
+credentials to an ODK Central server. The test suite ensures that the
+"canned" data is identical to the "live" data.
 }
 \seealso{
 Other included: \code{\link{fq_data_strata}},
   \code{\link{fq_data_taxa}}, \code{\link{fq_data}},
-  \code{\link{fq_fs}}, \code{\link{fq_meta}},
-  \code{\link{fq_raw_strata}}, \code{\link{fq_raw_taxa}},
+  \code{\link{fq_meta}}, \code{\link{fq_raw_strata}},
+  \code{\link{fq_raw_taxa}}, \code{\link{fq_raw}},
   \code{\link{fq_svc}}
 }
 \concept{included}
