+++ man/fq_data.Rd
@@ -1,21 +1,24 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{fq_raw}
-\alias{fq_raw}
-\title{OData submission data for an ODK Central form.}
-\format{A list of lists}
+\name{fq_data}
+\alias{fq_data}
+\title{Parsed submission data for an ODK Central form.}
+\format{The output of `ruODK::odata_submission_get()` for a set of example
+  data. A tidy tibble referencing the attachments included in the vignettes
+  and documentation at a relative path `attachments/media/<filename>.<ext>`.}
 \source{
 See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
+  and `ruODK::odata_submission_get()`.
 }
 \usage{
-fq_raw
+fq_data
 }
 \description{
 \lifecycle{stable}
 }
 \details{
-The OData response for the submissions of an ODK Central form.
+The parsed OData response for the submissions of an ODK Central form.
 This form represents a Flora Quadrat, which is a ca 50 by 50 m quadrat of
 a uniform plant community.
 
@@ -31,9 +34,9 @@ forms.
 }
 \seealso{
 Other included: \code{\link{fq_data_strata}},
-  \code{\link{fq_data_taxa}}, \code{\link{fq_data}},
-  \code{\link{fq_fs}}, \code{\link{fq_meta}},
-  \code{\link{fq_raw_strata}}, \code{\link{fq_raw_taxa}},
+  \code{\link{fq_data_taxa}}, \code{\link{fq_fs}},
+  \code{\link{fq_meta}}, \code{\link{fq_raw_strata}},
+  \code{\link{fq_raw_taxa}}, \code{\link{fq_raw}},
   \code{\link{fq_svc}}
 }
 \concept{included}
