+++ man/fq_data_strata.Rd
@@ -1,21 +1,24 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{fq_raw_strata}
-\alias{fq_raw_strata}
-\title{OData submission data for a subgroup of an ODK Central form.}
-\format{A list of lists}
+\name{fq_data_strata}
+\alias{fq_data_strata}
+\title{Parsed submission data for a subgroup of an ODK Central form.}
+\format{The output of `ruODK::odata_submission_get()` for a set of example
+  data. A tidy tibble referencing the attachments included in the vignettes
+  and documentation at a relative path `attachments/media/<filename>.<ext>`.}
 \source{
 See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
+  and `ruODK::odata_submission_get()`.
 }
 \usage{
-fq_raw_strata
+fq_data_strata
 }
 \description{
 \lifecycle{stable}
 }
 \details{
-The OData response for the subgroup of an ODK Central form.
+The parsed OData response for the subgroup of an ODK Central form.
 
 This subgroup represents vegetation strata as per the NVIS classification.
 A vegetation stratum is a layer of plants with the same height, and dominated
@@ -28,9 +31,9 @@ The forms used to capture this data are development versions of real-world
 forms.
 }
 \seealso{
-Other included: \code{\link{fq_data_strata}},
-  \code{\link{fq_data_taxa}}, \code{\link{fq_data}},
-  \code{\link{fq_fs}}, \code{\link{fq_meta}},
+Other included: \code{\link{fq_data_taxa}},
+  \code{\link{fq_data}}, \code{\link{fq_fs}},
+  \code{\link{fq_meta}}, \code{\link{fq_raw_strata}},
   \code{\link{fq_raw_taxa}}, \code{\link{fq_raw}},
   \code{\link{fq_svc}}
 }

[…]
