### 2019-11-03 #335 @maelle
Any ETA on your response to reviewers @florianm? The [author guide](https://devguide.ropensci.org/guide-for-authors.html) states "We ask that you respond to reviewers’ comments within 2 weeks of the last-submitted review" (but I know you've been traveling 🙂).
