### 2019-11-25 #335 @florianm
Thanks for the patience, getting stuck into ruODK now. ODK are preparing to release v0.7 of ODK Central, I'll have to work my way through their API updates too.
