### 2019-11-25 dbca-wa/ruODK#42 @florianm
Test fails on unused argument to tempdir(check=TRUE)
## Reproducible example
Reported at https://github.com/ropensci/software-review/issues/335#issuecomment-542019403
by @jmt2080ad

```{r, eval=F}
test-attachment_get.R:80: error: get_one_attachment handles repeat
download and NA filenames
unused argument (check = TRUE)
1: .handleSimpleError(function (e)
   {
       handled <<- TRUE
       test_error <<- e
       options(expressions = expressions_opt_new)
       on.exit(options(expressions = expressions_opt), add = TRUE)
       e$expectation_calls <- frame_calls(11, 2)
       test_error <<- e
       register_expectation(e)
       e$handled <- TRUE
       test_error <<- e
   }, "unused argument (check = TRUE)", quote(tempdir(check = TRUE)))
   at /home/....../test-attachment_get.R:80

```

<summary><strong>Session Info</strong></summary>

```{r}
# utils::sessionInfo()
```
