### 2019-11-25 dbca-wa/ruODK#42 @florianm
I can't reproduce the error on my end. `tempdir(check=FALSE)` is the default shown in the help of `tempdir()`, `check=TRUE` should simply check and re-create the tempdir if required.

@jmt2080ad are you happy for me to close this as "can't reproduce"?
