### 2019-11-25 dbca-wa/ruODK#43 @florianm
Rename parse_datetime to avoid naming conflict with readr::parse_datetime
## Feature
Reported at https://github.com/ropensci/software-review/issues/335#issuecomment-542019403
