### 2019-11-25 dbca-wa/ruODK#34 @florianm
* Added Karissa and Jason as rev to DESC. Pending @jmt2080ad's ORCID.
* Added paragraph to README how to get access to the ODK Central sandbox.
* `ruODK` is backticked in all places unless already formatted as code (code comments, print.ru_settings, license agreement).
* Typo fixed in "Publish the form to..."
* Title of "Contributing" in pkgdown website fixed.
* `usethis` boilerplate deleted.
* Contributing > technical prerequisites now mentions required ODK Central credentials.
* `utils-tody-eval.R` might come in handy in the future, so will keep them around.
* `\code{}` dropped from `@examples` - but still used elsewhere in function docs.
* Rewrote large parts of vignettes to clarify use cases.
* Renamed vignettes to `odata-api` and `restful-api`.

Further fixes
* Enclosed long lines (URLs) with `# nolint start/end` as per https://github.com/MangoTheCat/goodpractice/issues/137#issuecomment-554323393
