### 2019-11-25 dbca-wa/ruODK#37 @florianm
Status quo:

* Hide sensitive data (password) from `print.ru_settings`, but show hint how to view them.
* Helpers `yell_if_missing` and `yell_if_error` will inform the user on missing credentials or failed authentication.
