### 2019-11-25 7c80c86
rOpenSci reviewer comments part 1
Fix #34 - Documentation improvements
Fix #35 - Define "." as global variable
Fix #37 - Hide sensitive settings
Fix #43 - Rename parse_datetime to ru_datetime

#33 Use canned data in vignette "odata-api"
