### 2019-11-25 dbca-wa/ruODK#36 @florianm
It would make sense to retain the `yell_if_missing` gate check for the use case of non-default credentials.

Example:

```{r}
# With defaults, this will fall back to get_default_un() and get_default_pw(), which will yell if the system variables WASTDR_DEFAULT_UN/PW are NULL.
ruODK::project_list()

# This will only be caught and diagnosed by `yell_if_missing(url, un, pw)`, as `get_default_*()` are not called.
ruODK::project_list(url = NULL, un = NULL, pw=NULL)
```
