### 2019-11-25 237c28b
Fix type in examples for odata_submission_get #31
