+++ R/odata_submission_get.R
@@ -85,13 +85,13 @@
 #' )
 #'
 #' form_tables <- ruODK::odata_service_get()
-#' data <- odata_submissions_get() # default: main data table
-#' data <- odata_submissions_get(table = form_tables$url[1]) # same, explicitly
-#' data_sub1 <- odata_submissions_get(table = form_tables$url[2]) # sub-table 1
-#' data_sub2 <- odata_submissions_get(table = form_tables$url[3]) # sub-table 2
+#' data <- odata_submission_get() # default: main data table
+#' data <- odata_submission_get(table = form_tables$url[1]) # same, explicitly
+#' data_sub1 <- odata_submission_get(table = form_tables$url[2]) # sub-table 1
+#' data_sub2 <- odata_submission_get(table = form_tables$url[3]) # sub-table 2
 #'
 #' # Skip one row, return the next 1 rows (top), include total row count
-#' data <- odata_submissions_get(
+#' data <- odata_submission_get(
 #'   table = form_tables$url[1],
 #'   skip = 1,
 #'   top = 1,
@@ -99,7 +99,7 @@
 #' )
 #'
 #' # Point coordinates (lat, lon, alt, acc) need to be renamed
-#' data <- odata_submissions_get(
+#' data <- odata_submission_get(
 #'   table = form_tables$url[1],
 #'   wkt = FALSE
 #' ) %>%
@@ -112,7 +112,7 @@
 #'   )
 #'
 #' # Parse point coordinates into WKT like "POINT (115.8840312 -31.9961844 0)"
-#' data <- odata_submissions_get(
+#' data <- odata_submission_get(
 #'   table = form_tables$url[1],
 #'   wkt = TRUE
 #' )
@@ -197,10 +197,12 @@ odata_submission_get <- function(table = "Submissions",
     magrittr::extract2("name")
 
   if (verbose == TRUE) {
-    message(glue::glue("Found date column: {dttm_cols}.\n"))
+    message(glue::glue("Found date column: {dttm_cols}. "))
   }
   for (colname in dttm_cols) {
-    if (verbose == TRUE) message(glue::glue("Parsing {colname} as {tz}...\n"))
+    if (verbose == TRUE) {
+      message(glue::glue("\nParsing {colname} with timezone {tz}...\n"))
+    }
     sub <- sub %>%
       ru_datetime(
         orders = orders,
@@ -217,7 +219,7 @@ odata_submission_get <- function(table = "Submissions",
     intersect(names(sub))
 
   if (verbose == TRUE) {
-    message(glue::glue("Downloading attachments...\n"))
+    message(glue::glue("\nDownloading attachments...\n"))
   }
 
   sub <- sub %>% dplyr::mutate_at(
@@ -234,6 +236,30 @@ odata_submission_get <- function(table = "Submissions",
       pw = pw
     )
   )
+
+  # Parse geopoints
+  gp_cols <- fs %>%
+    dplyr::filter(type=="geopoint") %>%
+    magrittr::extract2("name")
+
+  if (verbose == TRUE) {
+    message(glue::glue("\nFound geopoint column: {gp_cols}.\n"))
+  }
+  for (colname in gp_cols) {
+    if (verbose == TRUE) {
+      message(glue::glue("\nParsing {colname} into lat/lon/alt...\n"))
+    }
+
+    # TODO make this work
+    # split_geopoint <- function(data, cn){
+    #   dplyr::mutate(data, gps = str_replace(cn, "POINT \\(","")) %>%
+    #   dplyr::mutate(., gps = str_replace(gps, "\\)",""))  %>%
+    #   tidyr::separate(., gps, c("longitude", "latitude", "altitude"), sep = " ")
+    # }
[…]
