### 2019-11-26 9eade75
Improve docs, split geopoint
Fix #7 Split WPT POINT (ODK geopoint)
#34 crossreference ruODK functions
