+++ man-roxygen/param-iid.R
@@ -1,1 +1,2 @@
-#' @param iid The `instance_id`, a UUID, as returned by `submission_list`.
+#' @param iid The `instance_id`, a UUID, as returned by
+#'   \code{\link{submission_list}}.
