+++ R/data.R
@@ -139,11 +139,11 @@
 #' The forms used to capture this data are development versions of real-world
 #' forms.
 #'
-#' @format The output of `ruODK::odata_submission_get()` for a set of example
+#' @format The output of \code{\link{odata_submission_get}} for a set of example
 #'   data. A tidy tibble referencing the attachments included in the vignettes
 #'   and documentation at a relative path `attachments/media/<filename>.<ext>`.
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
-#'   and `ruODK::odata_submission_get()`.
+#'   and \code{\link{odata_submission_get}}.
 #' @family included
 "fq_data"
 
@@ -163,11 +163,11 @@
 #' The forms used to capture this data are development versions of real-world
 #' forms.
 #'
-#' @format The output of `ruODK::odata_submission_get()` for a set of example
+#' @format The output of \code{\link{odata_submission_get}} for a set of example
 #'   data. A tidy tibble referencing the attachments included in the vignettes
 #'   and documentation at a relative path `attachments/media/<filename>.<ext>`.
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
-#'   and `ruODK::odata_submission_get()`.
+#'   and \code{\link{odata_submission_get}}.
 #' @family included
 "fq_data_strata"
 
@@ -189,10 +189,10 @@
 #' The forms used to capture this data are development versions of real-world
 #' forms.
 #'
-#' @format The output of `ruODK::odata_submission_get()` for a set of example
+#' @format The output of \code{\link{odata_submission_get}} for a set of example
 #'   data. A tidy tibble referencing the attachments included in the vignettes
 #'   and documentation at a relative path `attachments/media/<filename>.<ext>`.
 #' @source See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
-#'   and `ruODK::odata_submission_get()`.
+#'   and \code{\link{odata_submission_get}}.
 #' @family included
 "fq_data_taxa"
