+++ R/submission_detail.R
@@ -8,7 +8,9 @@
 #' @template param-url
 #' @template param-auth
 #' @return A nested list of submission metadata.
+# nolint start
 #' @seealso \url{https://odkcentral.docs.apiary.io/#reference/forms-and-submissions/submissions/getting-submission-details}
+# nolint end
 #' @family restful-api
 #' @export
 #' @examples
