+++ NEWS.md
@@ -1,3 +1,5 @@
+# ruODK (development version)
+
 # `ruODK` (development version)
 This version addresses ROpenSci reviewer comments from 
 @karissawhiting and @jmt2080ad.
