+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: ruODK
 Title: An R Client for the ODK Central API
-Version: 0.6.6.9002
+Version: 0.6.6.9003
 Authors@R: 
     c(person(given = c("Florian", "W."),
              family = "Mayer",
