### 2019-11-26 9d92d59
fix NEWS.md
