### 2019-11-27 dbca-wa/ruODK#42 @florianm
Also, what's your sessionInfo? This could be an issue with R < 3.5 as per https://github.com/mkearney/pkgverse/issues/3
