### 2019-11-27 dbca-wa/ruODK#38 @florianm
* Renamed `submission_get` into `get_one_submission`
* Created `submission_get` as shown above, so that submissions are downloaded for all supplied `iid` submission instance ids.
* Adjusted docs, tests, simplified vignettes.
