### 2019-11-27 17fd4a8
Second batch of rOpenSci reviewer comments.
Fix #33 - all vignettes use canned data
Fix #38 - submission_get now in batch mode
Fix #45 - add asciicast to README
Add new logo
