+++ tests/testthat/test-attachment_list.R
@@ -9,7 +9,7 @@ test_that("attachment_list works", {
     pw = get_test_pw()
   )
 
-  al <- attachment_list(
+  al <- get_one_submission_attachment_list(
     sl$instance_id[[1]],
     pid = get_test_pid(),
     fid = get_test_fid(),
@@ -17,7 +17,15 @@ test_that("attachment_list works", {
     un = get_test_un(),
     pw = get_test_pw()
   )
-  al %>% knitr::kable(.)
+
+  all <- attachment_list(
+    sl$instance_id,
+    pid = get_test_pid(),
+    fid = get_test_fid(),
+    url = get_test_url(),
+    un = get_test_un(),
+    pw = get_test_pw()
+  )
 
   # attachment_list returns a tibble
   testthat::expect_equal(class(al), c("tbl_df", "tbl", "data.frame"))
