### 2019-11-27 dbca-wa/ruODK#40 @florianm
Good catch, cheers!
The problem seems to be between throwing `mutate_at` and `attachment_get` when parsing [attachments odata_submission_get](https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L249-L262) resulting in a list column.
