### 2019-11-27 dbca-wa/ruODK#42 @florianm
Current status:

* `ruODK` requires R >=3.4.
* `tempdir(check=T/F)` works with R >=3.5 but fails with R <=3.4.
* Should we drop `check` or bump requirements?
