### 2019-11-28 ef4e878
Fix #42 - drop "check=TRUE" from tempdir in tests
