### 2019-12-03 b15499f
Move isodt_to_local to own source file
