+++ tests/testthat/test-ru_setup.R
@@ -14,6 +14,7 @@ test_that("ru_setup does not update settings if given NULL", {
     test_fid = NULL,
     test_fid_zip = NULL,
     test_fid_att = NULL,
+    test_fid_gap = NULL,
     test_url = NULL,
     test_un = NULL,
     test_pw = NULL
@@ -35,6 +36,7 @@ test_that("ru_setup does not update settings if given NULL", {
   testthat::expect_equal(x$test_fid, get_test_fid())
   testthat::expect_equal(x$test_fid_zip, get_test_fid_zip())
   testthat::expect_equal(x$test_fid_att, get_test_fid_att())
+  testthat::expect_equal(x$test_fid_gap, get_test_fid_gap())
 })
 
 test_that("ru_setup resets settings if given empty string", {
@@ -50,6 +52,7 @@ test_that("ru_setup resets settings if given empty string", {
   test_fid <- get_test_fid()
   test_fid_zip <- get_test_fid_zip()
   test_fid_att <- get_test_fid_att()
+  test_fid_gap <- get_test_fid_gap()
 
   # Hammertime
   ru_setup(
@@ -62,6 +65,7 @@ test_that("ru_setup resets settings if given empty string", {
     test_fid = "",
     test_fid_zip = "",
     test_fid_att = "",
+    test_fid_gap = "",
     test_url = "",
     test_un = "",
     test_pw = ""
@@ -81,6 +85,7 @@ test_that("ru_setup resets settings if given empty string", {
   testthat::expect_warning(get_test_fid())
   testthat::expect_warning(get_test_fid_zip())
   testthat::expect_warning(get_test_fid_att())
+  testthat::expect_warning(get_test_fid_gap())
 
   testthat::expect_equal(x$url, "")
   testthat::expect_equal(x$un, "")
@@ -92,6 +97,7 @@ test_that("ru_setup resets settings if given empty string", {
   testthat::expect_equal(x$test_fid, "")
   testthat::expect_equal(x$test_fid_zip, "")
   testthat::expect_equal(x$test_fid_att, "")
+  testthat::expect_equal(x$test_fid_gap, "")
 
   # Reset
   ru_setup(
@@ -104,7 +110,8 @@ test_that("ru_setup resets settings if given empty string", {
     test_pid = test_pid,
     test_fid = test_fid,
     test_fid_zip = test_fid_zip,
-    test_fid_att = test_fid_att
+    test_fid_att = test_fid_att,
+    test_fid_gap = test_fid_gap
   )
 })
 
