+++ man/isodt_to_local.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/isodt_to_local.R
 \name{isodt_to_local}
 \alias{isodt_to_local}
 \title{Parse an ISO8601 datetime string to a timezone aware datetime.}
