### 2019-12-03 b9833b9
Update contributing guidelines with new test form setting required for #46
