### 2019-12-03 e14d1e9
Midway through #46, preparing rewrite of odata_submission_get/parse
