### 2019-12-04 1d88387
Fix #46: require tidyr 1.0.0.9000
