+++ tests/testthat/test-odata_submission_parse.R
@@ -11,7 +11,7 @@ test_that("odata_submission_parse works with gaps in first submission", {
     parse = FALSE
   )
 
-  testthat::expect_error(
+  testthat::expect_warning(
     fresh_raw_parsed <- odata_submission_get(
       pid = get_test_pid(),
       fid = get_test_fid_gap(),
@@ -23,6 +23,7 @@ test_that("odata_submission_parse works with gaps in first submission", {
       local_dir = t
     )
 
+    # https://github.com/dbca-wa/ruODK/issues/46
     # Throws error:
     # test-odata_submission_parse.R:13: error:
     # odata_submission_get works with one known dataset
@@ -30,9 +31,9 @@ test_that("odata_submission_parse works with gaps in first submission", {
     # <vctrs_unspecified>.
   )
 
-  # fresh_parsed <- fresh_raw %>% odata_submission_parse()
-  # testthat::expect_gte(nrow(fresh_parsed), length(fresh_raw$value))
-  # testthat::expect_gte(nrow(fresh_parsed), nrow(fresh_raw_parsed))
+  fresh_parsed <- fresh_raw %>% odata_submission_parse()
+  testthat::expect_gte(nrow(fresh_parsed), length(fresh_raw$value))
+  testthat::expect_gte(nrow(fresh_parsed), nrow(fresh_raw_parsed))
 
   # testthat::expect_equal(
   #   class(fresh_raw_parsed$encounter_start_datetime),
