+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: ruODK
 Title: An R Client for the ODK Central API
-Version: 0.6.6.9005
+Version: 0.6.6.9006
 Authors@R: 
     c(person(given = c("Florian", "W."),
              family = "Mayer",
@@ -50,7 +50,7 @@ Imports:
     purrr (>= 0.3.2.9000),
     stringr (>= 1.4.0),
     tibble (>= 2.1.3),
-    tidyr (>= 1.0.0),
+    tidyr (>= 1.0.0.9000),
     xml2 (>= 1.2.2)
 Suggests: 
     covr (>= 3.3.1),
@@ -65,8 +65,9 @@ Suggests:
     vcr (>= 0.3.0)
 Remotes:
     github::tidyverse/purrr,
-    github::r-lib/rlang,
+    github::tidyverse/tidyr,
     github::r-lib/lifecycle,
+    github::r-lib/rlang,
     github::r-lib/usethis
 VignetteBuilder: 
     knitr
