### 2019-12-06 9957c38
Add colours and clisymbols to messages
