+++ data-raw/record_asciicast.R
@@ -1,4 +1,4 @@
-cast <- asciicast::record("data-raw/make_asciicast.R")
-asciicast::write_json(cast, "data-raw/odata.json")
+cast <- asciicast::record(here::here("data-raw/make_asciicast.R"))
+asciicast::write_json(cast, here::here("data-raw/odata.json"))
 asciicast::write_svg(cast, here::here("man", "figures", "odata.svg"))
 asciicast::play(cast)
