### 2019-12-16 #335 @florianm
Hi @maelle - apologies for the delay, hoping to finish this issue this week. Turtle season is in full swing!

Remaining work:

* Dates and attachments come out as list columns. Something is going wrong in all the mutate_at/purrring in `odata_submission_parse`.
* I'd like to make downloading attachment files (which can be several GBs) optional and rather hot-link the files. This would be handy for e.g. Shiny apps as previews.
* Add a few examples where missing, see last open issue.
* Use vcr in tests if I can get it to work. Maybe that goes in the "later" basket.
