### 2019-12-19 9c65259
Add more examples. #32
