+++ tests/testthat/test-attachment_get.R
@@ -16,7 +16,6 @@ test_that("attachment_get works", {
 
   fresh_parsed <- fresh_raw %>%
     odata_submission_parse() %>%
-    janitor::clean_names() %>%
     dplyr::mutate(
       quadrat_photo = attachment_get(
         id,
@@ -33,7 +32,6 @@ test_that("attachment_get works", {
 
   fresh_parsed_sep <- fresh_raw %>%
     odata_submission_parse() %>%
-    janitor::clean_names() %>%
     dplyr::mutate(
       quadrat_photo = attachment_get(
         id,
