+++ man/strip_uuid.Rd
@@ -15,6 +15,10 @@ The string with every occurrence of "uuid:" deleted.
 \description{
 \lifecycle{stable}
 }
+\examples{
+strip_uuid("uuid:1234")
+strip_uuid("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
+}
 \seealso{
 Other utilities: 
 \code{\link{\%>\%}()},
