### 2019-12-19 6c08736
Fix test failure, add examples, close #32
