+++ tests/testthat/test-attachment_get.R
@@ -61,13 +61,15 @@ test_that("attachment_url works", {
   pid <- get_test_pid()
   fid <- get_test_fid()
 
-  expected_url <-
-    glue::glue("{url}/v1/projects/14/forms/{fid}/submissions/{uuid}/attachments/{fn}")
+  expected_url <- glue::glue(
+    "{url}/v1/projects/14/forms/{fid}/submissions/{uuid}/attachments/{fn}"
+  )
   calculated_url <- ruODK:::attachment_url(uuid,
-                                           fn,
-                                           pid = pid,
-                                           fid = fid,
-                                           url = url)
+    fn,
+    pid = pid,
+    fid = fid,
+    url = url
+  )
 
   testthat::expect_equal(calculated_url, expected_url)
 })
@@ -87,10 +89,11 @@ test_that("get_one_attachment handles repeat download and NA filenames", {
 
   pth <- fs::path(t, fn)
   src <- ruODK:::attachment_url(uuid,
-                                fn,
-                                pid = pid,
-                                fid = fid,
-                                url = url)
+    fn,
+    pid = pid,
+    fid = fid,
+    url = url
+  )
 
   # Happy path: get one attachment should work
   fn_local <- get_one_attachment(
