+++ R/submission_get.R
@@ -83,7 +83,7 @@ get_one_submission <- function(iid,
 #' @examples
 #' \dontrun{
 #' # Step 1: Setup ruODK with OData Service URL (has url, pid, fid)
-#' ruODK::ru_setup(svc="...")
+#' ruODK::ru_setup(svc = "...")
 #'
 #' # Step 2: List all submissions of form
 #' sl <- submission_list()
