+++ R/attachment_list.R
@@ -106,7 +106,7 @@ get_one_submission_attachment_list <- function(iid,
 #' @examples
 #' \dontrun{
 #' # Step 1: Setup ruODK with OData Service URL (has url, pid, fid)
-#' ruODK::ru_setup(svc="...")
+#' ruODK::ru_setup(svc = "...")
 #'
 #' # Step 2: List all submissions of form
 #' sl <- submission_list()
