+++ tests/testthat/test-project_list.R
@@ -1,11 +1,11 @@
 test_that("project_list works", {
   # TODO enable vcr for all tests once vcr works with httr basicauth
   vcr::use_cassette("project_list", {
-  p <- project_list(
-    url = get_test_url(),
-    un = get_test_un(),
-    pw = get_test_pw()
-  )
+    p <- project_list(
+      url = get_test_url(),
+      un = get_test_un(),
+      pw = get_test_pw()
+    )
   })
   testthat::expect_true(nrow(p) > 0)
 
