### 2020-01-10 #335 @stefaniebutland
Happy new year @florianm. Leaving a note here so you can let me know (after turtles, and ruODK completion 😄) when you're ready to dig into a blog post.
