### 2020-01-16 dbca-wa/ruODK#40 @florianm
Found a fix. `purrr::pmap` generates a list, which needed to be cast into its type - `as.character` did the trick.
