### 2020-01-16 c0c245b
Drop old default issue template in favour of specific ones.
