+++ ISSUE_TEMPLATE.md
@@ -1,13 +1,0 @@
-Please briefly describe your problem and what output you expect. 
-
-Please include a minimal reproducible example (AKA a reprex). 
-If you've never heard of a [reprex](http://reprex.tidyverse.org/) before, 
-start by reading <https://www.tidyverse.org/help/#reprex>.
-
----
-
-Brief description of the problem
-
-```r
-# insert reprex here
-```
