### 2020-01-16 9a41b92
Fix #40 parse attachment file paths to character
* Refresh canned data (contains parsed file paths)
* Add test to verify att file paths are of class character
* Drop number of submissions (hard to parse, was incorrect) from verbose message at `odata_submission_get`
* bump version number
