### 2020-01-16 b875e32
bump version number, update README
