+++ README.md
@@ -184,6 +184,7 @@ read-permitted ODK Central web user.
     #>   Test ODK Central Form ID: build_Flora-Quadrat-0-4_1564384341 
     #>   Test ODK Central Form ID (ZIP tests): build_Spotlighting-0-6_1558333698 
     #>   Test ODK Central Form ID (Attachment tests): build_Flora-Quadrat-0-1_1558330379 
+    #>   Test ODK Central Form ID (Parsing tests): build_Turtle-Track-or-Nest-1-0_1569907666 
     #>   Test ODK Central URL: https://sandbox.central.opendatakit.org 
     #>   Test ODK Central Username: Florian.Mayer@dbca.wa.gov.au 
     #>   Test ODK Central Password: run ruODK::get_test_pw() to show
@@ -215,12 +216,12 @@ proj %>% head() %>% knitr::kable(.)
 
 | id | name                     | forms | app\_users | created\_at         | updated\_at | last\_submission    | archived |
 | -: | :----------------------- | ----: | ---------: | :------------------ | :---------- | :------------------ | :------- |
+| 59 | 1Horizon test project    |     1 |          2 | 2020-01-13 12:45:32 | NA          | 2020-01-13 15:06:15 | FALSE    |
+| 57 | AB’s                     |     1 |          1 | 2020-01-06 00:56:00 | NA          | NA                  | FALSE    |
 |  8 | AmenazasRD               |     0 |          0 | 2019-04-14 06:21:00 | NA          | NA                  | FALSE    |
 |  9 | BAFCO Test               |     1 |          1 | 2019-04-19 00:30:45 | NA          | 2019-04-19 01:26:49 | FALSE    |
 | 47 | Building Materials Reuse |     0 |          0 | 2019-10-16 05:07:56 | NA          | NA                  | FALSE    |
 | 11 | CaVaTeCo                 |     2 |          1 | 2019-04-30 06:43:22 | NA          | 2019-04-30 08:07:55 | FALSE    |
-| 18 | CherryPatch              |     2 |          1 | 2019-05-27 21:09:33 | NA          | 2019-05-28 17:32:15 | FALSE    |
-| 32 | Collect3289v2            |     1 |          2 | 2019-08-07 12:52:56 | NA          | 2019-08-07 13:44:19 | FALSE    |
 
 ``` r
 # List forms of default project
@@ -238,6 +239,7 @@ frms %>% knitr::kable(.)
 | Spotlighting 0.6              | build\_Spotlighting-0-6\_1558333698              |         | open    | 18          | 2019-05-20 06:30:21 |              57 | <florian.mayer@dbca.wa.gov.au> | NA                  | 2019-10-09 01:15:23 | 456daaa9a4f96670e6eef3cf4a7dd0db |
 | Spotlighting Survey End 0.3   | build\_Spotlighting-Survey-End-0-3\_1558320208   |         | open    | 3           | 2019-05-20 02:44:38 |              57 | <florian.mayer@dbca.wa.gov.au> | NA                  | 2019-10-09 02:43:40 | 5fdfac8e773834b1267f7ca7e1c9a428 |
 | Spotlighting Survey Start 0.3 | build\_Spotlighting-Survey-Start-0-3\_1558320795 |         | open    | 9           | 2019-05-20 02:53:50 |              57 | <florian.mayer@dbca.wa.gov.au> | NA                  | 2019-10-09 01:15:25 | f548a064cca13bca746f3c0b1a8b5a32 |
[…]
