### 2020-01-16 #335 @florianm
Update:
* Fixed https://github.com/dbca-wa/ruODK/issues/40 for attachments
* Hotlinking attachments is not yet possible (after discussion with ODK devs)
* All issues I see as critical are fixed
* ruODK has been extensively used in anger, looks robust enough now

I'll address the review comments here shortly.
