### 2020-01-28 dbca-wa/ruODK#49 @florianm
odata_submission_get: parse but don't download
## Feature
HT @karissawhiting

* Make it possible to parse the data, but not download the attachments.
* What URL parameter should we add for this? Or default local_dir to NULL as toggle to not download?

When ODK Central allows hotlinking attachments (sometime in the future),it would be useful to store the fully qualified attachment URL if not downloading attachments.

Source: https://github.com/ropensci/software-review/issues/335
