### 2020-01-28 0917cc4
Fix #49: parse odata submissions, don't download attachments
