+++ tests/testthat/test-odata_submission_parse.R
@@ -8,7 +8,8 @@ test_that("odata_submission_parse works with gaps in first submission", {
     url = get_test_url(),
     un = get_test_un(),
     pw = get_test_pw(),
-    parse = FALSE
+    parse = FALSE,
+    verbose=TRUE
   )
 
   testthat::expect_warning(
