### 2020-01-28 #335 @florianm
@maelle @karissawhiting @jmt2080ad thanks for your patience and the excellent reviews!

I've addressed all comments, added some more features, tested ruODK in production over the last few months and believe we're ready for the next step.

Answers to your comments below! I've linked to the GH issues for brevity.

# Reviewer comments Karissa Whiting

## Build

> I wasn't able to build without first specifying ODK credentials (because of api.Rmd and odata.Rmd vignettes). I’m not sure if there are established best practices with API packages and this issue, but @noamross pointed out this rcites package vignette as an example of one way this is addressed in API packages. Vignettes decide whether or not to eval code chunks based on an environment variable (in this case NOT_CRAN) and both ways are tested in the Travis-CI testing grid.

The vignettes now build from canned data.
https://github.com/dbca-wa/ruODK/issues/33
https://twitter.com/fistful_of_bass/status/1186422684504092673

## README

> You could add a bit more description about ODK Central Sandbox in this doc to help with multiple points of entry. New prospective users may be looking for the quickest way to test and play around with ODK/ruODK to see if it fits the needs of their project, and it seems this would be via a Sandbox account. To this point, you could adjust the "ODK Central Setup" section to say something like “Create a web user account on an ODK Central instance, or request an account on Sandbox..."

Rewrote https://dbca-wa.github.io/ruODK/index.html#access-to-an-odk-central-instance
to lower the barier for entry :-)

Detailed summary on documentation improvements at https://github.com/dbca-wa/ruODK/issues/34

> Put “(ruODK)” in backticks for consistency (in CONTRIBUTING doc as well)

`ruODK` backticked.

> Typo in link “Publish the formto ODK Central”.

Typo fixed. I'm also running a [spellcheck](https://github.com/dbca-wa/ruODK/blob/master/data-raw/make_release.R#L20) during the release process.

## Tests and CONTRIBUTING

> There is a barrier to entry in contributing to the package and running tests because of the need for testing credentials. That being said, the CONTRIBUTING document and the new github form you set up to request credentials definitely does help with this. Is there a way to set up a public set of ODK credentials using a dummy account? If not, consider me a +1 vote for setting up cached responses in the future when the ODK API is more stable if it will provide a way for tests to work out of the box :)

Until all tests are converted to vcr (the CRAN 5MB package size limit might blow a fuse),
the only way to contribute without compromising the ODK Central sandbox's security
is to request an account, either through the account request issue template
https://github.com/dbca-wa/ruODK/issues/new/choose or the ODK forum.

Also added friendly instructions to the README https://dbca-wa.github.io/ruODK/index.html#access-to-an-odk-central-instance.

> Title of CONTRIBUTING is NA.

Fixed: https://dbca-wa.github.io/ruODK/CONTRIBUTING.html

> You still have some leftover language from the usethis contributing doc. For example: “Before you do a pull request, you should always file an issue and make sure someone from the tidyverse team…”

Usethis boilerplate cleaned up.

> "To run tests, you’ll need access to the ODK Central sandbox instance…" I think this should go right at the top or even in the “prerequisites” section of technical contributions.

Retructured [contributing guide](https://dbca-wa.github.io/ruODK/CONTRIBUTING.html)
and added account request issue template.

Also https://github.com/dbca-wa/ruODK/issues/20

## Namespace

> Not all exported functions have examples, but some internal helper functions are exported that may not need to be (e.g. prepend_uuid(), yell_if_missing()).

Added examples for all exported functions.
Helpers are still exported, as they might be useful for manual parsing.
Helpers are grouped in docs as family
["utilities"](https://dbca-wa.github.io/ruODK/reference/index.html#section-utilities).

> Do you need utils-tidy-eval.R, and these rlang functions exported?

I intended to streamline some functions using NSE, which will require tidy eval.
This prepares us for the day I get my head around NSE.

## Vignettes
> I’m having some trouble distinguishing between the OData API and RESTful API functionalities. Do these workflows mirror each other? Does one workflow have functionality available that the other does not? I think some additional delineation between these could be helpful. Here are a few ideas on how to address this (suggestions only):

> If these are parallel workflows, consider having the vignette structures mirror each other, and add the same blurb at the top of each (like “three ways to happiness") with links to the other vignette. If there are any large benefits or drawbacks of one method or the other, call them out there or in the README.

> Does your preferred workflow consist of a mixture of both OData and RESTful calls? If so, consider creating just a typical workflow vignette that goes through a case study using both kinds of functions.

Vignettes are re-worded and restructured and put into context of each other
both at their introductory paragraphs and in the README.

While there is no preferred workflow, the most elegant workflow is the
OData path, which is shown both in the README's screencast and part of the
Rmd template.

https://github.com/dbca-wa/ruODK/issues/34

> Consider renaming vignette file names to odata-api, and restful-api

Done, also vignette titles are adjusted.
https://github.com/dbca-wa/ruODK/issues/34

## code

> I see several functions have . <- NULL. I’m assuming this is to avoid “no visible bindings for global variables”. One solutions is to add utils::globalVariables(".") in ruODK.R instead of specifying this in each function.

Done, and shamelessly stolen into my other packages.
https://github.com/dbca-wa/ruODK/issues/35

> Do you need yell_if_missing() in functions that have the get functions as their default args? In those cases, it looks like if you are missing credentials the get_default_<things>() functions already do the yelling, but I may be missing some edge case here.

Retained for the use case of directly specified, non-default credentials:
https://github.com/dbca-wa/ruODK/issues/36#issuecomment-558040212

> ru_setup()
I was expecting a confirmation message of some sort.
Maybe it could print ru_settings() when you run it by default?
Is there a way to actually check the authentication of your username/project/form
credentials here or in another function?
ggmaps::register_google() has an optional argument to write = TRUE where it will
write credentials to your .Renviron. Just wanted to point this out as a cool enhancement option.

Made settings great again: helpfully verbose, but not Wikileaks-level verbose.
https://github.com/dbca-wa/ruODK/issues/37

> @examples sections are formatted as R code so you can take out \code{ } tags (e.g. ru_setup())

Fixed and included in https://github.com/dbca-wa/ruODK/issues/34

> odata_submission_get()
    It could be helpful to separate arguments for parsing the data and getting attachments in this function.
    It looks like odata_submission_get() returns all submissions, but the equivalent function for the RESTful workflow submission_get() returns one submission at a time. I expected these functions to mirror each other in terms of outputs. Perhaps it would make sense to turn this code chunk you have in the vignette into a function? This would mirror the attachment_get() / get_one_attachment() functions too.

Mostly addressed in https://github.com/dbca-wa/ruODK/issues/38
Parse but don't download: https://github.com/dbca-wa/ruODK/issues/49

# Reviewer comments Jason Taylor

Discussed in person with Jason during [rOpenSci beer review](https://twitter.com/fistful_of_bass/status/1183811515067043840)
(apologies to Karissa I didn't make it to NY!)

## Installation
> Building vignettes require a properly setup account. Even with a
setup account, devtools::install(build_vignettes = TRUE) fails.

See KW's comment: fixed by using canned data under the bonnet,
but showing the code to use live data.

> When attempting to build vignettes via devtools fails, the user
password is printed to the console. ru_settings() also prints the
password to the console. I would feel better if that was hidden from
the user.

Addressed in https://github.com/dbca-wa/ruODK/issues/37

## Notes on setup functions and vignette

> Why provide a 'hard way' to pass credentials via a function call? My
sense is that this will encourage bad practices in writing code, as
this would give users a way of hard coding passwords in workflows,
which is typically discouraged.

Re-wrote vignette [setup](https://dbca-wa.github.io/ruODK/articles/setup.html)
to discuss the preferred method (`.Renviron`) and warn against bad practices.

Use cases of "the hard way" are the tests, where test functions call different
projects and forms explicitly.

## Notes on API functions and vignette

> parse_datetime() appears to parse any column name with the text
"time" in it by default or a user defined string. This could
potentially try to parse unintended columns containing words like
"candidate" or "centimeter". I think this functionality might be
best left to the user to define for themselves.

We now use the form schema to detect date and datetime columns.

> Also, parse_datetime() is a readr function. During my review,
those namespaces became confused in a couple of the examples and I
was no longer able to run ruODK vignette code. Would a different
name be better suited?

Renamed to `ru_datetime`: https://github.com/dbca-wa/ruODK/issues/43

## Notes on accessing odata functions and vignette
>  This error message should quote email address.
   Not Found (HTTP 404). Failed to get desired response from server
   https://sandbox.central.opendatakit.org/ as user
   "myemail@email.com".

Settings print now a more intelligible error message.

https://github.com/dbca-wa/ruODK/issues/37

https://github.com/dbca-wa/ruODK/issues/16

> In the odata vignette you have an expression fq_svc <- odata_service_get().
Everywhere else you define the namespace for
an ruODK function, i.e. ruODK::odata_service_get(). For
constancy I would edit this to include the namespace here and
elsewhere where that is missing.

Namespaced all the things!

> Also in the odata vignette, in the section "OData service document",
there appears to be a network call to ODK Central to assign the
fc_svc variable. The next comment however indicates that the
canned data is used. It doesn't appear that the datasets associated
with this package that are loaded early in this vignette are used.

Vignettes should only use canned data, but show the code for live data
https://github.com/dbca-wa/ruODK/issues/33

> In the odata vignette, in the section "Summarising data", the text
is a bit odd as it implies the previous leaflet examples
performed data transformations. Consider revising the language.

I think I've revised the language there.

> The rmarkdown template suggestion is great and the template would
make a great quick guide.

Done and battle-tested in four workshops! Also included a binderhub link in the README.
