### 2020-01-29 dbca-wa/ruODK#49 @florianm
Test: https://github.com/dbca-wa/ruODK/blob/master/tests/testthat/test-odata_submission_get.R#L3-L23

Param "download" toggles attachment download https://dbca-wa.github.io/ruODK/reference/odata_submission_get.html
