### 2020-02-06 7e4da0c
Fix #48: Use httr::RETRY
