+++ R/submission_detail.R
@@ -44,14 +44,15 @@ submission_detail <- function(iid,
                               un = get_default_un(),
                               pw = get_default_pw()) {
   yell_if_missing(url, un, pw, pid = pid, fid = fid)
-  glue::glue("{url}/v1/projects/{pid}/forms/{fid}/submissions/{iid}") %>%
-    httr::GET(
-      httr::add_headers(
-        "Accept" = "application/json; extended",
-        "X-Extended-Metadata" = "true"
-      ),
-      httr::authenticate(un, pw)
-    ) %>%
+  httr::RETRY(
+    "GET",
+    glue::glue("{url}/v1/projects/{pid}/forms/{fid}/submissions/{iid}"),
+    httr::add_headers(
+      "Accept" = "application/json; extended",
+      "X-Extended-Metadata" = "true"
+    ),
+    httr::authenticate(un, pw)
+  ) %>%
     yell_if_error(., url, un, pw) %>%
     httr::content(.) %>%
     {
