+++ README.md
@@ -217,7 +217,7 @@ proj %>% head() %>% knitr::kable(.)
 
 | id | name                     | forms | app\_users | created\_at         | updated\_at | last\_submission    | archived |
 | -: | :----------------------- | ----: | ---------: | :------------------ | :---------- | :------------------ | :------- |
-| 59 | 1Horizon test project    |     1 |          2 | 2020-01-13 12:45:32 | NA          | 2020-01-13 15:06:15 | FALSE    |
+| 59 | 1Horizon test project    |     1 |          3 | 2020-01-13 12:45:32 | NA          | 2020-01-23 14:20:13 | FALSE    |
 | 57 | AB’s                     |     1 |          1 | 2020-01-06 00:56:00 | NA          | NA                  | FALSE    |
 |  8 | AmenazasRD               |     0 |          0 | 2019-04-14 06:21:00 | NA          | NA                  | FALSE    |
 |  9 | BAFCO Test               |     1 |          1 | 2019-04-19 00:30:45 | NA          | 2019-04-19 01:26:49 | FALSE    |
@@ -258,41 +258,41 @@ meta <- ruODK::form_schema()
 meta %>% knitr::kable(.)
 ```
 
-| type      | name                                  | path                            |
-| :-------- | :------------------------------------ | :------------------------------ |
-| structure | meta                                  | Submissions                     |
-| dateTime  | encounter\_start\_datetime            | Submissions                     |
-| string    | reporter                              | Submissions                     |
-| string    | device\_id                            | Submissions                     |
-| structure | location                              | Submissions                     |
-| structure | habitat                               | Submissions                     |
-| repeat    | vegetation\_stratum                   | Submissions                     |
-| structure | perimeter                             | Submissions                     |
-| repeat    | taxon\_encounter                      | Submissions                     |
-| dateTime  | encounter\_end\_datetime              | Submissions                     |
-| string    | instanceID                            | Submissions.meta                |
-| string    | area\_name                            | Submissions.location            |
-| binary    | quadrat\_photo                        | Submissions.location            |
-| geopoint  | corner1                               | Submissions.location            |
-| select1   | morphological\_type                   | Submissions.habitat             |
-| binary    | morphological\_type\_photo            | Submissions.habitat             |
-| select1   | nvis\_level3\_broad\_floristic\_group | Submissions.vegetation\_stratum |
-| decimal   | max\_height\_m                        | Submissions.vegetation\_stratum |
-| select1   | foliage\_cover                        | Submissions.vegetation\_stratum |
-| string    | dominant\_species\_1                  | Submissions.vegetation\_stratum |
-| string    | dominant\_species\_2                  | Submissions.vegetation\_stratum |
-| string    | dominant\_species\_3                  | Submissions.vegetation\_stratum |
-| string    | dominant\_species\_4                  | Submissions.vegetation\_stratum |
-| geopoint  | corner2                               | Submissions.perimeter           |
-| geopoint  | corner3                               | Submissions.perimeter           |
-| geopoint  | corner4                               | Submissions.perimeter           |
-| binary    | mudmap\_photo                         | Submissions.perimeter           |
-| string    | field\_name                           | Submissions.taxon\_encounter    |
-| binary    | photo\_in\_situ                       | Submissions.taxon\_encounter    |
-| geopoint  | taxon\_encounter\_location            | Submissions.taxon\_encounter    |
[…]
