+++ .travis.yml
@@ -9,5 +9,11 @@ r:
 - release
 - devel
 
+before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
+  deploy:
+    provider: script
+    script: Rscript -e 'pkgdown::deploy_site_github()'
+    skip_cleanup: true
+
 after_success:
   - Rscript -e 'covr::codecov()'
