+++ .travis.yml
@@ -10,10 +10,11 @@ r:
 - devel
 
 before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
-  deploy:
-    provider: script
-    script: Rscript -e 'pkgdown::deploy_site_github()'
-    skip_cleanup: true
+deploy:
+  provider: script
+  script: Rscript -e 'pkgdown::deploy_site_github()'
+  cleanup: false
+  skip_cleanup: true
 
 after_success:
   - Rscript -e 'covr::codecov()'
