### 2020-02-11 770479d
Un-gitignore media files for docs
