+++ data-raw/make_data.R
@@ -94,4 +94,8 @@ usethis::use_data(fq_attachments, overwrite = T)
 
 # Update header of vignettes/odata-api.Rmd with (delete abs path):
 # - attachments/media/1568786958640.jpg
-ymlthis::yml_resource_files(ymlthis::yml(), fs::dir_ls(here::here("vignettes/attachments/media/")))
+ymlthis::yml_resource_files(
+  ymlthis::yml(),
+  fs::dir_ls(here::here("vignettes/attachments/media/"))
+)
+
