### 2020-02-11 def9395
Move vignette media to vignettes/media
