### 2020-02-11 bf3f052
Update vignette odata-api header resources
