### 2020-02-11 268ee41
Fix #51: add global verbosity setting
* ru_setup: add getter, setter, read env var RU_VERBOSE
* rd: update template param-verbose
* readme, vignettes: update
* rebuild man
