+++ data-raw/make_data.R
@@ -2,6 +2,7 @@
 # ODK Central example data
 #
 library(magrittr)
+library(ruODK)
 ruODK::ru_setup(
   svc = Sys.getenv("ODKC_TEST_SVC"),
   un = Sys.getenv("ODKC_TEST_UN"),
@@ -35,30 +36,31 @@ fq_form_schema <- ruODK::form_schema(parse = TRUE)
 fq_form_detail <- ruODK::form_detail()
 
 t <- fs::dir_create("attachments")
+tz <- "Australia/Perth"
 
 fid <- ruODK::get_test_fid()
 fid_csv <- fs::path(t, glue::glue("{fid}.csv"))
 fid_csv_tae <- fs::path(t, glue::glue("{fid}-taxon_encounter.csv"))
 fid_csv_veg <- fs::path(t, glue::glue("{fid}-vegetation_stratum.csv"))
 
-se <- ruODK::submission_export(local_dir = t, overwrite = FALSE, verbose = TRUE)
+se <- ruODK::submission_export(local_dir = ".", overwrite = FALSE, verbose = TRUE)
 f <- unzip(se, exdir = t)
 fq_zip_data <- fid_csv %>%
   readr::read_csv(na = c("", "NA", "na")) %>% # form uses "na" for NA
   janitor::clean_names(.) %>%
   attachment_link(.) %>%
-  ru_datetime(tz = "Australia/Perth") # an example timezone
+  ru_datetime(tz = tz) # an example timezone
 fq_zip_strata <- fid_csv_veg %>%
   readr::read_csv(na = c("", "NA", "na")) %>%
   janitor::clean_names(.) %>%
   attachment_link(.) %>%
-  ru_datetime(tz = "Australia/Perth") %>%
+  ru_datetime(tz = tz) %>%
   dplyr::left_join(fq_zip_data, by = c("parent_key" = "meta_instance_id"))
 fq_zip_taxa <- fid_csv_tae %>%
   readr::read_csv(na = c("", "NA", "na")) %>%
   janitor::clean_names(.) %>%
   attachment_link(.) %>%
-  ru_datetime(tz = "Australia/Perth") %>%
+  ru_datetime(tz = tz) %>%
   dplyr::left_join(fq_zip_data, by = c("parent_key" = "meta_instance_id"))
 
 fq_submission_list <- ruODK::submission_list()
@@ -92,9 +94,27 @@ usethis::use_data(fq_submission_list, overwrite = T)
 usethis::use_data(fq_submissions, overwrite = T)
 usethis::use_data(fq_attachments, overwrite = T)
 
-# Update header of vignettes/odata-api.Rmd with (delete abs path):
-# - attachments/media/1568786958640.jpg
+# Update header of vignettes/odata-api.Rmd with:
+# - media/1568786958640.jpg
+fs::dir_ls(here::here("media/media"), glob="*.jpg") %>%
+fs::file_copy(here::here("media"), overwrite = TRUE)
 ymlthis::yml_resource_files(
   ymlthis::yml(),
-  fs::dir_ls(here::here("vignettes/attachments/media/"))
+  fs::dir_ls(fs::path("media"), glob="*.jpg")
 )
+
+# -----------------------------------------------------------------------------#
+# If attachments in vignettes change
+#
+# Vignettes are big
+# the repo is small
+# so what shall we do
+# let's mogrify all
+# fs::dir_ls(here::here("media"), glob="*.jpg") %>%
+  # fs::file_copy(here::here("vignettes/media/"), overwrite = TRUE)
+# system("find vignettes/media -type f -exec mogrify -resize 200x150 {} \\;")
+
+# Cleanup temp files
[…]
