+++ man-roxygen/param-verbose.R
@@ -1,1 +1,3 @@
-#' @param verbose Whether to display debug messages or not (default)
+#' @param verbose Whether to display debug messages or not.
+#'   Read `vignette("setup", package = "ruODK")` to learn how `ruODK`'s
+#'   verbosity can be set globally or per function.

[…]
