+++ README.md
@@ -173,7 +173,7 @@ descriptions of the steps below.
 ## Configure `ruODK`
 
 Set up `ruODK` with an OData Service URL and credentials of a
-read-permitted ODK Central web user.
+read-permitted ODK Central web user. Adjust verbosity to your liking.
 
     #> <ruODK settings>
     #>   Default ODK Central Project ID:  14 
@@ -188,13 +188,15 @@ read-permitted ODK Central web user.
     #>   Test ODK Central Form ID (Parsing tests): build_Turtle-Track-or-Nest-1-0_1569907666 
     #>   Test ODK Central URL: https://sandbox.central.opendatakit.org 
     #>   Test ODK Central Username: Florian.Mayer@dbca.wa.gov.au 
-    #>   Test ODK Central Password: run ruODK::get_test_pw() to show
+    #>   Test ODK Central Password: run ruODK::get_test_pw() to show 
+    #>   Verbose messages: TRUE
 
 ``` r
 ruODK::ru_setup(
   svc = "https://sandbox.central.opendatakit.org/v1/projects/14/forms/build_Flora-Quadrat-0-4_1564384341.svc",
   un = "me@email.com",
-  pw = "..."
+  pw = "...",
+  verbose = TRUE # great for demo or debugging
 )
 ```
 
@@ -255,6 +257,284 @@ frmd %>% knitr::kable(.)
 ``` r
 # Form schema
 meta <- ruODK::form_schema()
+#> 'data.frame':    10 obs. of  3 variables:
+#>  $ type: chr  "structure" "dateTime" "string" "string" ...
+#>  $ name: chr  "meta" "encounter_start_datetime" "reporter" "device_id" ...
+#>  $ path: chr  "Submissions" "Submissions" "Submissions" "Submissions" ...
+#> 
+#> ℹ
+#> 
+#> Found child: meta
+#> 'data.frame':    0 obs. of  3 variables:
+#>  $ type: logi 
+#>  $ name: logi 
+#>  $ path: 'glue' chr
+#> 
+#> 'data.frame':    0 obs. of  3 variables:
+#>  $ type: logi 
+#>  $ name: logi 
+#>  $ path: 'glue' chr
+#> 
+#> ℹ
+#> 
+#> Found child: list(name = "instanceID", type = "string")
+#> 'data.frame':    1 obs. of  3 variables:
+#>  $ type: chr "string"
+#>  $ name: chr "instanceID"
+#>  $ path: 'glue' chr "Submissions.meta"
+#> 
+#> 'data.frame':    1 obs. of  3 variables:
+#>  $ type: chr "string"
+#>  $ name: chr "instanceID"
+#>  $ path: 'glue' chr "Submissions.meta"
+#> 
+#> ℹ
+#> 
+#> Found child: structure
+#> 'data.frame':    0 obs. of  3 variables:
+#>  $ type: logi 
+#>  $ name: logi 
+#>  $ path: 'glue' chr
+#> 
+#> 'data.frame':    0 obs. of  3 variables:
+#>  $ type: logi 
+#>  $ name: logi 
+#>  $ path: 'glue' chr
+#> 
+#> ℹ
+#> 
+#> Found child: location
+#> 'data.frame':    0 obs. of  3 variables:
+#>  $ type: logi 
+#>  $ name: logi 
+#>  $ path: 'glue' chr
+#> 
+#> 'data.frame':    0 obs. of  3 variables:
+#>  $ type: logi 
+#>  $ name: logi 
+#>  $ path: 'glue' chr
+#> 
+#> ℹ
+#> 
+#> Found child: list(name = "area_name", type = "string")ℹ
+#> 
[…]
