### 2020-02-11 e1022b1
update NEWS
