### 2020-02-11 dbca-wa/ruODK#14 @florianm
Picking up the publication plans, currently I have about 2 months of urgent end of field season work before I can realistically dedicate time to attempt an MEE paper.
