### 2020-02-11 d6fd4af
Add tests for verbose setting
* Found and fixed bug: need to type cast logical setting
