### 2020-02-12 bb85681
Fix #53: add timezone default
* Add setters and getters
* Add news
* Functions now use `tz = get_default_tz()`
* Amend README, examples and vignettes
* Rebuild data
* Add man `@template param-tz`
