### 2020-02-13 c9e8870
#50 Toggle console output of ru_setup with get_ru_verbose
