### 2020-02-14 54da50d
Adjust test-ru_settings for tz and verbose
