### 2020-02-14 e03d64f
Let tz default to UTC. Still warn in getter if empty.
