### 2020-02-14 d23361b
test-ru_settings: make (more) independent of env vars
