+++ tests/testthat/test-ru_setup.R
@@ -29,18 +29,18 @@ test_that("ru_setup does not update settings if given NULL", {
   # If env vars not set, default getters will issue warnings
   testthat::expect_equal(x$url, get_default_url())
   testthat::expect_equal(xx$url, get_default_url())
-  testthat::expect_equal(x$un, get_default_un())
-  testthat::expect_equal(x$pw, get_default_pw())
-  testthat::expect_equal(x$tz, get_default_tz())
-  testthat::expect_equal(x$test_url, get_test_url())
-  testthat::expect_equal(x$test_un, get_test_un())
-  testthat::expect_equal(x$test_pw, get_test_pw())
-  testthat::expect_equal(x$test_pid, get_test_pid())
-  testthat::expect_equal(x$test_fid, get_test_fid())
-  testthat::expect_equal(x$test_fid_zip, get_test_fid_zip())
-  testthat::expect_equal(x$test_fid_att, get_test_fid_att())
-  testthat::expect_equal(x$test_fid_gap, get_test_fid_gap())
-  testthat::expect_equal(x$verbose, get_ru_verbose())
+  testthat::expect_equal(xx$un, get_default_un())
+  testthat::expect_equal(xx$pw, get_default_pw())
+  testthat::expect_equal(xx$tz, get_default_tz())
+  testthat::expect_equal(xx$test_url, get_test_url())
+  testthat::expect_equal(xx$test_un, get_test_un())
+  testthat::expect_equal(xx$test_pw, get_test_pw())
+  testthat::expect_equal(xx$test_pid, get_test_pid())
+  testthat::expect_equal(xx$test_fid, get_test_fid())
+  testthat::expect_equal(xx$test_fid_zip, get_test_fid_zip())
+  testthat::expect_equal(xx$test_fid_att, get_test_fid_att())
+  testthat::expect_equal(xx$test_fid_gap, get_test_fid_gap())
+  testthat::expect_equal(xx$verbose, get_ru_verbose())
 })
 
 test_that("ru_setup resets settings if given empty string", {
