### 2020-02-14 5fec31d
test-ru_settings: defuse tz tests which pass on travis but fail on appveyor
