### 2020-02-16 #335 @jmt2080ad
> We now use the form schema to detect date and datetime columns.

`ru_datetime` still parses fields by sub-string. My concern here is that running `ru_datetime` will try to parse columns that have matching sub-strings but are not date/time. I understand this is a convenience function and works for particular cases. I'll leave it up to you to decide if the convenience outweighs the risk of inadvertently converting fields like in the example below.

```

df <- data.frame(
  dbh_centimeter = runif(10) * 100,
  timedate       = format(Sys.time(), "%H:%M:%S %Y/%m/%d"),
  datetime       = format(Sys.time(), "%Y/%m/%d %H:%M:%S"),
  posixtime      = as.POSIXct(Sys.time(), tz = "UTC")
)

df %>% ru_datetime(tz = Sys.timezone())

```

In your vignette, you use this function on data read in using `readr::read_csv` which automatically detects and sets the columns to "POSIXct POSIXt" where appropriate. These columns are also automatically assigned to the UTC timezone, which this function then converts to "Australia/Perth". Does this package always pull time data from the ODK servers in UTC? If that cannot be guaranteed, then this could be a problem. If the server is storing time data in a local timezone, `read_csv` is assigning that to UTC when reading in data to R, if you convert to local time using ru_datetime, wouldn't that transpose the time incorrectly?

Because this function calls `isodt_to_local`, it appears that this function has two purposes: to convert between UTC and local, and to set the date type where a specific sub-string (default "time") is present in the column name.

### For bulk timezone conversions:
Keeping in mind my comment above, if you are going to do bulk timezone conversion, I think it would be safer to use type introspection instead of searching by some sub-string, e.g. where there is a "POSIXct POSIXt" type, convert the zone set in the `tz` argument of the `ru_datetime` call.

### For finding date/time columns:
For bulk date/time type assignment, I think you need to make a handler for columns that are not converted correctly. In my example above, the `dbh_centimeter` and `timedate` columns are set to all NA due to conversion failure. A handler in `isodt_to_local` that checks for this then returns the original data would prevent errors trickling down a given processing pipeline that uses these functions.

I think these are my only comments. Everything else looks good.
