### 2020-02-17 #335 @florianm
Thanks for the detailed feedback @jmt2080ad!

## ru_datetime field names
The "golden path" of accessing data from ODK Central uses the OData API.
`odata_submission_get()` with default param `parse = TRUE` does the right thing here- it [parses the `form_schema` for date and datetime fields](https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L247). Here we won't need to touch `ru_datetime()` unless going down the manual ["DIY parse"](https://dbca-wa.github.io/ruODK/articles/odata-api.html#detour-data-rectangling) path.

Accessing data through the RESTful API on the other hand (both bulk export and individual access) involves more manual parsing rectangling, where indeed we call `ru_datetime()` directly.

* Timezones have a very recent default (env var RU_TIMEZONE). See the tutorial in vignette [setup](https://dbca-wa.github.io/ruODK/articles/setup.html#best-practice-ru_setup) and the reference docs for [settings](https://dbca-wa.github.io/ruODK/reference/ru_setup.html) (argh, formatting bug with what should be bullet points under `tz`). The abundant "Australia/Perth" is my Perth-flavoured go-to example timezone. The default tz (if unspecified) is UTC (as is the default export format from ODK Central, as opposed to Sys.timezone()).
* Bulk exports as zipped CSVs always export to UTC. The ODK Central API [docs](https://odkcentral.docs.apiary.io/#reference/forms-and-submissions/submissions/exporting-form-submissions-to-csv) don't specify this explicitly, but doesn't specify anything to the contrary either.
* `ru_datetime()` guarantees 1. type (POSIXct or POSIXt) and 2. timezone for (valid) date/datetime fields. I assumed that any mishap (parsing "cenTIMEter" as datetime) would be picked up sooner or later by the user.

@jmt2080ad how would you feel about a refactor of `ru_datetime`:

* apply `ru_datetime()` to all date/time columns of a submission table based on the `form_schema`
* test for correct type after conversion, warn if not time/datetime
* test for all NA after conversion, warn if all NA
* Re-write vignette RESTful API > ru_datetime
* Do the same for [geopoints](https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L337)
* Keep default tz at UTC instead of `Sys.timezone()` - think of RStudio Server hosted in other timezones (e.g. we're in Perth, our data centers are in Sydney) or analyses run by a multi-national team.
