### 2020-02-17 dbca-wa/ruODK#54 @florianm
Make ru_datetime more robust
## Feature
Source @jmt2080ad https://github.com/ropensci/software-review/issues/335#issuecomment-586763042

* apply `ru_datetime()` to all date/time columns of a submission table based on the `form_schema`
* test for correct type after conversion, warn if not time/datetime [here](https://github.com/dbca-wa/ruODK/blob/master/R/ru_datetime.R#L45)
* test for all NA after conversion, warn if all NA [here](https://github.com/dbca-wa/ruODK/blob/master/R/ru_datetime.R#L45)
* Re-write vignette RESTful API > application of ru_datetime
* Do the same for [geopoints](https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L337)?
* Keep default tz at UTC instead of `Sys.timezone()` - think of RStudio Server hosted in other timezones (e.g. we're in Perth, our data centers are in Sydney) or analyses run by a multi-national team.
