### 2020-02-17 b2e106f
Fix formatting bug in ru_setup docs
