### 2020-02-17 81a1156
Typo fix in ru_setup and ru_datetime man
