### 2020-02-17 f57d73c
Update Rmd template skeleton to defaults for tz and verbose
