+++ inst/rmarkdown/templates/odata/skeleton/skeleton.Rmd
@@ -72,20 +72,22 @@ library(readr)
 library(skimr)
 library(ruODK)
 
-# Set ruODK defaults to an ODK Central form
+# Set ruODK defaults to an ODK Central form, choose own tz and verbosity
 ruODK::ru_setup(
   svc=paste0("https://sandbox.central.opendatakit.org/v1/projects",
-             "/14/forms/build_Flora-Quadrat-0-4_1564384341.svc"))
-tz <- "Australia/Perth"
-loc <- fs::path("attachments", "media")
+             "/14/forms/build_Flora-Quadrat-0-4_1564384341.svc"),
+  tz = "Australia/Perth",
+  verbose = TRUE)
+
+loc <- fs::path("media")
 fs::dir_create(loc)
 ```
 
 # Download data
 <!-- Get table names. `form_tables` will contain one row per separate table. -->
 ```{r svc_get}
-form_tables <- ruODK::odata_service_get()
-form_tables %>% knitr::kable(.)
+ft <- ruODK::odata_service_get()
+ft %>% knitr::kable(.)
 ```
 
 <!-- 
@@ -102,34 +104,33 @@ Repeat for each table name in form_tables$url.
 -->
 ```{r dl_submissions}
 data <- ruODK::odata_submission_get(
-  table = form_tables$url[1], verbose = TRUE, tz = tz, local_dir = loc
-) %>%
-  dplyr::rename(
-    # Adjust coordinate colnames as needed
-    # longitude = x13,
-    # latitude = x14,
-    # altitude = x15
-  )
+  table = ft$url[1], 
+  local_dir = loc, 
+  wkt = TRUE
+)
+
 data_sub1 <- ruODK::odata_submission_get(
-  table = form_tables$url[2], verbose = TRUE, tz = tz, local_dir = loc
+  table = ft$url[2], 
+  local_dir = loc
 ) %>%
-  dplyr::left_join(data, by = c("submissions_id" = "id"))
+  dplyr::left_join(
+    data, 
+    by = c("submissions_id" = "id")
+  )
 
 data_sub2 <- ruODK::odata_submission_get(
-  table = form_tables$url[3], verbose = TRUE, tz = tz, local_dir = loc
-) %>% 
-  dplyr::rename(
-    # Adjust coordinate colnames as needed, avoid name conflicts with main data
-    # lon = x6,
-    # lat = x7,
-    # alt = x8
-  ) %>%
-  dplyr::left_join(data, by = c("submissions_id" = "id"))
+  table = ft$url[3], 
+  local_dir = loc
+) %>%
+  dplyr::left_join(
+    data, 
+    by = c("submissions_id" = "id")
+)
 
 # repeat for any remaining nested tables
 ```
 
-# Analyse data
+# Visualise data
 <!-- Summarise, analyse, visualise the tibble(s) data (, data_sub1, data_sub2) -->
 ```{r data_vis}
 skimr::skim(data)
@@ -144,19 +145,23 @@ leaflet::leaflet(width = 800, height = 600) %>%
   leaflet::clearBounds() %>%
   leaflet::addAwesomeMarkers(
     data = data,
-    lng = ~longitude, lat = ~latitude,
+    # 
+    # # Adjust to your coordinate field names
+    # 
+    lng = ~location_corner1_longitude, 
+    lat = ~location_corner1_latitude, 
     icon = leaflet::makeAwesomeIcon(text = "Q", markerColor = "red"),
     # 
     # # With your own field names
     # 
-    # label = ~ glue::glue("{area_name} {encounter_start_datetime}"),
+    # label = ~ glue::glue("{location_area_name} {encounter_start_datetime}"),
     # popup = ~ glue::glue(
-    #   "<h3>{area_name}</h3>",
[…]
