### 2020-02-17 8a70c64
Clean up Rmd template
