### 2020-02-17 2939a4a
Fix formatting bug in ruODK-package docs
