+++ R/ruODK.R
@@ -1,9 +1,9 @@
 #' @description \code{\link{ruODK}} is an R Client for the ODK Central API.
 #'
 #'   Please see the `ruODK` website for full documentation:
-#'   * <https://dbca-wa.github.io/ruODK/>
+#'   <https://dbca-wa.github.io/ruODK/>
 #'
-#'   `ruODK` is "pipe-friendly" and re-exports `%>%` and `%||%`, but does not
+#'   `ruODK` is "pipe-friendly" and re-exports `\%>\%` and `\%||\%`, but does not
 #'   require their use.
 #'
 #' @importFrom rlang %||%
