### 2020-02-20 dbca-wa/ruODK#54 @florianm
Idea:
```
handle_ru_datetimes() <- https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L247-L278

handle_ru_attachments <- https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L280-L320

handle_ru_geopoints <- https://github.com/dbca-wa/ruODK/blob/master/R/odata_submission_get.R#L322-L351

rename odata_submission_parse() to odata_submission_rectangle()

odata_submission_parse <- function(data, form_schema, some opts){
  odata_submission_rectangle()
  handle_ru_datetimes()
  handle_ru_attachments()
  handle_ru_geopoints()
}
```
