### 2020-02-21 9010b28
Refactor parsing of submissions and messages.
* Rename `odata_submission_parse()` to `odata_submission_rectangle()`
* Factor out `ru_handle_{datetimes, attachments, geopoints}` (plus man, examples, tests)
* Update package data and `data-raw/make_data.R`
* Update vignettes
* Fixes #54

* Factor out messages into `ru_msg_{info, success, noop}`, reducing duplication
* Clarify warning message for `ru_default_tz()`
