+++ R/odata_submission_get.R
@@ -30,10 +30,11 @@
 #'
 #' With parameter `parse=FALSE`, submission data is presented as nested list,
 #' which is the R equivalent of the returned form JSON.
-#' From there, \code{\link{odata_submission_parse}} will parse the data into a
-#' tibble, and subsequent lines of \code{\link{ru_datetime}} and
-#' \code{\link{attachment_get}} parse dates and download and link file
-#' attachments.
+#' From there, \code{\link{odata_submission_rectangle}} will rectangle the data
+#' into a tibble, and subsequent lines of \code{\link{handle_ru_datetimes}},
+#' \code{\link{handle_ru_attachments}} and \code{\link{handle_ru_geopoints}}
+#' parse dates, download and link file attachments, and split geopoints into
+#' coordinates.
 #' As any of these steps might fail on unexpected errors, `ruODK` offers this
 #' longer, more manual pathway as an option to investigate and narrow down
 #' unexpected or unwanted behaviour.
@@ -163,26 +164,16 @@ odata_submission_get <- function(table = "Submissions",
                                  un = get_default_un(),
                                  pw = get_default_pw(),
                                  tz = get_default_tz(),
-                                 verbose = get_ru_verbose()
-                                 ) {
-  type <- NULL
+                                 verbose = get_ru_verbose()) {
   yell_if_missing(url, un, pw)
-
-  # Parse params
   skip <- skip %||% ""
   top <- top %||% ""
   count <- ifelse(count == FALSE, "false", "true")
   wkt <- ifelse(wkt == FALSE, "false", "true")
 
-  # Get submissions
-  if (verbose == TRUE) {
-    message(crayon::cyan(
-      glue::glue(
-        "{clisymbols::symbol$info}",
-        " Downloading submissions...\n"
-      )
-    ))
-  }
+  #----------------------------------------------------------------------------#
+  # Download submissions
+  if (verbose == TRUE) ru_msg_info("Downloading submissions...")
 
   sub <- httr::RETRY(
     "GET",
@@ -196,35 +187,18 @@ odata_submission_get <- function(table = "Submissions",
     yell_if_error(., url, un, pw) %>%
     httr::content(.)
 
-  if (verbose == TRUE) {
-    message(crayon::green(
-      glue::glue(
-        "{clisymbols::symbol$tick}",
-        " Downloaded submissions.\n"
-      )
-    ))
-  }
+  if (verbose == TRUE) ru_msg_success("Downloaded submissions.")
 
   if (parse == FALSE) {
+    if (verbose == TRUE) ru_msg_success("Returning unparsed submissions.")
     return(sub)
   } else {
-    if (verbose == TRUE) {
-      message(crayon::cyan(
-        glue::glue(
-          "{clisymbols::symbol$info}",
-          " Parsing submissions...\n"
-        )
-      ))
-    }
+    if (verbose == TRUE) ru_msg_info("Parsing submissions...")
   }
 
   #----------------------------------------------------------------------------#
   # Get form schema
-  if (verbose == TRUE) {
-    message(crayon::cyan(
-      glue::glue("{clisymbols::symbol$info} Reading form schema...\n")
-    ))
-  }
+  if (verbose == TRUE) ru_msg_info("Reading form schema...")
 
   fs <- form_schema(
     parse = TRUE,
@@ -237,130 +211,42 @@ odata_submission_get <- function(table = "Submissions",
 
   #----------------------------------------------------------------------------#
   # Parse submission data
-  if (verbose == TRUE) {
-    message(crayon::cyan(
[…]
