### 2020-02-21 434ccaa
Update test-ru_setup
