### 2020-02-21 903c323
test ru_setup: debug windows tests for tz
