+++ tests/testthat/test-ru_setup.R
@@ -129,8 +129,19 @@ test_that("ru_setup resets settings if given empty string", {
 })
 
 test_that("get_default_tz warns if tz explicitly set to empty string", {
+
+  tz <- get_default_tz()
+
+  # Override tz
   ru_setup(tz = "")
+
+  # Appveyor seems to be UTC
+  message("Default time zone after setting to empty string: ")
+  message(get_default_tz())
+
   testthat::expect_warning(get_default_tz())
+
+  ru_setup(tz = tz) # restore previous settings
 })
 
 test_that("ru_setup sets pid, fid, url if given service url", {
