### 2020-02-21 9e8ac4f
test-ru_setup: test assumptions about tz
