### 2020-02-21 c7a3dc4
test-ru_setup: add mufflers
