### 2020-02-21 0d0c97e
TravisCI to only deploy gh-pages on R release
