+++ .travis.yml
@@ -6,17 +6,15 @@ sudo: false
 r:
 - oldrel
 - release
+  before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
+  deploy:
+    provider: script
+    script: Rscript -e 'pkgdown::deploy_site_github()'
+    cleanup: false
+    skip_cleanup: true
 - devel
 
-before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
-
 cache: packages
-
-deploy:
-  provider: script
-  script: Rscript -e 'pkgdown::deploy_site_github()'
-  cleanup: false
-  skip_cleanup: true
 
 after_success:
   - Rscript -e 'covr::codecov()'
