### 2020-02-21 #335 @florianm
For `ru_datetime` I think we have a workable solution now - could I solicit your opinions?
See https://github.com/dbca-wa/ruODK/issues/54 for details - after a complete refactor of submission parsing, `ru_datetime` is only called internally, and superceded by new helpers `handle_ru_{datetimes, attachments, geopoints}` as shown in vignettes. These are driven by the form schema definition, so they will never mistake `cenTIMEmeter` for a time field.

If you suggest to go even stricter, we could now drop the default field name component `col_contains="date"` from `ru_datetime` without complicating everyday usage.
