### 2020-02-27 20dc261
Fix versions for clisymbols and crayon
