+++ DESCRIPTION
@@ -36,8 +36,8 @@ BugReports: https://github.com/dbca-wa/ruODK/issues
 Depends: 
     R (>= 3.4)
 Imports: 
-    clisymbols,
-    crayon,
+    clisymbols (>= 1.2.0),
+    crayon (>= 1.3.4),
     dplyr (>= 0.8.3),
     fs (>= 1.3.1),
     glue (>= 1.3.1),
