### 2020-03-07 #335 @jmt2080ad
Ok, this is looking really good. The handlers are using the form schema, which is great.

## Concerning the handler functions and the functions they handle

I think your idea of dropping the default argument for `col_contains` makes sense. Also, do you see any reason for `ru_datetime` or `isodt_to_local` to be exported now that they are only used internally?

I see that there are a few places in the documentation where `ru_datetime` is mentioned. In the details for `form_schema_parse`:

> form_schema_parse recursively unpacks the form and extracts the name and type of each field. This information then can be used to inform the user which columns require ru_datetime, attachment_get, or attachment_link, respectively

And I see that `ru_datatime` is used in the example for `attachment_link`.

If you decide not to export the functions that you are using only internally (totally your choice), then I think should edit the documentation so that it does not point users to them. If you do export them, and use cases are not demonstrated in the vignettes, then they need to be well documented. Just take an extra moment to make sure that you are saying all you want to in their help.

These same points, exporting internal functions and mentioning internal functions in the documentation, should be considered for the other functions you wrote handlers for: `split_geopoint` and `attachment_get`.

I also noticed that `attachment_link` appears to be a place holder for functionality you intend to bring up in the future. It isn't demonstrated in the vignettes, and it has a TODO note in it that indicates that you will be improving this function to use the form schema to find attachment fields, like you did for `ru_datatime`. I am not sure the best approach here, but I think it's worth considering putting this in a development branch for now if it is not fully implemented. I might be wrong about your intentions here, so pardon me if that is the case. Would it make sense to do what this function does when a user calls `handle_ru_attachments`?

## Other notes:

Here are some outdated function references I found in the vignettes:

* Line 171 of odata-api.Rmd references the superseded `odata_submission_parse`.
* Lines 193 and 197 of odata-api.Rmd references the superseded `parse_datetime`.
* The section "Spatial formats - how granular..." that starts on line 319 of odata-api.Rmd has a few references to `parse_submissions`, which appears to have been superseded. I think this section needs to be updated to reflect the modifications to the workflow you have made.
* Line 337 of restful-api.Rmd references the superseded `odata_submission_parse`.
