### 2020-03-11 0dcc958
Refactor messages, bump rlang version
* Refactor colourful messages with custom icons into helper functions
* Bump rlang version to 0.4.5 to resolve downstream dependency conflict
* Update other dependencies
