+++ R/ru_msg.R
@@ -43,3 +43,35 @@ ru_msg_noop <- function(message) {
   x <- clisymbols::symbol$circle_filled
   message(crayon::green(glue::glue("{x} {message}\n")))
 }
+
+
+#' rlang::warn() with a yellow warning message with a warning symbol.
+#'
+#' \lifecycle{stable}
+#'
+#' @param message <chr> A message to print
+#' @return NULL
+#' @export
+#' @family helpers
+#' @examples
+#' ru_msg_warn("This is a warning.")
+ru_msg_warn <- function(message) {
+  x <- clisymbols::symbol$warning
+  rlang::warn(crayon::yellow(glue::glue("{x} {message}\n")))
+}
+
+
+#' rlang::abort() with a red error message with a cross symbol.
+#'
+#' \lifecycle{stable}
+#'
+#' @param message <chr> A message to print
+#' @return NULL
+#' @export
+#' @family helpers
+#' @examples
+#' ru_msg_abort("This is an error, abort.")
+ru_msg_abort <- function(message) {
+  x <- clisymbols::symbol$cross
+  rlang::abort(crayon::red(glue::glue("{x} {message}\n")))
+}
