+++ R/ru_setup.R
@@ -267,9 +267,8 @@ ru_setup <- function(svc = NULL,
 #' @rdname ru_settings
 get_default_pid <- function() {
   x <- Sys.getenv("ODKC_PID")
-  if (identical(x, "")) {
-    rlang::warn("No default ODK Central Project ID set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No default ODK Central Project ID set. ru_setup()?")
   x
 }
 
@@ -278,9 +277,8 @@ get_default_pid <- function() {
 #' @rdname ru_settings
 get_default_fid <- function() {
   x <- Sys.getenv("ODKC_FID")
-  if (identical(x, "")) {
-    rlang::warn("No default ODK Central Form ID set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No default ODK Central Form ID set. ru_setup()?")
   x
 }
 
@@ -289,9 +287,8 @@ get_default_fid <- function() {
 #' @rdname ru_settings
 get_default_url <- function() {
   x <- Sys.getenv("ODKC_URL")
-  if (identical(x, "")) {
-    rlang::warn("No default ODK Central URL set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No default ODK Central URL set. ru_setup()?")
   x
 }
 
@@ -300,9 +297,8 @@ get_default_url <- function() {
 #' @rdname ru_settings
 get_default_un <- function() {
   x <- Sys.getenv("ODKC_UN")
-  if (identical(x, "")) {
-    rlang::warn("No default ODK Central username set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No default ODK Central username set. ru_setup()?")
   x
 }
 
@@ -311,9 +307,8 @@ get_default_un <- function() {
 #' @rdname ru_settings
 get_default_pw <- function() {
   x <- Sys.getenv("ODKC_PW")
-  if (identical(x, "")) {
-    rlang::warn("No default ODK Central password set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No default ODK Central password set. ru_setup()?")
   x
 }
 
@@ -323,7 +318,7 @@ get_default_pw <- function() {
 get_default_tz <- function() {
   x <- Sys.getenv("RU_TIMEZONE")
   if (identical(x, "")) {
-    rlang::warn("No default timezone set, defaulting to 'UTC'. ru_setup()?")
+    ru_msg_warn("No default timezone set, defaulting to 'UTC'. ru_setup()?")
     x <- "UTC"
   }
   x
@@ -334,9 +329,8 @@ get_default_tz <- function() {
 #' @rdname ru_settings
 get_test_url <- function() {
   x <- Sys.getenv("ODKC_TEST_URL")
-  if (identical(x, "")) {
-    rlang::warn("No test ODK Central URL set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No test ODK Central URL set. ru_setup()?")
   x
 }
 
@@ -345,9 +339,8 @@ get_test_url <- function() {
 #' @rdname ru_settings
 get_test_un <- function() {
   x <- Sys.getenv("ODKC_TEST_UN")
-  if (identical(x, "")) {
-    rlang::warn("No test ODK Central username set. ru_setup()?")
-  }
+  if (identical(x, ""))
+    ru_msg_warn("No test ODK Central username set. ru_setup()?")
   x
 }
 
@@ -356,9 +349,8 @@ get_test_un <- function() {
 #' @rdname ru_settings
[…]
