+++ man/ru_msg_noop.Rd
@@ -17,7 +17,9 @@ ru_msg_noop("This is a noop message.")
 }
 \seealso{
 Other helpers: 
+\code{\link{ru_msg_abort}()},
 \code{\link{ru_msg_info}()},
-\code{\link{ru_msg_success}()}
+\code{\link{ru_msg_success}()},
+\code{\link{ru_msg_warn}()}
 }
 \concept{helpers}
