+++ man/ru_msg_warn.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/ru_msg.R
-\name{ru_msg_success}
-\alias{ru_msg_success}
-\title{Print a green success message with a tick symbol.}
+\name{ru_msg_warn}
+\alias{ru_msg_warn}
+\title{rlang::warn() with a yellow warning message with a warning symbol.}
 \usage{
-ru_msg_success(message)
+ru_msg_warn(message)
 }
 \arguments{
 \item{message}{<chr> A message to print}
@@ -13,13 +13,13 @@ ru_msg_success(message)
 \lifecycle{stable}
 }
 \examples{
-ru_msg_success("This is a success message.")
+ru_msg_warn("This is a warning.")
 }
 \seealso{
 Other helpers: 
 \code{\link{ru_msg_abort}()},
 \code{\link{ru_msg_info}()},
 \code{\link{ru_msg_noop}()},
-\code{\link{ru_msg_warn}()}
+\code{\link{ru_msg_success}()}
 }
 \concept{helpers}
