+++ man/ru_msg_abort.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/ru_msg.R
-\name{ru_msg_success}
-\alias{ru_msg_success}
-\title{Print a green success message with a tick symbol.}
+\name{ru_msg_abort}
+\alias{ru_msg_abort}
+\title{rlang::abort() with a red error message with a cross symbol.}
 \usage{
-ru_msg_success(message)
+ru_msg_abort(message)
 }
 \arguments{
 \item{message}{<chr> A message to print}
@@ -13,13 +13,13 @@ ru_msg_success(message)
 \lifecycle{stable}
 }
 \examples{
-ru_msg_success("This is a success message.")
+ru_msg_abort("This is an error, abort.")
 }
 \seealso{
 Other helpers: 
-\code{\link{ru_msg_abort}()},
 \code{\link{ru_msg_info}()},
 \code{\link{ru_msg_noop}()},
+\code{\link{ru_msg_success}()},
 \code{\link{ru_msg_warn}()}
 }
 \concept{helpers}
