+++ NAMESPACE
@@ -52,9 +52,11 @@ export(quo)
 export(quo_name)
 export(quos)
 export(ru_datetime)
+export(ru_msg_abort)
 export(ru_msg_info)
 export(ru_msg_noop)
 export(ru_msg_success)
+export(ru_msg_warn)
 export(ru_settings)
 export(ru_setup)
 export(split_geopoint)
