+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: ruODK
 Title: An R Client for the ODK Central API
-Version: 0.6.6.9014
+Version: 0.6.6.9015
 Authors@R: 
     c(person(given = c("Florian", "W."),
              family = "Mayer",
@@ -43,11 +43,11 @@ Imports:
     glue (>= 1.3.1),
     httr (>= 1.4.1),
     janitor (>= 1.2.0),
-    lifecycle (>= 0.1.0.9000),
+    lifecycle (>= 0.2.0),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5),
     readr (>= 1.3.1),
-    rlang (>= 0.4.0.9000),
+    rlang (>= 0.4.5),
     rlist (>= 0.4.6.1),
     purrr (>= 0.3.2.9000),
     stringr (>= 1.4.0),
@@ -55,21 +55,19 @@ Imports:
     tidyr (>= 1.0.0.9000),
     xml2 (>= 1.2.2)
 Suggests: 
-    covr (>= 3.3.1),
+    covr (>= 3.4.0),
     DT (>= 0.9),
     ggplot2 (>= 3.2.1),
-    knitr (>= 1.25),
-    leaflet (>= 2.0.2),
-    listviewer (>= 2.1.0),
-    rmarkdown (>= 1.16),
-    testthat (>= 2.2.1),
+    knitr (>= 1.26),
+    leaflet (>= 2.0.3),
+    listviewer (>= 3.0.0),
+    rmarkdown (>= 1.17),
+    testthat (>= 2.3.0),
     usethis (>= 1.5.1.9000),
     vcr (>= 0.3.0)
 Remotes:
     github::tidyverse/purrr,
     github::tidyverse/tidyr,
-    github::r-lib/lifecycle,
-    github::r-lib/rlang,
     github::r-lib/usethis
 VignetteBuilder: 
     knitr
