### 2020-03-12 c4b8b91
Muffle ru_msg_warn/abort examples
