### 2020-03-12 affe017
Bump tidyr, purrr, usethis from dev to newer stable versions
* All tests are passing locally
* Intended fix for reverse dep CI failing on installing dev versions on esotheric OS like Windows
