+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: ruODK
 Title: An R Client for the ODK Central API
-Version: 0.6.6.9015
+Version: 0.6.6.9016
 Authors@R: 
     c(person(given = c("Florian", "W."),
              family = "Mayer",
@@ -49,10 +49,10 @@ Imports:
     readr (>= 1.3.1),
     rlang (>= 0.4.5),
     rlist (>= 0.4.6.1),
-    purrr (>= 0.3.2.9000),
+    purrr (>= 0.3.3),
     stringr (>= 1.4.0),
     tibble (>= 2.1.3),
-    tidyr (>= 1.0.0.9000),
+    tidyr (>= 1.0.1),
     xml2 (>= 1.2.2)
 Suggests: 
     covr (>= 3.4.0),
@@ -63,12 +63,8 @@ Suggests:
     listviewer (>= 3.0.0),
     rmarkdown (>= 1.17),
     testthat (>= 2.3.0),
-    usethis (>= 1.5.1.9000),
+    usethis (>= 1.5.1),
     vcr (>= 0.3.0)
-Remotes:
-    github::tidyverse/purrr,
-    github::tidyverse/tidyr,
-    github::r-lib/usethis
 VignetteBuilder: 
     knitr
 RdMacros: 
