### 2020-03-12 bcc4b3d
Relax lifecycle dep version to unbreak Appveyor (Windows) CI
