+++ DESCRIPTION
@@ -43,7 +43,7 @@ Imports:
     glue (>= 1.3.1),
     httr (>= 1.4.1),
     janitor (>= 1.2.0),
-    lifecycle (>= 0.2.0),
+    lifecycle (>= 0.1.0),
     lubridate (>= 1.7.4),
     magrittr (>= 1.5),
     readr (>= 1.3.1),
