### 2020-03-12 39f08db
Revert lifecycle to 0.2.0
