### 2020-03-12 1765c92
Relax versions of roxygen and xml2 to max avail on CRAN for Win R3.5
