### 2020-03-12 a033bcf
Relax versions of roxygen and xml2 to max avail on CRAN for Win R3.5
