+++ DESCRIPTION
@@ -53,7 +53,7 @@ Imports:
     stringr (>= 1.4.0),
     tibble (>= 2.1.3),
     tidyr (>= 1.0.1),
-    xml2 (>= 1.2.5)
+    xml2 (>= 1.2.2)
 Suggests: 
     covr (>= 3.4.0),
     DT (>= 0.9),
@@ -62,7 +62,7 @@ Suggests:
     leaflet (>= 2.0.3),
     listviewer (>= 3.0.0),
     rmarkdown (>= 1.17),
-    roxygen2 (>= 7.1.0),
+    roxygen2 (>= 7.0.2),
     testthat (>= 2.3.0),
     usethis (>= 1.5.1),
     vcr (>= 0.3.0)
