### 2020-03-12 fe1e370
Relax roxygen version to latest available Windows binaries; fix Rbuildignore
