+++ man/fq_raw_taxa.Rd
@@ -4,7 +4,9 @@
 \name{fq_raw_taxa}
 \alias{fq_raw_taxa}
 \title{OData submission data for a subgroup of an ODK Central form.}
-\format{A list of lists}
+\format{
+A list of lists
+}
 \source{
 See `system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`
 }
