### 2020-04-16 fd1e5ec
README: Update funding wording
