### 2020-04-16 f013d1f
Update references to renamed functions in vignettes
* Source https://github.com/ropensci/software-review/issues/335#issuecomment-596120460
* odata_submission_parse > odata_submission_rectangle
* link vignette odata-api from restful-api
* explain how to disable form group prefixing of field names in odata_submission_rectangle
* Fixes #58
