### 2020-04-16 dbca-wa/ruODK#56 @florianm
Do not export internal helper functions
## Feature
Source https://github.com/ropensci/software-review/issues/335#issuecomment-596120460

> I think your idea of dropping the default argument for col_contains makes sense. Also, do you see any reason for ru_datetime or isodt_to_local to be exported now that they are only used internally?
I see that there are a few places in the documentation where ru_datetime is mentioned. In the details for form_schema_parse:
"form_schema_parse recursively unpacks the form and extracts the name and type of each field. This information then can be used to inform the user which columns require ru_datetime, attachment_get, or attachment_link, respectively"
And I see that ru_datatime is used in the example for attachment_link.
If you decide not to export the functions that you are using only internally (totally your choice), then I think you should edit the documentation so that it does not point users to them. If you do export them, and use cases are not demonstrated in the vignettes, then they need to be well documented. Just take an extra moment to make sure that you are saying all you want to in their help.
These same points, exporting internal functions and mentioning internal functions in the documentation, should be considered for the other functions you wrote handlers for: split_geopoint and attachment_get.

* Internal helpers: ru_datetime, isodt_to_local, split_geopoint, attachment_get
* Extend documentation for internal helpers with "how to do it by hand" examples
* If not exporting, remove mentions in documentation?
* If exporting, explain clearly in documentation (maybe separate vignette "under the bonnet")?
