### 2020-04-16 dbca-wa/ruODK#58 @florianm
Fix stale references in documentation
Source https://github.com/ropensci/software-review/issues/335#issuecomment-596120460

> Line 171 of odata-api.Rmd references the superseded odata_submission_parse.
Lines 193 and 197 of odata-api.Rmd references the superseded parse_datetime.
The section "Spatial formats - how granular..." that starts on line 319 of odata-api.Rmd has a few references to parse_submissions, which appears to have been superseded. I think this section needs to be updated to reflect the modifications to the workflow you have made.
Line 337 of restful-api.Rmd references the superseded odata_submission_parse.
