### 2020-04-23 4d9949d
README: adjust to ODK re-branding, remove beloved intro line
