+++ README.Rmd
@@ -10,7 +10,10 @@ knitr::opts_chunk$set(
   fig.path = "man/figures/",
   out.width = "100%")
 ```
-# `ruODK`: An R Client for the ODK Central API <img src="man/figures/ruODK.png" align="right" alt="Are you ODK?" width="120" />
+# `ruODK`: An R Client for the ODK Central API <img 
+src="man/figures/ruODK.png" 
+align="right" 
+alt="Especially in these trying times, it is important to ask: ruODK? width="120" />
 <!-- badges: start -->
 [![ROpenSci submission status](https://badges.ropensci.org/335_status.svg)](https://github.com/ropensci/software-review/issues/335)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -21,34 +24,36 @@ knitr::opts_chunk$set(
 [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 <!-- badges: end -->
-   
-Especially in these trying times, it is important to ask: "ruODK?"
 
-[OpenDataKit](https://opendatakit.org/) (ODK) is a free and open-source software 
-for collecting, managing, and using data in resource-constrained environments.
+`ruODK` is an R client to access and parse data from ODK Central.
 
-ODK consists of a range of [software packages and apps](https://opendatakit.org/software/).
+[OpenDataKit](https://getodk.org/) (ODK) is 
+[free-and open-source software](https://getodk.org/software/) 
+that helps millions of people collect data quickly, accurately, offline, 
+and at scale. The software is in active use in every country in the world and is 
+supported by a large and helpful community.
+
 `ruODK` assumes some familiarity of its users with the ODK ecosystem and workflows.
-For a detailed overview, read the extensive [ODK documentation](https://docs.opendatakit.org/)
-and visit the friendly [ODK forum](https://forum.opendatakit.org/).
+For a detailed overview, read the extensive [ODK documentation](https://docs.getodk.org/)
+and visit the friendly [ODK forum](https://forum.getodk.org/).
 
-[ODK Central](https://docs.opendatakit.org/central-intro/) is a cloud-based
+[ODK Central](https://docs.getodk.org/central-intro/) is a cloud-based
 data clearinghouse for digitally captured data, replacing the older software 
-[ODK Aggregate](https://docs.opendatakit.org/aggregate-intro/). 
+[ODK Aggregate](https://docs.getodk.org/aggregate-intro/). 
 ODK Central manages user accounts and permissions, stores form definitions, 
 and allows data collection clients like 
-[ODK Collect](https://docs.opendatakit.org/collect-intro/) to connect to it for 
+[ODK Collect](https://docs.getodk.org/collect-intro/) to connect to it for 
 form download and submission upload.
 
 ![An ODK setup with ODK Build, Central, Collect, and ruODK](
 https://www.lucidchart.com/publicSegments/view/952c1350-3003-48c1-a2c8-94bad74cdb46/image.png)
 
-A typical [ODK workflow](https://docs.opendatakit.org/#how-is-odk-used):
-An XForm is designed e.g. in [ODK Build](https://build.opendatakit.org/),
-[published to ODK Central](https://docs.opendatakit.org/central-forms/), 
[…]
