### 2020-04-26 514d77f
Fix form_schema for ODK Central v0.7 and 0.8
