+++ tests/testthat/test-form_schema_parse.R
@@ -15,23 +15,24 @@ test_that("predict_ruodk_name works", {
   )
 })
 
-test_that("form_schema_parse works", {
+test_that("form_schema works with ODK Central v0.8", {
   fs <- form_schema(
     flatten = FALSE,
     parse = FALSE,
+    odata = FALSE,
     pid = get_test_pid(),
     fid = get_test_fid(),
     url = get_test_url(),
     un = get_test_un(),
-    pw = get_test_pw()
+    pw = get_test_pw(),
+    odkc_version = 0.8,
+    verbose = TRUE
   )
 
-  fsp <- form_schema_parse(fs)
-
-  testthat::expect_equal(class(fsp), c("tbl_df", "tbl", "data.frame"))
-  testthat::expect_equal(names(fsp), c("type", "name", "path"))
-  testthat::expect_true("encounter_start_datetime" %in% fsp$name)
-  testthat::expect_true("quadrat_photo" %in% fsp$name)
+  testthat::expect_equal(class(fs), "data.frame")
+  testthat::expect_equal(names(fs), c("path", "name", "type", "ruodk_name"))
+  testthat::expect_true("encounter_start_datetime" %in% fs$name)
+  testthat::expect_true("quadrat_photo" %in% fs$name)
 
   # # Attachments
   # fsp %>% dplyr::filter(type=="binary")
@@ -43,24 +44,26 @@ test_that("form_schema_parse works", {
   # fsp %>% dplyr::filter(type=="geopoint")
 })
 
-test_that("form_schema_parse debug messages work", {
-  fs <- form_schema(
-    flatten = FALSE,
-    parse = FALSE,
-    pid = get_test_pid(),
-    fid = get_test_fid(),
-    url = get_test_url(),
-    un = get_test_un(),
-    pw = get_test_pw()
-  )
-
-  testthat::capture_output(
-    testthat::expect_message(
-      fsp <- form_schema_parse(fs, verbose = TRUE)
-    )
-  )
-
-  testthat::expect_equal(class(fsp), c("tbl_df", "tbl", "data.frame"))
-  testthat::expect_true("encounter_start_datetime" %in% fsp$name)
-  testthat::expect_true("quadrat_photo" %in% fsp$name)
-})
+# TODO run this test against ODK Central v0.7
+# test_that("form_schema_parse debug messages work", {
+#   fs <- form_schema(
+#     flatten = FALSE,
+#     parse = FALSE,
+#     pid = get_test_pid(),
+#     fid = get_test_fid(),
+#     url = get_test_url(),
+#     un = get_test_un(),
+#     pw = get_test_pw(),
+#     odkc_version = 0.8
+#   )
+#
+#   testthat::capture_output(
+#     testthat::expect_message(
+#       fsp <- form_schema_parse(fs, verbose = TRUE)
+#     )
+#   )
+#
+#   testthat::expect_equal(class(fsp), c("tbl_df", "tbl", "data.frame"))
+#   testthat::expect_true("encounter_start_datetime" %in% fsp$name)
+#   testthat::expect_true("quadrat_photo" %in% fsp$name)
+# })
