### 2020-04-27 9ac9ca3
Fix tests depending on form_schema
