+++ tests/testthat/test-form_schema_parse.R
@@ -29,8 +29,8 @@ test_that("form_schema works with ODK Central v0.8", {
     verbose = TRUE
   )
 
-  testthat::expect_equal(class(fs), "data.frame")
-  testthat::expect_equal(names(fs), c("path", "name", "type", "ruodk_name"))
+  testthat::expect_true(tibble::is_tibble(fs))
+  testthat::expect_equal(names(fs), c("name", "type", "path", "ruodk_name"))
   testthat::expect_true("encounter_start_datetime" %in% fs$name)
   testthat::expect_true("quadrat_photo" %in% fs$name)
 
