+++ man/fq_data.Rd
@@ -41,7 +41,6 @@ Other included:
 \code{\link{fq_data_taxa}},
 \code{\link{fq_form_detail}},
 \code{\link{fq_form_list}},
-\code{\link{fq_form_schema_raw}},
 \code{\link{fq_form_schema}},
 \code{\link{fq_form_xml}},
 \code{\link{fq_meta}},
