### 2020-04-28 726514c
Update README: drop introduction line.
