### 2020-04-28 dbca-wa/ruODK#56 @florianm
> I think your idea of dropping the default argument for col_contains makes sense.

The default argument is now superseded by form schema introspection, therefore happy to drop.

> do you see any reason for ru_datetime or isodt_to_local to be exported now that they are only used internally?

Removed export, turned into internal: simple parsing helpers: ru_datetime, isodt_to_local
Deleted: `ru_datetime` - `handle_ru_datetimes` now uses `isodt_to_local` directly
Still exported and well (enough?) documented in examples and vignettes: functions that are core to ruODK's parsing workflow: `split_geopoint`, `attachment_get`

> I see that there are a few places in the documentation where ru_datetime is mentioned. In the details for form_schema_parse: <...>

Less of an issues now: `form_schema_parse` is only used in legacy mode (ODK Central version < 0.8)

> And I see that ru_datatime is used in the example for attachment_link.
Not any more! `attachment_link` now uses `handle_ru_datetimes` in examples - thanks for the prompt
