### 2020-04-28 cba1022
Fix #56: make helpers internal
