+++ tests/testthat/test-ru_datetime.R
@@ -1,4 +1,0 @@
-
-# tests run as part of test_attachment_link
-
-# usethis::edit_file("R/ru_datetime.R")
