+++ R/project_list.R
@@ -37,8 +37,8 @@
 #'
 #' # columns are project metadata
 #' names(pl)
-#' #> "id" "name" "forms" "app_users" "created_at" "updated_at"
-#' #> "last_submission" "archived"
+#' # > "id" "name" "forms" "app_users" "created_at" "updated_at"
+#' # > "last_submission" "archived"
 #' }
 project_list <- function(url = get_default_url(),
                          un = get_default_un(),
