+++ R/isodt_to_local.R
@@ -1,6 +1,10 @@
 #' Parse an ISO8601 datetime string to a timezone aware datetime.
 #'
 #' \lifecycle{stable}
+#'
+#' This function is used internally by \code{ruODK} to parse ISO timestamps
+#' to timezone-aware local times.
+#'
 #' @param datetime_string (character) An ISO8601 datetime string as produced by
 #'   XForms exported from ODK Central.
 #' @param orders (vector of character) Orders of datetime elements for
@@ -9,7 +13,7 @@
 #' @template param-tz
 #' @return A lubridate PosixCT datetime in the given timezone.
 #' @family utilities
-#' @export
+#' @keywords internal
 isodt_to_local <- function(datetime_string,
                            orders = c("YmdHMS", "YmdHMSz"),
                            tz = get_default_tz()) {

[…]
