### 2020-04-28 dbca-wa/ruODK#57 @florianm
* `submission_export` exports a ZIP archive of all submissions as CSV, plus all attachment media.
* `attachment_link` prepends the attachment filenames in the CSV with the attachment media folder.

The TODO note refers to the intent to introspect the form_schema to find column names of attachment files. This is now implemented as follows:

* `attachment_link` introspects `form_schema` to get attachment field names
* `attachment_link` params: added `form_schema`, dropped `att_contains`

`attachment_link` usage is (still) demonstrated in examples and (updated) in vignette "restful-api" under "submission_export".
