+++ R/attachment_get.R
@@ -9,8 +9,10 @@
 #' @family utilities
 #' @keywords internal
 #' @examples
+#' \donttest{}
 #' strip_uuid("uuid:1234")
 #' strip_uuid("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
+#' }
 strip_uuid <- function(uuid) {
   uuid %>% stringr::str_replace_all(., pattern = "uuid:", replacement = "")
 }
@@ -26,8 +28,10 @@ strip_uuid <- function(uuid) {
 #' @family utilities
 #' @keywords internal
 #' @examples
+#' \donttest{
 #' prepend_uuid("1234")
 #' prepend_uuid("d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
+#' }
 prepend_uuid <- function(md5hash) {
   glue::glue("uuid:{md5hash}") %>% as.character(.)
 }
@@ -50,11 +54,13 @@ prepend_uuid <- function(md5hash) {
 #' @family utilities
 #' @keywords internal
 #' @examples
+#' \donttest{
 #' ruODK:::attachment_url("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0",
 #'   "filename.jpg",
 #'   pid = 1, fid = "form1",
 #'   url = "https://sandbox.central.getodk.org"
 #' )
+#' }
 attachment_url <- function(uuid,
                            fn,
                            pid = get_default_pid(),
