+++ man/attachment_url.Rd
@@ -45,11 +45,13 @@ This function is vectorised and accepts single values or vectors for uuid and
 fn.
 }
 \examples{
+\donttest{
 ruODK:::attachment_url("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0",
   "filename.jpg",
   pid = 1, fid = "form1",
   url = "https://sandbox.central.getodk.org"
 )
+}
 }
 \seealso{
 Other utilities: 
