+++ man/prepend_uuid.Rd
@@ -19,8 +19,10 @@ The string with a prepended "uuid:"
 This is the inverse of the helper function \code{\link{strip_uuid}}.
 }
 \examples{
+\donttest{
 prepend_uuid("1234")
 prepend_uuid("d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
+}
 }
 \seealso{
 Other utilities: 
