+++ man/predict_ruodk_name.Rd
@@ -21,6 +21,7 @@ Predict a field name after \code{tidyr::unnest_wider(names_sep="_")} prefixes
 the form path.
 }
 \examples{
+\donttest{
 testthat::expect_equal(
   predict_ruodk_name("bar", "Submissions.foo"), "foo_bar"
 )
@@ -30,6 +31,7 @@ testthat::expect_equal(
 testthat::expect_equal(
   predict_ruodk_name("bar", "Submissions.foo_fighters"), "foo_fighters_bar"
 )
+}
 }
 \seealso{
 Other utilities: 
