### 2020-04-28 5f7ebaf
#64 fix examples of predict_ruodk_name
