+++ R/form_schema_parse.R
@@ -94,15 +94,12 @@ form_schema_parse <- function(fs,
 #' @keywords internal
 #' @examples
 #' \donttest{
-#' testthat::expect_equal(
-#'   predict_ruodk_name("bar", "Submissions.foo"), "foo_bar"
-#' )
-#' testthat::expect_equal(
-#'   predict_ruodk_name("bar", "Submissions"), "bar"
-#' )
-#' testthat::expect_equal(
-#'   predict_ruodk_name("bar", "Submissions.foo_fighters"), "foo_fighters_bar"
-#' )
+#'   predict_ruodk_name("bar", "Submissions.foo")
+#'   # > "foo_bar"
+#'   predict_ruodk_name("bar", "Submissions")
+#'   # > "bar"
+#'   predict_ruodk_name("rockr", "Submissions.foo_fighters")
+#'   # > "foo_fighters_rock"
 #' }
 predict_ruodk_name <- function(name_str, path_str) {
   prefix <- path_str %>%
