### 2020-04-28 408ef84
Debug messages for failing tests
