+++ R/form_schema_parse.R
@@ -98,7 +98,7 @@ form_schema_parse <- function(fs,
 #'   # > "foo_bar"
 #'   predict_ruodk_name("bar", "Submissions")
 #'   # > "bar"
-#'   predict_ruodk_name("rockr", "Submissions.foo_fighters")
+#'   predict_ruodk_name("rock", "Submissions.foo_fighters")
 #'   # > "foo_fighters_rock"
 #' }
 predict_ruodk_name <- function(name_str, path_str) {
