### 2020-04-28 776dbc2
Add debug messages to test-odata_submission_get
These tests fail on CI (both Travis and Appveyor) but pass locally. Why?
