### 2020-04-28 9e21bd1
Better debug messages
