### 2020-04-28 cc9fb37
settings defaults for ODK Central version
