### 2020-04-29 1e08db8
#64 Try \dontrun{} for @examples of internal functions
