+++ R/attachment_get.R
@@ -9,7 +9,7 @@
 #' @family utilities
 #' @keywords internal
 #' @examples
-#' \donttest{}
+#' \dontrun{
 #' strip_uuid("uuid:1234")
 #' strip_uuid("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
 #' }
@@ -28,7 +28,7 @@ strip_uuid <- function(uuid) {
 #' @family utilities
 #' @keywords internal
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' prepend_uuid("1234")
 #' prepend_uuid("d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
 #' }
@@ -54,7 +54,7 @@ prepend_uuid <- function(md5hash) {
 #' @family utilities
 #' @keywords internal
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' ruODK:::attachment_url("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0",
 #'   "filename.jpg",
 #'   pid = 1, fid = "form1",
