+++ man/attachment_url.Rd
@@ -45,7 +45,7 @@ This function is vectorised and accepts single values or vectors for uuid and
 fn.
 }
 \examples{
-\donttest{
+\dontrun{
 ruODK:::attachment_url("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0",
   "filename.jpg",
   pid = 1, fid = "form1",
