+++ R/utils-tidy-eval.R
@@ -11,6 +11,9 @@
 #' @family utilities
 #' @aliases          quo quos enquo sym syms ensym expr exprs enexpr quo_name
 #' @importFrom rlang quo quos enquo sym syms ensym expr exprs enexpr quo_name
-#' @export           quo quos enquo sym syms ensym expr exprs enexpr quo_name
+#' @importFrom tidyselect all_of any_of one_of contains ends_with starts_with
+#'   everything
+#' @export  quo quos enquo sym syms ensym expr exprs enexpr quo_name all_of
+#'   any_of one_of contains ends_with starts_with everything
 #' @importFrom rlang UQ UQS .data :=
 NULL
