+++ DESCRIPTION
@@ -8,11 +8,11 @@ Authors@R:
              role = c("aut", "cre"),
              email = "Florian.Mayer@dbca.wa.gov.au",
              comment = c(ORCID = "0000-0003-4269-4242")),
-      person(given = "Maëlle", 
-             family = "Salmon", 
-             email = "maelle.salmon@yahoo.se", 
-             role = "rev", 
-             comment = c(ORCID = "0000-0002-2815-0399")), 
+      person(given = "Maëlle",
+             family = "Salmon",
+             role = "rev",
+             email = "maelle.salmon@yahoo.se",
+             comment = c(ORCID = "0000-0002-2815-0399")),
       person(given = "Karissa",
              family = "Whiting",
              role = "rev",
@@ -38,22 +38,23 @@ Depends:
 Imports: 
     clisymbols (>= 1.2.0),
     crayon (>= 1.3.4),
-    dplyr (>= 0.8.3),
-    fs (>= 1.3.1),
-    glue (>= 1.3.1),
+    dplyr (>= 0.8.5),
+    fs (>= 1.4.1),
+    glue (>= 1.4.0),
     httr (>= 1.4.1),
-    janitor (>= 1.2.0),
+    janitor (>= 2.0.1),
     lifecycle (>= 0.1.0),
-    lubridate (>= 1.7.4),
+    lubridate (>= 1.7.8),
     magrittr (>= 1.5),
+    purrr (>= 0.3.4),
     readr (>= 1.3.1),
     rlang (>= 0.4.5),
     rlist (>= 0.4.6.1),
-    purrr (>= 0.3.3),
     stringr (>= 1.4.0),
     tibble (>= 2.1.3),
-    tidyr (>= 1.0.1),
-    xml2 (>= 1.2.2)
+    tidyr (>= 1.0.2),
+    tidyselect (>= 1.0.0),
+    xml2 (>= 1.3.2)
 Suggests: 
     covr (>= 3.4.0),
     DT (>= 0.9),
@@ -62,10 +63,10 @@ Suggests:
     leaflet (>= 2.0.3),
     listviewer (>= 3.0.0),
     rmarkdown (>= 1.17),
-    roxygen2 (>= 7.0.2),
-    testthat (>= 2.3.0),
-    usethis (>= 1.5.1),
-    vcr (>= 0.3.0)
+    roxygen2 (>= 7.1.0),
+    testthat (>= 2.3.2),
+    usethis (>= 1.6.0),
+    vcr (>= 0.5.4)
 VignetteBuilder: 
     knitr
 RdMacros: 
