### 2020-04-29 3f316f2
#65 downgrade xml2 dep to un-break Appveyor (windows build)
