+++ DESCRIPTION
@@ -54,7 +54,7 @@ Imports:
     tibble (>= 2.1.3),
     tidyr (>= 1.0.2),
     tidyselect (>= 1.0.0),
-    xml2 (>= 1.3.2)
+    xml2 (>= 1.2.2)
 Suggests: 
     covr (>= 3.4.0),
     DT (>= 0.9),
