### 2020-04-29 0d7103f
Travis: update config
