### 2020-04-29 b91f1b1
fix test-strip_uuid: drop `ruODK::`
