### 2020-04-29 bb4d5a8
#65 skip failing tests after printing debug logs
