+++ tests/testthat/test-odata_submission_get.R
@@ -104,7 +104,9 @@ test_that("odata_submission_get skip omits number of results", {
   print(skip_parsed)
 
 
-  testthat::skip() # https://github.com/dbca-wa/ruODK/issues/65
+  # https://github.com/dbca-wa/ruODK/issues/65
+  skip_on_travis()
+  skip_on_appveyor()
   testthat::expect_true(nrow(fresh_parsed) == nrow(skip_parsed) + 1)
 })
 
@@ -125,8 +127,10 @@ test_that("odata_submission_get top limits number of results", {
   print("odata_submission_get with top=1 should return one record")
   print(top_parsed)
 
-  testthat::skip() # https://github.com/dbca-wa/ruODK/issues/65
-  # testthat::expect_true(nrow(top_parsed) == 1)
+  # https://github.com/dbca-wa/ruODK/issues/65
+  skip_on_travis()
+  skip_on_appveyor()
+  testthat::expect_true(nrow(top_parsed) == 1)
 
 })
 
@@ -151,20 +155,20 @@ test_that("odata_submission_get count returns total number or rows", {
   print(x_raw)
   print(x_parsed)
 
-  testthat::skip() # https://github.com/dbca-wa/ruODK/issues/65
+  # https://github.com/dbca-wa/ruODK/issues/65
+  skip_on_travis()
+  skip_on_appveyor()
 
   # Returned: one row
-  # testthat::expect_true(nrow(x_parsed) == 1)
+  testthat::expect_true(nrow(x_parsed) == 1)
 
   # Count: shows all records
-  # testthat::expect_true("@odata.count" %in% names(x_raw))
-  # testthat::expect_gte(x_raw$`@odata.count`, nrow(x_parsed))
+  testthat::expect_true("@odata.count" %in% names(x_raw))
+  testthat::expect_gte(x_raw$`@odata.count`, nrow(x_parsed))
 })
 
 
 test_that("odata_submission_get parses WKT geopoint", {
-
-  testthat::skip() # https://github.com/dbca-wa/ruODK/issues/65
 
   df <- odata_submission_get(
     wkt = TRUE,
@@ -182,19 +186,23 @@ test_that("odata_submission_get parses WKT geopoint", {
   print(df)
   print(names(df))
 
-  # testthat::expect_true("location_corner1_latitude" %in% names(df))
-  # testthat::expect_true("location_corner1_longitude" %in% names(df))
-  # testthat::expect_true("location_corner1_altitude" %in% names(df))
+  # https://github.com/dbca-wa/ruODK/issues/65
+  skip_on_travis()
+  skip_on_appveyor()
 
-  # testthat::expect_true("perimeter_corner4_latitude" %in% names(df))
-  # testthat::expect_true("perimeter_corner4_longitude" %in% names(df))
-  # testthat::expect_true("perimeter_corner4_altitude" %in% names(df))
+  testthat::expect_true("location_corner1_latitude" %in% names(df))
+  testthat::expect_true("location_corner1_longitude" %in% names(df))
+  testthat::expect_true("location_corner1_altitude" %in% names(df))
+
+  testthat::expect_true("perimeter_corner4_latitude" %in% names(df))
+  testthat::expect_true("perimeter_corner4_longitude" %in% names(df))
+  testthat::expect_true("perimeter_corner4_altitude" %in% names(df))
 
   print("df$location_corner1_longitude class: ")
   print(class(df$location_corner1_longitude))
-  # testthat::expect_equal(class(df$location_corner1_longitude), "numeric")
-  # testthat::expect_equal(class(df$perimeter_corner2_latitude), "numeric")
-  # testthat::expect_equal(class(df$perimeter_corner3_altitude), "numeric")
[…]
