### 2020-04-29 858c736
TravisCI: re-enable deprecated "skip_cleanup: true"
