### 2020-04-29 ccb8aac
Close #66: use httr::modify_url to build URLs over glue::glue
