+++ R/form_detail.R
@@ -48,7 +48,7 @@ form_detail <- function(pid = get_default_pid(),
   yell_if_missing(url, un, pw, pid = pid, fid = fid)
   httr::RETRY(
     "GET",
-    glue::glue("{url}/v1/projects/{pid}/forms/{fid}"),
+    httr::modify_url(url, path = glue::glue("v1/projects/{pid}/forms/{fid}")),
     httr::add_headers(
       "Accept" = "application/xml",
       "X-Extended-Metadata" = "true"
