+++ R/attachment_get.R
@@ -69,6 +69,14 @@ attachment_url <- function(uuid,
   glue::glue(
     "{url}/v1/projects/{pid}/forms/{fid}/submissions/{uuid}/attachments/{fn}"
   )
+  # See https://github.com/dbca-wa/ruODK/issues/66
+  # This breaks attachment_get tests:
+  # httr::modify_url(
+  #   url,
+  #   path = glue::glue(
+  #     "v1/projects/{pid}/forms/{fid}/submissions/{uuid}/attachments/{fn}"
+  #   )
+  # )
 }
 
 #' Download one media attachment.
