+++ R/form_schema.R
@@ -161,19 +161,15 @@ form_schema <- function(flatten = FALSE,
                         odkc_version = get_default_odkc_version(),
                         verbose = get_ru_verbose()) {
   yell_if_missing(url, un, pw, pid = pid, fid = fid)
+  if (verbose == TRUE) ru_msg_info(glue::glue("form schema v{odkc_version}"))
 
   if (odkc_version < 0.8) {
-    url_v7 <- httr::modify_url(
-      url,
-      path = glue::glue("v1/projects/{pid}/forms/{fid}.schema.json")
-    )
-    if (verbose == TRUE) {
-      ru_msg_info(glue::glue("Reading v0.7 form schema from {url_v7}"))
-    }
-
     fs <- httr::RETRY(
       "GET",
-      url_v7,
+      httr::modify_url(
+        url,
+        path = glue::glue("v1/projects/{pid}/forms/{fid}.schema.json")
+      ),
       httr::add_headers("Accept" = "application/json"),
       httr::authenticate(un, pw),
       query = list(flatten = flatten, odata = odata)
@@ -185,7 +181,7 @@ form_schema <- function(flatten = FALSE,
       if (flatten == TRUE) {
         ru_msg_warn(
           "Cannot parse flattened form schema, returning unparsed and flattened.
-        Use flatten=FALSE with parse=TRUE for a parsed form_schema."
+           Use flatten=FALSE with parse=TRUE for a parsed form_schema."
         )
         return(fs)
       }
@@ -195,17 +191,12 @@ form_schema <- function(flatten = FALSE,
     }
     return(fs)
   } else {
-    url_v8 <- httr::modify_url(
-      url,
-      path = glue::glue("v1/projects/{pid}/forms/{fid}/fields")
-    )
-    if (verbose == TRUE) {
-      ru_msg_info(glue::glue("Reading v0.8 form fields from {url_v8}"))
-    }
-
     httr::RETRY(
       "GET",
-      url_v8,
+      httr::modify_url(
+        url,
+        path = glue::glue("v1/projects/{pid}/forms/{fid}/fields")
+      ),
       httr::add_headers("Accept" = "application/json"),
       httr::authenticate(un, pw),
       query = list(flatten = flatten, odata = odata)
