### 2020-04-30 f4a64d7
test-odata_submission_get: turn verbose to trigger handle_ru_geopoint verbose messages
