+++ .github/workflows/main.yml
@@ -44,6 +44,25 @@ jobs:
       BUILD_PKGDOWN: ${{ matrix.config.pkgdown }}
       # macOS >= 10.15.4 linking
       SDKROOT: /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk
+      # ruODK test env variables
+      ODKC_URL: ${{ secrets.ODKC_URL }}
+      ODKC_UN: ${{ secrets.ODKC_UN }}
+      ODKC_PW: ${{ secrets.ODKC_PW }}
+      ODKC_PID: ${{ secrets.ODKC_PID }}
+      ODKC_FID: ${{ secrets.ODKC_FID }}
+      ODKC_TEST_SVC: ${{ secrets.ODKC_TEST_SVC }}
+      ODKC_TEST_URL: ${{ secrets.ODKC_TEST_URL }}
+      ODKC_TEST_PID: ${{ secrets.ODKC_TEST_PID }}
+      ODKC_TEST_FID: ${{ secrets.ODKC_TEST_FID }}
+      ODKC_TEST_FID_ZIP: ${{ secrets.ODKC_TEST_FID_ZIP }}
+      ODKC_TEST_FID_ATT: ${{ secrets.ODKC_TEST_FID_ATT }}
+      ODKC_TEST_FID_GAP: ${{ secrets.ODKC_TEST_FID_GAP }}
+      ODKC_TEST_UN: ${{ secrets.ODKC_TEST_UN }}
+      ODKC_TEST_PW: ${{ secrets.ODKC_TEST_PW }}
+      ODKC_VERSION: ${{ secrets.ODKC_VERSION }}
+      ODKC_TEST_VERSION: ${{ secrets.ODKC_TEST_VERSION }}
+      RU_VERBOSE: ${{ secrets.RU_VERBOSE }}
+      RU_TIMEZONE: ${{ secrets.RU_TIMEZONE }}
 
     steps:
       - uses: actions/checkout@v2
