### 2020-04-30 d2f4d75
Move pkgdown deploy site github from TravisCI to GHA
