### 2020-05-12 7169676
Re-enabling Leroy Jenkins tests
