### 2020-05-12 4cc4560
GHA: firmly step on build matrix landmines to see whether they still blow up
