+++ .github/workflows/main.yml
@@ -18,16 +18,15 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          # comment out lines if you do not want to build on certain platforms
           - { os: windows-latest, r: "release" }
-          # - { os: windows-latest, r: "oldrel" } # fails on install: rlang not avail for tic
+          - { os: windows-latest, r: "oldrel" } # fails on install: rlang not avail for tic
           - { os: windows-latest, r: "devel" }
           - { os: macOS-latest, r: "release"}
           - { os: macOS-latest, r: "oldrel"}
           - { os: macOS-latest, r: "devel" }
           - { os: ubuntu-latest, r: "release", pkgdown: "true"}
           - { os: ubuntu-latest, r: "oldrel" }
-          # - { os: ubuntu-latest, r: "devel" } # fails on setup-r
+          - { os: ubuntu-latest, r: "devel" } # fails on setup-r
 
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
