+++ .github/workflows/main.yml
@@ -19,14 +19,17 @@ jobs:
       matrix:
         config:
           - { os: windows-latest, r: "release" }
-          - { os: windows-latest, r: "oldrel" } # fails on install: rlang not avail for tic
+          # - { os: windows-latest, r: "oldrel" } # fails on script: dependency 'tidyselect' is not available
           - { os: windows-latest, r: "devel" }
           - { os: macOS-latest, r: "release"}
           - { os: macOS-latest, r: "oldrel"}
           - { os: macOS-latest, r: "devel" }
           - { os: ubuntu-latest, r: "release", pkgdown: "true"}
           - { os: ubuntu-latest, r: "oldrel" }
-          - { os: ubuntu-latest, r: "devel" } # fails on setup-r
+          # - { os: ubuntu-latest, r: "devel" } # fails on setup-r
+          - { os: ubuntu-16.04, r: "release", pkgdown: "true"}
+          - { os: ubuntu-16.04, r: "oldrel" }
+          - { os: ubuntu-16.04, r: "devel" }
 
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
