### 2020-05-12 3bb3154
GHA: add libicu-dev for stringi on ubuntu 16.04
