### 2020-05-12 4a69879
GHA: drop ubuntu 16.04 R devel
