+++ .github/workflows/main.yml
@@ -27,9 +27,9 @@ jobs:
           - { os: ubuntu-latest, r: "release", pkgdown: "true"}
           - { os: ubuntu-latest, r: "oldrel" }
           # - { os: ubuntu-latest, r: "devel" } # fails on setup-r
-          - { os: ubuntu-16.04, r: "release", pkgdown: "true"}
+          - { os: ubuntu-16.04, r: "release"}
           - { os: ubuntu-16.04, r: "oldrel" }
-          - { os: ubuntu-16.04, r: "devel" }
+          # - { os: ubuntu-16.04, r: "devel" } # fails on setup-r
 
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
