+++ .github/workflows/main.yml
@@ -27,10 +27,6 @@ jobs:
           - { os: ubuntu-latest, r: "release", pkgdown: "true"}
           - { os: ubuntu-latest, r: "oldrel" }
           # - { os: ubuntu-latest, r: "devel" } # fails on setup-r
-          - { os: ubuntu-16.04, r: "release"}
-          - { os: ubuntu-16.04, r: "oldrel" }
-          # - { os: ubuntu-16.04, r: "devel" } # fails on setup-r
-
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
@@ -121,7 +117,7 @@ jobs:
       - name: "[Linux] ccache"
         if: runner.os == 'Linux'
         run: |
-          sudo apt install ccache libcurl4-openssl-dev libicu-dev
+          sudo apt install ccache libcurl4-openssl-dev
           mkdir -p ~/.R && echo -e 'CC=ccache gcc -std=gnu99\nCXX=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
 
       # install ccache and write config file
