+++ README.md
@@ -13,15 +13,11 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 [![Last-changedate](https://img.shields.io/github/last-commit/dbca-wa/ruODK.svg)](https://github.com/dbca-wa/ruODK/commits/master)
 [![GitHub
 issues](https://img.shields.io/github/issues/dbca-wa/ruodk.svg?style=popout)](https://github.com/dbca-wa/ruODK/issues/)
-[![Travis build
-status](https://travis-ci.org/dbca-wa/ruODK.svg?branch=master)](https://travis-ci.org/dbca-wa/ruODK)
-[![AppVeyor build
-status](https://ci.appveyor.com/api/projects/status/github/dbca-wa/ruODK?branch=master&svg=true)](https://ci.appveyor.com/project/dbca-wa/ruODK)
+[![CI via
+{tic}](https://github.com/dbca-wa/ruODK/workflows/R%20CMD%20Check%20via%20%7Btic%7D/badge.svg?branch=master)](https://github.com/dbca-wa/ruODK/actions)
 [![Coverage
 status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
-[![R CMD Check via
-{tic}](https://github.com/dbca-wa/ruODK/workflows/R%20CMD%20Check%20via%20%7Btic%7D/badge.svg?branch=master)](https://github.com/dbca-wa/ruODK/actions)
 <!-- badges: end -->
 
 `ruODK` is an R client to access and parse data from ODK Central.
@@ -73,13 +69,13 @@ Benefits of using the R ecosystem in combination with ODK:
     Scaling to many users does not incur license fees.
   - Ubiquity: R is known to many scientists and is widely taught at
     universities.
-  - Automation: The entire data access workflow can be automated through
-    R scripts.
+  - Automation: The entire data access and analysis workflow can be
+    automated through R scripts.
   - Reproducible reporting (e.g. 
     [Sweave](https://support.rstudio.com/hc/en-us/articles/200552056-Using-Sweave-and-knitr),
     [RMarkdown](https://rmarkdown.rstudio.com/)), interactive web apps
     ([Shiny](https://shiny.rstudio.com/)), workflow scaling
-    ([drake](https://docs.ropensci.org/drake/))
+    ([drake](https://docs.ropensci.org/drake/)).
   - Rstudio-as-a-Service (RaaS) at
     [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 
@@ -225,8 +221,8 @@ proj %>% head() %>% knitr::kable(.)
 | 57 | AB’s                     |     1 |          1 | 2020-01-06 08:56:00 | NA          | NA                  | FALSE    |
 |  8 | AmenazasRD               |     1 |          1 | 2019-04-14 14:21:00 | NA          | 2020-04-21 14:23:06 | FALSE    |
 |  9 | BAFCO Test               |     1 |          1 | 2019-04-19 08:30:45 | NA          | 2019-04-19 09:26:49 | FALSE    |
+| 79 | Birhan Project           |     1 |          1 | 2020-05-11 20:58:06 | NA          | 2020-05-12 00:09:16 | FALSE    |
 | 47 | Building Materials Reuse |     0 |          0 | 2019-10-16 13:07:56 | NA          | NA                  | FALSE    |
[…]
