+++ README.Rmd
@@ -16,11 +16,9 @@ knitr::opts_chunk$set(
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Last-changedate](https://img.shields.io/github/last-commit/dbca-wa/ruODK.svg)](https://github.com/dbca-wa/ruODK/commits/master)
 [![GitHub issues](https://img.shields.io/github/issues/dbca-wa/ruodk.svg?style=popout)](https://github.com/dbca-wa/ruODK/issues/)
-[![Travis build status](https://travis-ci.org/dbca-wa/ruODK.svg?branch=master)](https://travis-ci.org/dbca-wa/ruODK)
-[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/dbca-wa/ruODK?branch=master&svg=true)](https://ci.appveyor.com/project/dbca-wa/ruODK)
+[![CI via {tic}](https://github.com/dbca-wa/ruODK/workflows/R%20CMD%20Check%20via%20{tic}/badge.svg?branch=master)](https://github.com/dbca-wa/ruODK/actions)
 [![Coverage status](https://codecov.io/gh/dbca-wa/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/dbca-wa/ruODK?branch=master)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
-[![R CMD Check via {tic}](https://github.com/dbca-wa/ruODK/workflows/R%20CMD%20Check%20via%20{tic}/badge.svg?branch=master)](https://github.com/dbca-wa/ruODK/actions)
 <!-- badges: end -->
 
 `ruODK` is an R client to access and parse data from ODK Central.
@@ -69,12 +67,13 @@ Benefits of using the R ecosystem in combination with ODK:
 *  Scalability: Both R and ODK are free and open source software. Scaling to
    many users does not incur license fees.
 *  Ubiquity: R is known to many scientists and is widely taught at universities.
-*  Automation: The entire data access workflow can be automated through R scripts.
+*  Automation: The entire data access and analysis workflow can be automated 
+   through R scripts.
 *  Reproducible reporting (e.g. 
    [Sweave](https://support.rstudio.com/hc/en-us/articles/200552056-Using-Sweave-and-knitr), 
    [RMarkdown](https://rmarkdown.rstudio.com/)), interactive web apps 
    ([Shiny](https://shiny.rstudio.com/)), 
-   workflow scaling ([drake](https://docs.ropensci.org/drake/))
+   workflow scaling ([drake](https://docs.ropensci.org/drake/)).
 * Rstudio-as-a-Service (RaaS) at 
   [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 
@@ -245,12 +244,10 @@ citation("ruODK")
 ```
 
 ## Acknowledgements
-The Department of Biodiversity, Conservation and Attractions (DBCA) respectfully 
-acknowledges Aboriginal people as the traditional owners of the lands and waters 
-it manages.
-
-One of the Department's core missions is to conserve and protect the value of 
-the land to the culture and heritage of Aboriginal people.
+The Department of Biodiversity, Conservation and Attractions (DBCA) acknowledges 
+the traditional owners of country throughout Western Australia and their continuing 
[…]
