### 2020-05-12 d8d36a3
Fix #60 encode fid
See also https://github.com/getodk/central-backend/issues/256
