+++ R/submission_detail.R
@@ -48,7 +48,10 @@ submission_detail <- function(iid,
     "GET",
     httr::modify_url(
       url,
-      path = glue::glue("v1/projects/{pid}/forms/{fid}/submissions/{iid}")
+      path = glue::glue(
+        "v1/projects/{pid}/forms/",
+        "{URLencode(fid, reserved = TRUE)}/submissions/{iid}"
+        )
     ),
     httr::add_headers(
       "Accept" = "application/json; extended",
