### 2020-05-12 4d29459
odata_submission_get: improve URL query string
* Drop: building query as URL path
* Replace with: build query as list
* Streamline verbose comments: remove duplicate "parsing" message
* Streamline: combine building query with parsing function parameters
