### 2020-05-12 3d1a543
Skip Heisentests on Travis/Appveyor
* odata_submission_get query not honoured on Travis or Appveyor, but OK on GH Actions (win/mac/ubuntu)
* Heisentests are not fixed on Travis and Appveyor after odata_submission_get query upgrade
