+++ tests/testthat/test-odata_submission_get.R
@@ -99,13 +99,11 @@ test_that("odata_submission_get skip omits number of results", {
     table = fq_svc$name[2] # brittle: depends on form used
   )
 
-  print("odata_submission_get without skip")
-  print(fresh_parsed)
-  print("odata_submission_get with skip=1 should return one less record")
-  print(skip_parsed)
-
-
   # https://github.com/dbca-wa/ruODK/issues/65
+  # print("odata_submission_get without skip")
+  # print(fresh_parsed)
+  # print("odata_submission_get with skip=1 should return one less record")
+  # print(skip_parsed)
   skip_on_travis()
   skip_on_appveyor()
   testthat::expect_true(nrow(fresh_parsed) == nrow(skip_parsed) + 1)
@@ -125,10 +123,10 @@ test_that("odata_submission_get top limits number of results", {
     parse = TRUE
   )
 
-  print("odata_submission_get with top=1 should return one record")
-  print(top_parsed)
 
   # https://github.com/dbca-wa/ruODK/issues/65
+  # print("odata_submission_get with top=1 should return one record")
+  # print(top_parsed)
   skip_on_travis()
   skip_on_appveyor()
   testthat::expect_true(nrow(top_parsed) == 1)
@@ -152,11 +150,11 @@ test_that("odata_submission_get count returns total number or rows", {
   )
   x_parsed <- x_raw %>% odata_submission_rectangle()
 
-  ru_msg_info("odata_submission_get count")
-  print(x_raw)
-  print(x_parsed)
 
   # https://github.com/dbca-wa/ruODK/issues/65
+  # ru_msg_info("odata_submission_get count")
+  # print(x_raw)
+  # print(x_parsed)
   skip_on_travis()
   skip_on_appveyor()
 
@@ -183,11 +181,11 @@ test_that("odata_submission_get parses WKT geopoint", {
     download = FALSE
   )
 
-  ru_msg_info("WKT test")
-  print(df)
-  print(names(df))
 
   # https://github.com/dbca-wa/ruODK/issues/65
+  # ru_msg_info("WKT test")
+  # print(df)
+  # print(names(df))
   skip_on_travis()
   skip_on_appveyor()
 
