+++ R/odata_submission_get.R
@@ -177,7 +177,7 @@ odata_submission_get <- function(table = "Submissions",
   # Download submissions
   if (verbose == TRUE) ru_msg_info("Downloading submissions...")
 
-  sub <- httr::RETRY(
+    sub <- httr::RETRY(
     "GET",
     httr::modify_url(
       url,
@@ -245,7 +245,7 @@ odata_submission_get <- function(table = "Submissions",
       }
     } %>%
     {
-      if (wkt == "true") {
+      if (wkt != FALSE) {
         handle_ru_geopoints(data = ., form_schema = fs, verbose = verbose)
       } else {
         .
