+++ .travis.yml
@@ -19,5 +19,5 @@ jobs:
 
 cache: packages
 
-after_success:
-  - Rscript -e 'covr::codecov()'
+# after_success:
+#   - Rscript -e 'covr::codecov()'
