### 2020-05-12 9fe29d0
Update README
re-enable Travis and Appveyor
