### 2020-05-14 9737379
WIP #69
* add new test form `fid_wkt` containing geopoints, geotraces, geoshapes with each widget appearance
* add test form to settings, settings tests, instructions to contributing
* odata_submission_rectangle now takes a form_schema (and passes it on to unnest_all) to exempt geo{point, trace, shape} from unnesting
* geo fields GeoJSON become nested lists, geo fields WKT remain character strings
* TODO: add handlers to value-add geo fields
