### 2020-05-18 92aa196
#69 Handle Geopoints either way up
* odata_submission_rectangle takes param form_schema to prevent unnesting of GeoJSON (nested list of fixed (geopoint) or arbitrary (geotrace, -shape) length == rectangling drama)
* odata_submission_get handballs all geopoint business to handle_ru_geopoints
* handle_ru_geopoints takes param form_schema to infer geopoint fields and wkt to tell split_geopoint whether to expect GeoJSON or WKT
* split_geopint takes param wkt from handle_ru_geopoints
* geopoint tests refactored

Remaining work: geotrace and geoshape
