### 2020-05-18 9c9eb45
#69 patch vignette restful-api
