### 2020-05-18 649288a
#69 patch geopoint handlers
