+++ tests/testthat/test-handle_ru_geopoints.R
@@ -1,4 +1,6 @@
-test_that("handle_geopoint annotates GeoJSON points with lon lat alt acc", {
+test_that(
+  "handle_ru_geopoints annotates GeoJSON points with lon lat alt acc",
+  {
 
   fs <- form_schema(
     pid = get_test_pid(),
@@ -21,7 +23,7 @@ test_that("handle_geopoint annotates GeoJSON points with lon lat alt acc", {
   )
 
   geo_fields <- fs %>%
-    dplyr::filter(type %in% c("geopoint", "geotrace", "geoshape")) %>%
+    dplyr::filter(type  == "geopoint") %>%
     magrittr::extract2("ruodk_name")
 
   # This test requires a form with geo fields
@@ -39,16 +41,18 @@ test_that("handle_geopoint annotates GeoJSON points with lon lat alt acc", {
 
     # GeoJSON should still exist  after handle_ru_geopoint as nested lists
     testthat::expect_true(
-      is.list(d[[geo_fields[1]]]),
-      label = glue::glue("GeoJSON field {geo_fields[i]} should be a nested list")
+      is.list(d[[geo_fields[i]]]),
+      label = glue::glue(
+        "GeoJSON field {geo_fields[i]} should be a nested list"
+      )
     )
 
     # handle_ru_geopoints should have appended new fields with postfixes
     # for lon, lat, alt, acc
-    geofield_lon <- glue::glue("{geo_fields[1]}_longitude")
-    geofield_lat <- glue::glue("{geo_fields[1]}_latitude")
-    geofield_alt <- glue::glue("{geo_fields[1]}_altitude")
-    geofield_acc <- glue::glue("{geo_fields[1]}_accuracy")
+    geofield_lon <- glue::glue("{geo_fields[i]}_longitude")
+    geofield_lat <- glue::glue("{geo_fields[i]}_latitude")
+    geofield_alt <- glue::glue("{geo_fields[i]}_altitude")
+    geofield_acc <- glue::glue("{geo_fields[i]}_accuracy")
     testthat::expect_true(
       geofield_lon %in% names(d),
       label = glue::glue("handle_ru_geopoint appends {geofield_lon}")
@@ -65,11 +69,30 @@ test_that("handle_geopoint annotates GeoJSON points with lon lat alt acc", {
       geofield_acc %in% names(d),
       label = glue::glue("handle_ru_geopoint appends {geofield_acc}")
     )
+    # Fields must be numeric
+    testthat::expect_true(
+      is.numeric(d[,geofield_lon][[1]]),
+      label = glue::glue("handle_ru_geopoint {geofield_lon} is numeric")
+    )
+    testthat::expect_true(
+      is.numeric(d[,geofield_lat][[1]]),
+      label = glue::glue("handle_ru_geopoint {geofield_lat} is numeric")
+    )
+    testthat::expect_true(
+      is.numeric(d[,geofield_alt][[1]]),
+      label = glue::glue("handle_ru_geopoint {geofield_alt} is numeric")
+    )
+    testthat::expect_true(
+      is.numeric(d[,geofield_acc][[1]]),
+      label = glue::glue("handle_ru_geopoint {geofield_acc} is numeric")
+    )
   }
 
 })
 
-test_that("handle_geopoint annotates WKT points with lon lat alt (no acc)", {
+test_that(
+  "handle_ru_geopoints annotates WKT points with lon lat alt (no acc)",
+  {
 
   fs <- form_schema(
     pid = get_test_pid(),
@@ -110,16 +133,16 @@ test_that("handle_geopoint annotates WKT points with lon lat alt (no acc)", {
 
     # GeoJSON should still exist after handle_ru_geopoint as nested lists
[…]
