### 2020-05-19 d68ae59
#69 First light for geotrace
* handle_ru_geotrace() retains original format (GeoJSON or WKT) and extracts lon, lat, alt from the first point of the geotrace.
* Add package data with example geofields, use in tests and examples, document
* Add source for geolocation test form as odkbuild and xml
* Add ODKC_TEST_FID_WKT to GHA
* RMD skeleton updated with comments re geofields
* Bump dev version
