+++ man-roxygen/param-wkt.R
@@ -1,0 +1,2 @@
+#' @param wkt Whether geofields are GeoJSON (if FALSE) or WKT
+#'   strings (if TRUE), default: FALSE.
