### 2020-05-19 50ef769
#69 First light for geoshape
* bump version
* handle_ru_geoshape, split_geoshape work very similar to geotrace equivalents
* refactor tests
* update packaged geofields test data with parsed geoshape
* make_data: bugfix geo_fs: set form_schema
