+++ tests/testthat/test-handle_ru_geoshapes.R
@@ -1,0 +1,4 @@
+test_that("multiplication works", {
+  expect_equal(2 * 2, 4)
+})
+# usethis::use_r("handle_ru_geoshapes")
