### 2020-05-19 f35caae
#69 Update vignette odata-api
* streamline spatial bits
* insert comments on how to split up and simplify vignettes
