+++ vignettes/odata-api.Rmd
@@ -37,7 +37,11 @@ library(dplyr)
 library(ggplot2)
 library(ruODK)
 ```
-
+<!-- TODO split out vignettes -->
+<!-- * This vignette: the typical workflow -->
+<!-- * New vignette: spatial is not special -->
+<!-- * New vignette: DIY rectangling -->
+<!-- * New vignette: data visualisation (refer to that from odata and resful api) -->
 This vignette demonstrates `ruODK`'s workflow to extract data from ODK Central's
 OData service endpoint, and to prepare the data and the media attachments for
 further analysis and visualisation.
@@ -97,20 +101,22 @@ fq_svc <- ruODK::odata_service_get()
 ```
 
 The same data is included as example data `fq_svc`.
+
 ```{r view_service}
 fq_svc %>% knitr::kable(.)
 ```
 
 `ruODK` provides the names and urls of the service endpoints as `tibble`.
 We see the main data available under the url `Submissions`, and repeating 
-groups called `taxon_encounter` and `vegetation_stratum` in the ODK form under 
+form groups called `taxon_encounter` and `vegetation_stratum` in the ODK form under 
 the url `Submissions.taxon_encounter` and `Submissions.vegetation_stratum`, 
 respectively.
 
-The number and names of repeating subgroups of course depends on the ODK form.
+The main value we get out of the service document are these names of the 
+form groups, which can differ between forms.
 
 ## OData metadata document
-Next, let's get the metadata to review our data schema.
+Let's inspect the form metadata to review our data schema.
 While we can download the submission data without it, the metadata document 
 contains information about field data types and attachment names.
 
@@ -125,6 +131,7 @@ listviewer::jsonedit(fq_meta)
 As an alternative to the OData metadata document, ODK Central also offers
 form metadata as a much cleaner JSON document, which `ruODK` can read and parse
 into a clean `tibble` of field type, name, and path. 
+
 `ruODK` uses this introspection to parse submission data.
 
 ```{r form_schema, eval=FALSE}
@@ -138,13 +145,19 @@ fq_form_schema %>% knitr::kable(.)
 ## OData submission data documents
 Now let's download the form submissions and, separately, repeating form groups.
 `ruODK::odata_submission_get()` defaults to download the submission data,
-parse it into a tidy tibble, parses dates and datetimes, and finally downloads
-and links file attachments. 
+parse it into a tidy tibble, parses dates and datetimes, downloads
+and links file attachments, and handles spatial datatypes.
+
 This vignette is built with canned data, so the verbose messages are not shown.
-If we use `wkt=TRUE`, geopoint fields are automatically split into their
-individual coordinates and prefixed with their respective original field name.
-If we use `wkt=FALSE` (the default), geopoints will be split but receive
-anonymous names like `x10`, `x11`, `x12` which then need to be renamed.
+
+With `wkt=TRUE`, we'll receive spatial types as Well Known Text, which `ruODK` parses
+as plain text.
+With `wkt=FALSE` (the default), we'll receive spatial types as GeoJSON, which
+`ruODK` parses into a nested list.
+`ruODK` retains the original spatial field, and annotates the data with 
+extracted longitude, latitude, altitude, and (where given) accuracy.
+These additional fields are prefixed with the original field name to prevent 
+name collisions between possibly multiple location fields.
 
 ```{r load_odata, eval=FALSE}
[…]
