### 2020-06-09 d51df30
odata_submission_get: update help and examples
