### 2020-06-15 4b30883
Address GHA build warning: update magrittr pipe import
* Issue: GHA fails only on Windows-latest / R-devel with `Non-file package-anchored link(s) in documentation object 'pipe.Rd':'[magrittr]{%>%}'`
* Fix: re-run `usethis::use_pipe()`, rebuild documentation (updates all "utilities" Rd)
