+++ R/utils-pipe.R
@@ -1,10 +1,9 @@
 #' Pipe operator
 #'
-#' See \code{magrittr::\link[magrittr]{\%>\%}} for details.
+#' See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
 #'
 #' @name %>%
 #' @rdname pipe
-#' @family utilities
 #' @keywords internal
 #' @export
 #' @importFrom magrittr %>%
