+++ man/strip_uuid.Rd
@@ -26,7 +26,6 @@ strip_uuid("uuid:d3bcefea-32a8-4dbc-80ca-4ecb0678e2b0")
 }
 \seealso{
 Other utilities: 
-\code{\link{\%>\%}()},
 \code{\link{attachment_get}()},
 \code{\link{attachment_url}()},
 \code{\link{form_schema_parse}()},
