+++ man/predict_ruodk_name.Rd
@@ -32,7 +32,6 @@ the form path.
 }
 \seealso{
 Other utilities: 
-\code{\link{\%>\%}()},
 \code{\link{attachment_get}()},
 \code{\link{attachment_url}()},
 \code{\link{form_schema_parse}()},
