+++ man/ru_msg_success.Rd
@@ -17,7 +17,6 @@ ru_msg_success("This is a success message.")
 }
 \seealso{
 Other utilities: 
-\code{\link{\%>\%}()},
 \code{\link{attachment_get}()},
 \code{\link{attachment_url}()},
 \code{\link{form_schema_parse}()},
