+++ man/tidyeval.Rd
@@ -22,7 +22,6 @@ To learn more about tidy eval and how to use these tools, read
 }
 \seealso{
 Other utilities: 
-\code{\link{\%>\%}()},
 \code{\link{attachment_get}()},
 \code{\link{attachment_url}()},
 \code{\link{form_schema_parse}()},
