### 2020-06-15 d5c5dee
GHA: re-enable known failing matrix envs
* R 4.0.0 has been released now, semantic R versions (devel, release, oldrel) have shifted, R ecosystem has updated
* add windows latest & R oldrel: failed on "tidyselect not available'
* add ubuntu latest & R devel: failed on setup-r
* add ubuntu 18.04
