### 2020-06-15 8e15bfb
Typeset references for handle_ru_geo*
* Change Markdown backticks to Latex `\code{}`
* https://github.com/dbca-wa/ruODK/issues/58
