+++ man/split_geoshape.Rd
@@ -18,10 +18,10 @@ strings (if TRUE), default: FALSE.}
 }
 \value{
 The given dataframe with the geoshape column <colname>, plus
-  three new columns, `<colname>_longitude`, `<colname>_latitude`,
-  `<colname>_altitude`.
-  The three new columns are prefixed with the original `colname` to avoid
-  naming conflicts with any other geoshape columns.
+  three new columns, \code{<colname>_longitude}, \code{<colname>_latitude},
+  \code{<colname>_altitude}.
+  The three new columns are prefixed with the original \code{colname} to
+  avoid naming conflicts with any other geoshape columns.
 }
 \description{
 \lifecycle{stable}
