+++ man/split_geopoint.Rd
@@ -17,10 +17,10 @@ strings (if TRUE), default: FALSE.}
 }
 \value{
 The given dataframe with the WKT POINT column <colname>, plus
-  three new columns, `<colname>_longitude`, `<colname>_latitude`,
-  `<colname>_altitude`.
-  The three new columns are prefixed with the original `colname` to avoid
-  naming conflicts with any other geopoint columns.
+  three new columns, \code{<colname>_longitude}, \code{<colname>_latitude},
+  \code{<colname>_altitude}.
+  The three new columns are prefixed with the original \code{colname} to
+  avoid naming conflicts with any other geopoint columns.
 }
 \description{
 \lifecycle{stable}
