### 2020-06-15 77c11e2
Fix stale references in vignettes
* https://github.com/dbca-wa/ruODK/issues/58
* odata Rmd template: sort imports
* odata vignette: replace stale ref to attachment_get / split_geopoint with handle_ru_{attachments/geopoints}
