### 2020-06-15 3f04ad9
GHA: drop failing Ubuntu / R-devel combos from build matrix
