### 2020-06-15 #335 @florianm
Hi all,

here's a summary of the [second round (Jason's comments)](https://github.com/ropensci/software-review/issues/335#issuecomment-596120460) of reviewer comments. I've split Jason's comments into the three issues linked under his comments, which I'll also summarise here.

### Internal helpers
[Details](https://github.com/dbca-wa/ruODK/issues/56)

**TLDR** Jason asked to re-think which functions to export, and which to keep purely internal. I have changed `ruODK` from "export all the functions" to export only the top level (`odata_submission_get`) and the next level of helpers (`handle_ru_*`). This allows users to choose the level of automation:

* First level: parse all the things in one go - the default: `odata_submission_get()`
* Second level: parse step by step - explained in vignettes under "DIY rectangling"
  * Several functions which are useful for step by step data rectangling have been moved to the reference family ["Utilities"](https://dbca-wa.github.io/ruODK/reference/index.html).
  * `handle_ru_{datetimes, attachments, geopoints, geotraces, geoshapes}`:
    * Naming streamlined: prefixed `handle_ru_` and pluralised
    * [man](https://dbca-wa.github.io/ruODK/reference/handle_ru_datetimes.html)
    * Mentioned in vignettes [odata-api > data rectangling](https://dbca-wa.github.io/ruODK/articles/odata-api.html#detour-data-rectangling)
      and [restful-api](https://dbca-wa.github.io/ruODK/articles/restful-api.html) where required in the context of manual rectangling
  * `split_geo{point, trace, shape}` have [examples](https://dbca-wa.github.io/ruODK/reference/split_geopoint.html#examples)
    to show order of lon / lat / alt (coordinate order is notorious for swap errors vs. geojson)
  * `attachment_get` is mentioned in vignette [odata-api > data rectangling](https://dbca-wa.github.io/ruODK/articles/odata-api.html#detour-data-rectangling)
* The third level helpers used by the exported `handle_ru_*` handlers then are all internal, and not mentioned in any of man / vignette / readme:
  * isodt_to_local
  * strip_uuid
  * prepend_uuid
  * attachment_url
  * unnest_all

### Update all documentation
[Details](https://github.com/dbca-wa/ruODK/issues/58)

**TLDR** Jason found some outdated references.

All documentation should now be aligned and tell the same story.
Some new functionality brought another round of sweeping changes:

* `split_geo{point, trace, shape}`: `ruODK` can now parse all spatial types,
  GeoJSON and WKT alike.
* The tests use canned data of a new spatial test form
* Rebuilding the canned data requires a new env var for the spatial test form. `CONTRIBUTING.md` has been updated to include the new test form.

### Re-think `attachment_link`
[Details](https://github.com/dbca-wa/ruODK/issues/57)

**TLDR** Jason asked whether `attachment_link` should be internal or exported.

* Listed under "RESTful API" in [man](https://dbca-wa.github.io/ruODK/reference/attachment_link.html)
* Demonstrated in vignette [restful-api](https://dbca-wa.github.io/ruODK/articles/restful-api.html#get-all-submissions-for-one-form) - we're looking at the "export everything including attachments as one giant ZIP file" use case here. The difference to `handle_ru_attachments` is that the attachments are already downloaded and only need to be linked to the relative path of the downloaded and freshly extracted attachments.

I've just refreshed the [GitHub Actions](https://github.com/dbca-wa/ruODK/actions/) build matrix, adding more environments (10 in total). Plus TravisCI, plus Appveyor.

@maelle does the above address the reviewer comments suitably?
