### 2020-06-15 956bdcf
Formatting
* styler::style_active_package()
* goodpgractice::goodpractice()
* add note to cran-comments.md: not feasible to test form_schema(odkc_version=0.7)
* GHA: use semantic versions (-latest), add ubuntu-20.04
