+++ R/form_schema_parse.R
@@ -96,12 +96,12 @@ form_schema_parse <- function(fs,
 #' @keywords internal
 #' @examples
 #' \dontrun{
-#'   predict_ruodk_name("bar", "Submissions.foo")
-#'   # > "foo_bar"
-#'   predict_ruodk_name("bar", "Submissions")
-#'   # > "bar"
-#'   predict_ruodk_name("rock", "Submissions.foo_fighters")
-#'   # > "foo_fighters_rock"
+#' predict_ruodk_name("bar", "Submissions.foo")
+#' # > "foo_bar"
+#' predict_ruodk_name("bar", "Submissions")
+#' # > "bar"
+#' predict_ruodk_name("rock", "Submissions.foo_fighters")
+#' # > "foo_fighters_rock"
 #' }
 predict_ruodk_name <- function(name_str, path_str) {
   prefix <- path_str %>%
