+++ R/ru_setup.R
@@ -225,7 +225,7 @@ odata_svc_parse <- function(svc) {
 #'   test_fid_zip = "build_Spotlighting-0-6_1558333698",
 #'   test_fid_att = "build_Flora-Quadrat-0-1_1558330379",
 #'   test_fid_gap = "build_Turtle-Track-or-Nest-1-0_1569907666",
-#'   test_fid_wkt="build_Locations_1589344221",
+#'   test_fid_wkt = "build_Locations_1589344221",
 #'   verbose = TRUE
 #' )
 ru_setup <- function(svc = NULL,
@@ -242,7 +242,7 @@ ru_setup <- function(svc = NULL,
                      test_fid_zip = NULL,
                      test_fid_att = NULL,
                      test_fid_gap = NULL,
-                     test_fid_wkt=NULL,
+                     test_fid_wkt = NULL,
                      test_url = NULL,
                      test_un = NULL,
                      test_pw = NULL,
