### 2020-06-15 3ee18d1
GHA: Drop failing ubuntu-20.04
* E: The repository 'http://ppa.launchpad.net/cran/travis/ubuntu focal Release' does not have a Release file.
