### 2020-06-15 ac1599e
Update docs, fix links
* README, vignette odata-api: update broken links
* cran-comments: update CI
* GHA: add ubuntu-16.04
* make_release: change check() to cran=T
* test-handle-ru_geopoint: make comparisons more explicit (fail in Appveyor, pass in GHA)
