+++ README.md
@@ -49,7 +49,7 @@ form download and submission upload.
 ruODK](https://www.lucidchart.com/publicSegments/view/952c1350-3003-48c1-a2c8-94bad74cdb46/image.png)
 
 A typical [ODK workflow](https://docs.getodk.org/#how-is-odk-used): An
-XForm is designed e.g. in [ODK Build](https://build.getodk.org//),
+XForm is designed e.g. in [ODK Build](https://build.getodk.org/),
 [published to ODK Central](https://docs.getodk.org/central-forms/), and
 downloaded onto an Android device running ODK Collect. After data have
 been captured digitally using [ODK
@@ -180,12 +180,13 @@ read-permitted ODK Central web user. Adjust verbosity to your liking.
     #>   Default ODK Central Username:  Florian.Mayer@dbca.wa.gov.au 
     #>   Default ODK Central Password: run ruODK::get_default_pw() to show 
     #>   Default Time Zone:  Australia/Perth 
-    #>   Default ODK Central Version:  0.8 
+    #>   Default ODK Central Version:  0.7 
     #>   Test ODK Central Project ID: 14 
     #>   Test ODK Central Form ID: build_Flora-Quadrat-0-4_1564384341 
     #>   Test ODK Central Form ID (ZIP tests): build_Spotlighting-0-6_1558333698 
     #>   Test ODK Central Form ID (Attachment tests): build_Flora-Quadrat-0-1_1558330379 
     #>   Test ODK Central Form ID (Parsing tests): build_Turtle-Track-or-Nest-1-0_1569907666 
+    #>   Test ODK Central Form ID (WKT tests): build_Locations_1589344221 
     #>   Test ODK Central URL: https://sandbox.central.getodk.org 
     #>   Test ODK Central Username: Florian.Mayer@dbca.wa.gov.au 
     #>   Test ODK Central Password: run ruODK::get_test_pw() to show 
@@ -219,14 +220,14 @@ proj <- ruODK::project_list()
 proj %>% head() %>% knitr::kable(.)
 ```
 
-| id | name                     | forms | app\_users | created\_at         | updated\_at | last\_submission    | archived |
-| -: | :----------------------- | ----: | ---------: | :------------------ | :---------- | :------------------ | :------- |
-| 59 | 1Horizon test project    |     1 |          3 | 2020-01-13 20:45:32 | NA          | 2020-01-23 22:20:13 | FALSE    |
-| 57 | AB’s                     |     1 |          1 | 2020-01-06 08:56:00 | NA          | NA                  | FALSE    |
-|  8 | AmenazasRD               |     1 |          1 | 2019-04-14 14:21:00 | NA          | 2020-04-21 14:23:06 | FALSE    |
-|  9 | BAFCO Test               |     1 |          1 | 2019-04-19 08:30:45 | NA          | 2019-04-19 09:26:49 | FALSE    |
-| 79 | Birhan Project           |     1 |          1 | 2020-05-11 20:58:06 | NA          | 2020-05-12 00:09:16 | FALSE    |
-| 47 | Building Materials Reuse |     0 |          0 | 2019-10-16 13:07:56 | NA          | NA                  | FALSE    |
+| id | name                  | forms | app\_users | created\_at         | updated\_at | last\_submission    | archived |
+| -: | :-------------------- | ----: | ---------: | :------------------ | :---------- | :------------------ | :------- |
+| 59 | 1Horizon test project |     1 |          4 | 2020-01-13 20:45:32 | NA          | 2020-01-23 22:20:13 | FALSE    |
+| 57 | AB’s                  |     1 |          1 | 2020-01-06 08:56:00 | NA          | NA                  | FALSE    |
[…]
