+++ vignettes/odata-api.Rmd
@@ -295,8 +295,8 @@ fq_data_strata <- fq_strata %>%
 ```
 
 Note: A manually resized version of the original photos in this example live in
-the package source under [`docs/articles/attachments`
-](https://github.com/dbca-wa/ruODK/tree/master/docs/articles/attachments). 
+the package source under [`articles/media`
+](https://github.com/dbca-wa/ruODK/tree/gh-pages/articles/media). 
 To minimise package size, they were resized with imagemagick:
 
 ```{sh, eval=FALSE}
