+++ README.Rmd
@@ -47,7 +47,7 @@ form download and submission upload.
 https://www.lucidchart.com/publicSegments/view/952c1350-3003-48c1-a2c8-94bad74cdb46/image.png)
 
 A typical [ODK workflow](https://docs.getodk.org/#how-is-odk-used):
-An XForm is designed e.g. in [ODK Build](https://build.getodk.org//),
+An XForm is designed e.g. in [ODK Build](https://build.getodk.org/),
 [published to ODK Central](https://docs.getodk.org/central-forms/), 
 and downloaded onto an Android device running ODK Collect.
 After data have been captured digitally using 
@@ -199,22 +199,23 @@ frms %>% knitr::kable(.)
 # Form details of default form
 frmd <- ruODK::form_detail()
 frmd %>% knitr::kable(.)
-# Form schema
-meta <- ruODK::form_schema()
+# Form schema: defaults to version 0.8
+meta <- ruODK::form_schema(odkc_version = get_test_odkc_version())
 meta %>% knitr::kable(.)
 # Part 2: Data access ---------------------------------------------------------#
 # Form tables
 srv <- ruODK::odata_service_get()
 srv %>% knitr::kable(.)
 # Form submissions
-data <- ruODK::odata_submission_get(local_dir = fs::path("vignettes/media"))
+data <- ruODK::odata_submission_get(local_dir = fs::path("vignettes/media"),
+                                    odkc_version = get_test_odkc_version())
 data %>% knitr::kable(.)
 ```
 
 A more detailed walk-through with some data visualisation examples is available
-in the [`vignette("odata-api", package="ruODK")`](https://dbca-wa.github.io/ruODK/articles/odata.html).
+in the [`vignette("odata-api", package="ruODK")`](https://dbca-wa.github.io/ruODK/articles/odata-api.html).
 
-See also [`vignette("restful-api", package="ruODK")`](https://dbca-wa.github.io/ruODK/articles/api.html)
+See also [`vignette("restful-api", package="ruODK")`](https://dbca-wa.github.io/ruODK/articles/restful-api.html)
 for examples using the alternative RESTful API.
 
 ## Try ruODK
@@ -269,7 +270,7 @@ There are several other R packages interacting with the ODK ecosystem, and/or
[…]
