### 2020-06-15 f11b70d
GHA: skip ubuntu-16-04
* GHA ubuntu-16.04 fails setup-r
* TravisCI still tests ubuntu-16.04 so we're good.
