### 2020-06-15 9ae4718
Clean up docs, update cran-comments
* Fix stale links: vignette, README
* Explain CMD CHECK notes
* Drop unused attachments folder in vignettes (using /media now)
