### 2020-06-16 918af8a
GHA: add windows-2016 with R release/oldrel
