+++ .github/workflows/main.yml
@@ -35,8 +35,8 @@ jobs:
           - { os: ubuntu-latest, r: "release", pkgdown: "true"} # ubuntu-18.04
           - { os: ubuntu-latest, r: "oldrel" }
           # - { os: ubuntu-16.04, r: "devel" } # fails setup-r
-          # - { os: ubuntu-16.04, r: "release"} # fails setup-r
-          # - { os: ubuntu-16.04, r: "oldrel" } # fails setup-r
+          - { os: ubuntu-16.04, r: "release"}
+          - { os: ubuntu-16.04, r: "oldrel" }
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
