### 2020-06-16 6d7f8b5
Support ODK v0.7 geotrace/shape output
* Add odkc_version to split_geo{trace, shape}, handle_geo(trace, shape}, take from odata_submission_get(
* Parse ODK linestrings depending on odkc_version
* Fix #70
