### 2020-06-16 52d2a8e
GHA: install r-cran-stringr for ubuntu-16.04
