### 2020-06-17 fc6b86b
GHA: ubuntu016.04 needs libicu-dev for stringr
