+++ .github/workflows/main.yml
@@ -125,11 +125,11 @@ jobs:
           restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-ccache-${{steps.datew.outputs.datew}}
 
       # install ccache and write config file
-      # install r-cran-stringr for ubuntu-16.04
+      # install libicu-dev for stringr on ubuntu-16.04
       - name: "[Linux] ccache"
         if: runner.os == 'Linux'
         run: |
-          sudo apt install ccache libcurl4-openssl-dev r-cran-stringr
+          sudo apt install ccache libcurl4-openssl-dev libicu-dev
           mkdir -p ~/.R && echo -e 'CC=ccache gcc -std=gnu99\nCXX=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
 
       # install ccache and write config file
