### 2020-06-17 804b60b
Attempt to fix GHA build error for xml2 on ubuntu-16.04
Add libxml2 system package as per https://github.com/r-lib/xml2/issues/250#issuecomment-490610400
