### 2020-06-17 f39296d
GHA: ubuntu-16.04 needs libxml2-dev headers for xml2
