### 2020-06-17 c4884f6
GHA: ubuntu-16.04 fails to install xml2, add GH PAT
