+++ .github/workflows/main.yml
@@ -35,8 +35,8 @@ jobs:
           - { os: ubuntu-latest, r: "release", pkgdown: "true"} # ubuntu-18.04
           - { os: ubuntu-latest, r: "oldrel" }
           # - { os: ubuntu-16.04, r: "devel" } # fails setup-r
-          - { os: ubuntu-16.04, r: "release"}
-          - { os: ubuntu-16.04, r: "oldrel" }
+          # - { os: ubuntu-16.04, r: "release"} # fails to install xml2
+          # - { os: ubuntu-16.04, r: "oldrel" } # fails to install xml2
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
@@ -56,6 +56,7 @@ jobs:
       BUILD_PKGDOWN: ${{ matrix.config.pkgdown }}
       # macOS >= 10.15.4 linking
       SDKROOT: /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       # ruODK test env variables
       ODKC_URL: ${{ secrets.ODKC_URL }}
       ODKC_UN: ${{ secrets.ODKC_UN }}
