### 2020-06-17 6ecb1a7
GHA: drop system deps for ubuntu-16.04
Suspected cause of segfault at https://github.com/dbca-wa/ruODK/runs/778881416
