### 2020-06-17 d84ba73
Update vignette odata-api: disable form group prefix for submission data columns
names_set=NULL works, names_sep=None fails.
HT @OdiljonQurbon
