### 2020-06-18 32256db
Updated examples for form_detail
Fix #74
Thanks to @dmenne
