### 2020-06-21 95d5875
GHA: address windows-oldrel failure, test ubuntu envs
* Windows (2019 and 2016) with R oldrel fail at stage "script" on missing R pkg backports
* Re-enable previously failing envs for a soundcheck
