### 2020-06-21 11dcf9e
Make GHA happy again
Sometimes the solution to more happiness is to shoot unhappy build environments
