### 2020-06-21 caf4a82
GHA: drop failing envs
* ubuntu-latest (bionic) R release segfaults on tic install (worked before, no changes to workflow) - that's a GHA / tic bug, not a fault in ruODK.
* ubuntu 16.04 (xenial) R oldrel fails on stringr / stringi / libicui18n.so.60
