+++ .github/workflows/main.yml
@@ -26,17 +26,17 @@ jobs:
           - { os: windows-2016, r: "release" }
           # - { os: windows-2016, r: "oldrel" }
           - { os: macOS-latest, r: "devel" } # macos-10.15
-          - { os: macOS-latest, r: "release"}
+          - { os: macOS-latest, r: "release", pkgdown: "true"}
           - { os: macOS-latest, r: "oldrel"}
           # - { os: ubuntu-20.04, r: "devel" } # # not released yet
           # - { os: ubuntu-20.04, r: "release"} # not released yet
           # - { os: ubuntu-20.04, r: "oldrel" } # not released yet
           # - { os: ubuntu-latest, r: "devel" } # fails setup-r
-          - { os: ubuntu-latest, r: "release", pkgdown: "true"} # ubuntu-18.04
+          # - { os: ubuntu-latest, r: "release"} # ubuntu-18.04 # segfaults on tic install
           - { os: ubuntu-latest, r: "oldrel" }
           # - { os: ubuntu-16.04, r: "devel" } # fails setup-r
           - { os: ubuntu-16.04, r: "release"}
-          - { os: ubuntu-16.04, r: "oldrel" }
+          # - { os: ubuntu-16.04, r: "oldrel" } # fails on stringi / libicui-dev
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
