### 2020-06-24 3aa5688
GHA soundcheck: enable failing envs
