### 2020-06-24 2567c82
GHA soundcheck: drop failing, welcome back ubuntu xenial and bionic
