### 2020-06-24 #335 @florianm
Meanwhile, GHA has a [heisen-segfault](https://github.com/ropensci/tic/issues/271) in ubuntu-latest / R devel and R release. I'm monitoring this over the next few days hoping it will resolve, so the red GHA badge is deliberate.
