### 2020-06-24 5b3d475
Add note about #79 to news.
