### 2020-06-24 3638996
Add note about #79 to news.
