### 2020-06-24 17d5a69
Add tests for split_geoshape and handle_ru_geo*
Fix #78
