### 2020-06-24 2503e08
Add data and tests for form_schema_parse, update docs
Fix #76
HT @dmenne
