+++ tests/testthat/test-form_schema_parse.R
@@ -45,25 +45,19 @@ test_that("form_schema works with ODK Central v0.8", {
 })
 
 # TODO run this test against ODK Central v0.7
-# test_that("form_schema_parse debug messages work", {
-#   fs <- form_schema(
-#     flatten = FALSE,
-#     parse = FALSE,
-#     pid = get_test_pid(),
-#     fid = get_test_fid(),
-#     url = get_test_url(),
-#     un = get_test_un(),
-#     pw = get_test_pw(),
-#     odkc_version = 0.8
-#   )
-#
-#   testthat::capture_output(
-#     testthat::expect_message(
-#       fsp <- form_schema_parse(fs, verbose = TRUE)
-#     )
-#   )
-#
-#   testthat::expect_equal(class(fsp), c("tbl_df", "tbl", "data.frame"))
-#   testthat::expect_true("encounter_start_datetime" %in% fsp$name)
-#   testthat::expect_true("quadrat_photo" %in% fsp$name)
-# })
+test_that("form_schema_parse debug messages work", {
+  data("fs_v7_raw")
+  data("fs_v7")
+
+  testthat::capture_output(
+    testthat::expect_message(
+      fs_v7_parsed <- form_schema_parse(fs_v7_raw, verbose = TRUE)
+    )
+  )
+
+  testthat::expect_equal(class(fs_v7_parsed), c("tbl_df", "tbl", "data.frame"))
+  testthat::expect_true("observation_start_time" %in% fs_v7_parsed$name)
+  testthat::expect_true("disturbance_cause" %in% fs_v7_parsed$name)
+
+  testthat::expect_equal(fs_v7, fs_v7_parsed)
+})
