### 2020-06-24 a680938
GHA: drop heisenbug-segfaulting envs
