+++ man/fq_form_xml.Rd
@@ -40,6 +40,8 @@ Other included:
 \code{\link{fq_zip_data}},
 \code{\link{fq_zip_strata}},
 \code{\link{fq_zip_taxa}},
+\code{\link{fs_v7_raw}},
+\code{\link{fs_v7}},
 \code{\link{geo_fs}},
 \code{\link{geo_gj_raw}},
 \code{\link{geo_gj}},
