+++ man/fs_v7.Rd
@@ -2,19 +2,17 @@
 % Please edit documentation in R/data.R
 \docType{data}
 \encoding{UTF-8}
-\name{fq_form_detail}
-\alias{fq_form_detail}
-\title{A tibble of form metadata.}
+\name{fs_v7}
+\alias{fs_v7}
+\title{The parsed XML form_schema of a form from ODK Central v0.6.}
 \format{
-A tibble of form metadata.
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.
 }
 \source{
-The output of \code{\link{form_detail}}
-run on submissions of the test form
-`system.file("extdata", "FloraQuadrat04.xml", package = "ruODK")`.
+\code{\link{form_schema_parse}(fs_v7_raw)}
 }
 \usage{
-fq_form_detail
+fs_v7
 }
 \description{
 \lifecycle{stable}
@@ -25,6 +23,7 @@ Other included:
 \code{\link{fq_data_strata}},
 \code{\link{fq_data_taxa}},
 \code{\link{fq_data}},
+\code{\link{fq_form_detail}},
 \code{\link{fq_form_list}},
 \code{\link{fq_form_schema}},
 \code{\link{fq_form_xml}},
@@ -41,7 +40,6 @@ Other included:
 \code{\link{fq_zip_strata}},
 \code{\link{fq_zip_taxa}},
 \code{\link{fs_v7_raw}},
-\code{\link{fs_v7}},
 \code{\link{geo_fs}},
 \code{\link{geo_gj_raw}},
 \code{\link{geo_gj}},

[…]
