### 2020-06-25 305db7e
GHA soundcheck - spurious ubuntu 16.04 fail at setup R
