### 2020-06-25 de9bb2a
covr: exclude untestable code
* form_schema supports ODK Central versions below 0.8
* ruODK CI and reviewers have no reliable access to ODK Central < 0.8
* form_schema calls form_schema_parse, which is fully tested with included example data
* excluded code downloads XML, calls form_schema_parse, and emits messages
