### 2020-06-25 bb1a19e
GHA: Drop options(Ncpus=4)
