### 2020-06-25 ff60adf
GHA: bring back options(Ncpus=4)
