### 2020-06-25 368e32f
Fix markup in docs for odata_submission_get
