### 2020-07-06 5a54171
Rename master to main branch, add libxml2-dev to Linux runners
* Address build failure on ubuntu (all) R oldrel: rlang > XML > libxml
