+++ .github/workflows/main.yml
@@ -1,7 +1,7 @@
 on:
   push:
   pull_request:
-  # for now, CRON jobs only run on the default branch of the repo (i.e. usually on master)
+  # for now, CRON jobs only run on the default branch of the repo (i.e. usually on main)
   schedule:
     # * is a special character in YAML so you have to quote this string
     - cron: "0 4 * * *"
@@ -131,7 +131,7 @@ jobs:
       - name: "[Linux] ccache"
         if: runner.os == 'Linux'
         run: |
-          sudo apt install ccache libcurl4-openssl-dev
+          sudo apt install ccache libcurl4-openssl-dev libxml2-dev
           mkdir -p ~/.R && echo -e 'CC=ccache gcc -std=gnu99\nCXX=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
 
       # install ccache and write config file
