### 2020-07-08 5e73139
Fix #81: update docs to install from main
* Fix in README and install.R (for urODK)
* Bump patch version
* Add mention to NEWS
