+++ README.md
@@ -129,11 +129,11 @@ Out of scope:
 
 ## Install
 
-You can install `ruODK` from GitHub with:
+You can install the development version (`main` branch) of `ruODK` with:
 
 ``` r
 if (!requireNamespace("remotes")) {install.packages("remotes")}
-remotes::install_github("dbca-wa/ruODK", dependencies = TRUE)
+remotes::install_github("dbca-wa/ruODK@main", dependencies = TRUE)
 ```
 
 ## ODK Central
@@ -224,7 +224,7 @@ proj %>% head() %>% knitr::kable(.)
 | -: | :-------------------- | ----: | ---------: | :------------------ | :---------- | :------------------ | :------- |
 | 59 | 1Horizon test project |     1 |          4 | 2020-01-13 20:45:32 | NA          | 2020-01-23 22:20:13 | FALSE    |
 | 57 | AB’s                  |     1 |          1 | 2020-01-06 08:56:00 | NA          | NA                  | FALSE    |
-|  8 | AmenazasRD            |     1 |          1 | 2019-04-14 14:21:00 | NA          | 2020-04-21 14:23:06 | FALSE    |
+|  8 | AmenazasRD            |     1 |          2 | 2019-04-14 14:21:00 | NA          | 2020-04-21 14:23:06 | FALSE    |
 | 88 | Amit-wfp-test         |     3 |          1 | 2020-06-12 10:09:57 | NA          | 2020-06-12 13:03:00 | FALSE    |
 |  9 | BAFCO Test            |     1 |          1 | 2019-04-19 08:30:45 | NA          | 2019-04-19 09:26:49 | FALSE    |
 | 79 | Birhan Project        |     1 |          1 | 2020-05-11 20:58:06 | NA          | 2020-05-12 00:09:16 | FALSE    |
@@ -323,8 +323,6 @@ data <- ruODK::odata_submission_get(local_dir = fs::path("vignettes/media"),
 #> ℹ Reading form schema...
 #> ℹ Form schema v0.8
 #> ℹ Parsing submissions...
-#> New names:
-#> * `@odata.context` -> .odata.context
 #> ℹ Not unnesting geo fields: value_location_corner1, value_perimeter_corner2, value_perimeter_corner3, value_perimeter_corner4, value_taxon_encounter_taxon_encounter_location
 #> ℹ Unnesting: value
 #> ℹ Unnesting column "value"
@@ -359,10 +357,9 @@ data <- ruODK::odata_submission_get(local_dir = fs::path("vignettes/media"),
 data %>% knitr::kable(.)
 ```
 
[…]
