### 2020-07-13 #335 @maelle
:wave: here! I see @jmt2080ad checked "The author has responded to my review and made changes to my satisfaction. I recommend approving this package."

@karissawhiting do you have further comments for @florianm? If not could you please update this thread or check the box in [your review](https://github.com/ropensci/software-review/issues/335#issuecomment-541438418)? Thank you!
