### 2020-07-13 d4f9159
GHA: drop ubuntu-{18,16}.04/R oldrel (rlist not avail, missing XML pkg)
