### 2020-07-13 82c58fd
GHA: drop failing envs
