### 2020-07-14 2ff8bbb
Add Dockerfile
* Based on rocker/geospatial:4.0.1
* ruODK and other packages used by DBCA
* Can access MS Access via mdbtools / DBI
* #83 likely incompatible with binder
