### 2020-07-16 #335 @maelle
Approved! Thanks @florianm for submitting and @karissawhiting @jmt2080ad  for your reviews! :turtle: :sparkles:

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`

Should you want to acknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field (with their consent).  [More info on this](https://devguide.ropensci.org/building.html#authorship).
