### 2020-07-16 #335 @maelle
@jmt2080ad Could you please add a rough estimate of the time you spent reviewing in [your review](https://github.com/ropensci/software-review/issues/335#issuecomment-542019403) near `Estimated hours spent reviewing:`? Thank you.
