### 2020-07-17 369e006
Transfer ownership to rOpenSci
* Rename links from dbca-wa to ropensci
* CI and docs likely broken for now and will be fixed next
* TODO: docs swapperoo to docs.ropensci.org
* Following instructions at https://github.com/ropensci/software-review/issues/335#issuecomment-659174910
