+++ inst/joss/paper.bib
@@ -3,7 +3,7 @@
   author = {Florian W. Mayer},
   note = {R package version 0.3.1},
   year = {2019},
-  url = {https://github.com/dbca-wa/ruODK},
+  url = {https://github.com/ropensci/ruODK},
 }
 
 @online{github,
