+++ tests/testthat/test-attachment_get.R
@@ -66,7 +66,7 @@ test_that("attachment_url works", {
   expected_url <- glue::glue(
     "{url}/v1/projects/14/forms/{fid}/submissions/{uuid}/attachments/{fn}"
   )
-  # https://github.com/dbca-wa/ruODK/issues/66
+  # https://github.com/ropensci/ruODK/issues/66
   # Not using here because it breaks attachment_get
   # expected_url <- httr::modify_url(
   #   url,
