+++ tests/testthat/test-split_geotrace.R
@@ -128,8 +128,8 @@ test_that("split_geotrace works with ODK Linestrings", {
   # ruODK::odata_submission_rectangle()
   # First three points shown with truncated decimal places
   odk_v7 <- tibble::tibble(
-      id = 1,
-      tx = "-14.80 128.40 10.9 5.9;-14.81 128.41 1.7 1.9;-14.82 128.42 1.9 1.7;"
+    id = 1,
+    tx = "-14.80 128.40 10.9 5.9;-14.81 128.41 1.7 1.9;-14.82 128.42 1.9 1.7;"
   )
   odk_v7_split <- split_geotrace(odk_v7, "tx", odkc_version = 0.7)
 
