### 2020-07-20 ropensci/ruODK#84 @florianm
GitHubActions Windows build fails with warning
## Problem
GHA fails its Windows builds with a warning relating to vignettes.

As we build documentation with pkgdown from the Ubuntu build (about to be changed to rOpenSci's own doc build&deployment method), Windows builds should not be concerned at all with vignettes.

I'd like to solve this issue in order to restore the green "tic" badge for https://github.com/ropensci/software-review/issues/335.

```
* checking files in 'vignettes' ... WARNING
Files in the 'vignettes' directory but no files in 'inst/doc':
  'media/1568786958640.jpg', 'media/1568787004467.jpg',
    'media/1568787172983.jpg', 'media/1568787220913.jpg',
    'media/1568787333689.jpg', 'media/1568787380833.jpg',
    'media/1568787456664.jpg', 'media/1568787504877.jpg',
    'media/1568794395624.jpg', 'media/1568794560256.jpg',
    'media/1568795382633.jpg', 'media/1568795579890.jpg',
    'odata-api.Rmd', 'restful-api.Rmd', 'setup.Rmd'
* checking examples ... OK
* checking for unstated dependencies in 'tests' ... OK
* checking tests ... OK
  Running 'testthat.R' [91s]
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes in 'inst/doc' ... WARNING
dir.exists(dir) is not TRUEdir.exists(dir) is not TRUEdir.exists(dir) is not TRUE
Package vignettes without corresponding single PDF/HTML:
   'odata-api.Rmd'
   'restful-api.Rmd'
   'setup.Rmd'

```
## Reproducible example
Failing builds with error messages e.g. https://github.com/ropensci/ruODK/runs/886204228?check_suite_focus=true
