### 2020-07-20 8043d73
#84: attempt to fix Windows build
* Try `%\VignetteEngine{knitr::knitr}` in each vignette as per https://stackoverflow.com/a/54337029/2813717
