+++ vignettes/setup.Rmd
@@ -3,7 +3,7 @@ title: "Setup"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{setup}
-  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEngine{knitr::knitr}
   %\VignetteEncoding{UTF-8}  
 ---
 
