### 2020-07-20 df7132f
#84: Add webshot dependency to resolve vignette build error, enable previously failing build envs
