+++ README.Rmd
@@ -16,10 +16,10 @@ knitr::opts_chunk$set(
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Last-changedate](https://img.shields.io/github/last-commit/ropensci/ruODK.svg)](https://github.com/ropensci/ruODK/commits/master)
 [![GitHub issues](https://img.shields.io/github/issues/ropensci/ruodk.svg?style=popout)](https://github.com/ropensci/ruODK/issues/)
-[![CI via {tic}](https://github.com/ropensci/ruODK/workflows/R%20CMD%20Check%20via%20{tic}/badge.svg?branch=master)](https://github.com/ropensci/ruODK/actions)
-[![Build Status](https://travis-ci.org/ropensci/ruODK.svg?branch=master)](https://travis-ci.org/ropensci/ruODK)
-[![Build status](https://ci.appveyor.com/api/projects/status/1cs19xx0t64bmd2q/branch/master?svg=true)](https://ci.appveyor.com/project/florianm/ruodk/branch/master)
-[![Coverage status](https://codecov.io/gh/ropensci/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/ruODK?branch=master)
+[![CI via {tic}](https://github.com/ropensci/ruODK/workflows/R%20CMD%20Check%20via%20{tic}/badge.svg?branch=main)](https://github.com/ropensci/ruODK/actions)
+[![Build Status](https://travis-ci.org/ropensci/ruODK.svg?branch=main)](https://travis-ci.org/ropensci/ruODK)
+[![Build status](https://ci.appveyor.com/api/projects/status/1cs19xx0t64bmd2q/branch/master?svg=true)](https://ci.appveyor.com/project/florianm/ruodk/branch/main)
+[![Coverage status](https://codecov.io/gh/ropensci/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/ruODK?branch=main)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 <!-- badges: end -->
 
