+++ .github/workflows/main.yml
@@ -27,17 +27,17 @@ jobs:
           - { os: windows-2016, r: "release" }
           # - { os: windows-2016, r: "oldrel" } # fails on missing RTools for 3.6.3
           - { os: macOS-latest, r: "devel" } # macos-10.15
-          - { os: macOS-latest, r: "release", pkgdown: "true"}
+          - { os: macOS-latest, r: "release", pkgdown: "false"} # >> docs.ropensci.org
           - { os: macOS-latest, r: "oldrel"}
           # - { os: ubuntu-20.04, r: "devel" } # # not released yet
           # - { os: ubuntu-20.04, r: "release"} # not released yet
           # - { os: ubuntu-20.04, r: "oldrel" } # not released yet
-          # - { os: ubuntu-latest, r: "devel" } # fails setup-r
+          - { os: ubuntu-latest, r: "devel" } # fails setup-r
           - { os: ubuntu-latest, r: "release"} # ubuntu-18.04
-          # - { os: ubuntu-latest, r: "oldrel" } # rlist not avail > XML
-          # - { os: ubuntu-16.04, r: "devel" } # fails setup-r
+          - { os: ubuntu-latest, r: "oldrel" } # rlist not avail > XML
+          - { os: ubuntu-16.04, r: "devel" } # fails setup-r
           - { os: ubuntu-16.04, r: "release"}
-          # - { os: ubuntu-16.04, r: "oldrel" } # rlist not avail > XML
+          - { os: ubuntu-16.04, r: "oldrel" } # rlist not avail > XML
     env:
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
