+++ DESCRIPTION
@@ -66,7 +66,8 @@ Suggests:
     roxygen2 (>= 7.1.0),
     testthat (>= 2.3.2),
     usethis (>= 1.6.0),
-    vcr (>= 0.5.4)
+    vcr (>= 0.5.4),
+    webshot (>= 0.5.2)
 VignetteBuilder: 
     knitr
 RdMacros: 
