### 2020-07-20 850ed50
#84 Add empty inst/doc/ folder
* GHA Windows build fails on missing inst/doc/
