+++ .github/workflows/main.yml
@@ -32,10 +32,10 @@ jobs:
           # - { os: ubuntu-20.04, r: "devel" } # # not released yet
           # - { os: ubuntu-20.04, r: "release"} # not released yet
           # - { os: ubuntu-20.04, r: "oldrel" } # not released yet
-          - { os: ubuntu-latest, r: "devel" } # fails setup-r
+          # - { os: ubuntu-latest, r: "devel" } # fails setup-r
           - { os: ubuntu-latest, r: "release"} # ubuntu-18.04
           - { os: ubuntu-latest, r: "oldrel" } # rlist not avail > XML
-          - { os: ubuntu-16.04, r: "devel" } # fails setup-r
+          # - { os: ubuntu-16.04, r: "devel" } # fails setup-r
           - { os: ubuntu-16.04, r: "release"}
           - { os: ubuntu-16.04, r: "oldrel" } # rlist not avail > XML
     env:
