### 2020-07-20 11eeba1
Vignette engine: use knitr::rmarkdown
