+++ vignettes/setup.Rmd
@@ -3,7 +3,7 @@ title: "Setup"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{setup}
-  %\VignetteEngine{knitr::knitr}
+  %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}  
 ---
 
