+++ vignettes/odata-api.Rmd
@@ -3,7 +3,7 @@ title: "Accessing the OData API"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{odata}
-  %\VignetteEngine{knitr::knitr}
+  %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 resource_files:
 - media/1568786958640.jpg
