### 2020-07-20 ropensci/ruODK#84 @florianm
Windows build runs with options `* using options '--no-manual --no-vignettes --no-build-vignettes --as-cran'`, then complains about missing vignettes in inst/doc.
