### 2020-07-21 d8f1657
Drop rlist/XML dependency
* Close #85
* Refactor fns to use tidyr/dplyr
* Remove dep from DESC, bump version
* Re-enable prev failing GHA builds
* Clean up vignette restful-api
* Update CITATION year
* Update packages data
