+++ vignettes/restful-api.Rmd
@@ -328,48 +328,13 @@ as e.g. JSON.
 listviewer::jsonedit(fq_submissions, mode = "code")
 ```
 
-## Rectangling and parsing the submissions
-For a more traditional analysis, we need the data in a rectangular shape.
-The unpredictably nested structure of the submission data (when containing 
-repeated subgroups) makes it challenging to do this in an automated fashion.
-
-While both the ZIP export and the OData API provide clean, rectangular versions
-of the data, we'll still attempt here to un-nest the nested list of submission 
-data into a rectangular `tibble`.
-
-In the example shown above, the repeated `taxon_encounter` groups are named
-`taxon_encounter`, `taxon_encounter.1` and so on. This is an artifact of the 
-`jsonedit::listviewer()` widget.
-
-With one additional step we can flatten the nested lists and bind them into one
-dataframe, then sanitise the column names, handle datetimes, geopoints and 
-attachments, like we did with the CSV export.
-
-It appears this operation loses all but the last repeat of repeating groups.
-
-<!-- TODO update and eval=T after #69 -->
-
-```{r parse_submissions, eval=F}
-sub_parsed <- fq_submissions %>% 
-  purrr::map_df(rlist::list.flatten) %>%
-  janitor::clean_names(.) %>%
-  dplyr::mutate(id = meta_instance_id) %>%
-  ruODK::handle_ru_datetimes(fq_form_schema) %>%
-  ruODK::handle_ru_geopoints(fq_form_schema) %>%
-  ruODK::handle_ru_attachments(fq_form_schema, local_dir = t)
-
-names(sub_parsed)
-head(sub_parsed)
-
-# List attachment files
-fs::dir_ls(t, recurse = TRUE, glob = "*.jpg")
-```
-
 # Outlook
+The approach shown here yields nested and stand-alone records, which is useful
+if the subsequent use requires records in nested JSON or XML format. 
+Complex forms with repeating sub-groups will result in highly nested lists, 
+whose structure heavily depends on the completeness of the submissions.
 
-From here, we can use the separately created data tibbles and locally downloaded
-attachments to build visualisations as shown in the 
-[`vignette("odata-api", package="ruODK")`](https://ropensci.github.io/ruODK/articles/odata-api.html).
-
-The best pathway to retrieve submission data seems to be the data export to ZIP 
-or access via OData.
+The other approach shown in 
+[`vignette("odata-api", package="ruODK")`](https://ropensci.github.io/ruODK/articles/odata-api.html) 
+yields rectangled data in several normalized tables, which is useful for 
+analysis and visualisation.
