### 2020-07-21 200b34c
#84 Attempt to prevent Windows builds failing on WARNINGS
* Add error_on to `tic::do_package_checks(error_on="error")`
* The WARNINGS we aim to suppress relate to documentation, which is not built from Windows
