### 2020-07-21 828a070
GHA: drop failing env
* macos-latest R-devel fails at r-lib/actions/setup-r@master
* ubuntu-latest fails with Heisensegfault (will resolve when re-run at correct moon phase)
