+++ .github/workflows/main.yml
@@ -26,7 +26,7 @@ jobs:
           - { os: windows-2016, r: "devel" }
           - { os: windows-2016, r: "release" }
           # - { os: windows-2016, r: "oldrel" } # fails on missing RTools for 3.6.3
-          - { os: macOS-latest, r: "devel" } # macos-10.15
+          # - { os: macOS-latest, r: "devel" } # macos-10.15 # fails on setup-r
           - { os: macOS-latest, r: "release", pkgdown: "false"} # see docs.ropensci.org
           - { os: macOS-latest, r: "oldrel"}
           # - { os: ubuntu-20.04, r: "devel" } # # not released yet
