### 2020-07-21 ropensci/ruODK#84 @florianm
Do not fail on WARNING: `tic::do_package_checks(error_on="error")`.
