### 2020-07-21 e78b89f
Tic tuning: build vignettes and manual #84
