### 2020-07-21 595bea3
GHA: drop heisensegfaulting ubuntu-latest envs
