### 2020-07-21 8343ad1
Update README and citation with DOI
