+++ README.Rmd
@@ -13,13 +13,14 @@ knitr::opts_chunk$set(
 # `ruODK`: An R Client for the ODK Central API <img src="man/figures/ruODK.png" align="right" alt="Especially in these trying times, it is important to ask: ruODK?" width="120" />
 <!-- badges: start -->
 [![ROpenSci submission status](https://badges.ropensci.org/335_status.svg)](https://github.com/ropensci/software-review/issues/335)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3953159.svg)](https://doi.org/10.5281/zenodo.3953159)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Last-changedate](https://img.shields.io/github/last-commit/ropensci/ruODK.svg)](https://github.com/ropensci/ruODK/commits/master)
 [![GitHub issues](https://img.shields.io/github/issues/ropensci/ruodk.svg?style=popout)](https://github.com/ropensci/ruODK/issues/)
 [![CI via {tic}](https://github.com/ropensci/ruODK/workflows/R%20CMD%20Check%20via%20{tic}/badge.svg?branch=main)](https://github.com/ropensci/ruODK/actions)
 [![Build Status](https://travis-ci.org/ropensci/ruODK.svg?branch=main)](https://travis-ci.org/ropensci/ruODK)
 [![Build status](https://ci.appveyor.com/api/projects/status/1cs19xx0t64bmd2q/branch/master?svg=true)](https://ci.appveyor.com/project/florianm/ruodk/branch/main)
-[![Coverage status](https://codecov.io/gh/ropensci/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/ruODK?branch=main)
+[![codecov](https://codecov.io/gh/ropensci/ruODK/branch/main/graph/badge.svg)](https://codecov.io/gh/ropensci/ruODK)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 <!-- badges: end -->
 
@@ -209,7 +210,7 @@ srv %>% knitr::kable(.)
 # Form submissions
 data <- ruODK::odata_submission_get(local_dir = fs::path("vignettes/media"),
                                     odkc_version = get_test_odkc_version())
-data %>% knitr::kable(.)
+data %>% dplyr::select(-"odata_context") %>% knitr::kable(.)
 ```
 
 A more detailed walk-through with some data visualisation examples is available
@@ -267,13 +268,13 @@ There are several other R packages interacting with the ODK ecosystem, and/or
 |------------------------------|---------------|---------------|---------------|---------------|---------------|---------------|
 | Elevator pitch               | "[ckanr](https://github.com/ropensci/ckanr) for ODK Central"  | "Drive ODK Briefcase through R" | "Export ODK Aggregate to SPSS" | "Export ODK Aggregate to STATA" | "Metapackage for the extended ODK ecosystem" | "What ruODK does, but as GUI" |
[…]
