+++ README.md
@@ -7,6 +7,7 @@
 
 [![ROpenSci submission
 status](https://badges.ropensci.org/335_status.svg)](https://github.com/ropensci/software-review/issues/335)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3953159.svg)](https://doi.org/10.5281/zenodo.3953159)
 [![Project Status: Active – The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -19,8 +20,7 @@ issues](https://img.shields.io/github/issues/ropensci/ruodk.svg?style=popout)](h
 Status](https://travis-ci.org/ropensci/ruODK.svg?branch=main)](https://travis-ci.org/ropensci/ruODK)
 [![Build
 status](https://ci.appveyor.com/api/projects/status/1cs19xx0t64bmd2q/branch/master?svg=true)](https://ci.appveyor.com/project/florianm/ruodk/branch/main)
-[![Coverage
-status](https://codecov.io/gh/ropensci/ruODK/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/ruODK?branch=main)
+[![codecov](https://codecov.io/gh/ropensci/ruODK/branch/main/graph/badge.svg)](https://codecov.io/gh/ropensci/ruODK)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/dbca-wa/urODK/master?urlpath=rstudio)
 <!-- badges: end -->
 
@@ -265,41 +265,41 @@ meta <- ruODK::form_schema(odkc_version = get_test_odkc_version())
 meta %>% knitr::kable(.)
 ```
 
-| name                                  | type      | path                                                       | ruodk\_name                                                |
-| :------------------------------------ | :-------- | :--------------------------------------------------------- | :--------------------------------------------------------- |
-| meta                                  | structure | /meta                                                      | meta                                                       |
-| instanceID                            | string    | /meta/instanceID                                           | meta\_instanceID                                           |
-| encounter\_start\_datetime            | dateTime  | /encounter\_start\_datetime                                | encounter\_start\_datetime                                 |
-| reporter                              | string    | /reporter                                                  | reporter                                                   |
-| device\_id                            | string    | /device\_id                                                | device\_id                                                 |
-| location                              | structure | /location                                                  | location                                                   |
-| area\_name                            | string    | /location/area\_name                                       | location\_area\_name                                       |
-| quadrat\_photo                        | binary    | /location/quadrat\_photo                                   | location\_quadrat\_photo                                   |
-| corner1                               | geopoint  | /location/corner1                                          | location\_corner1                                          |
-| habitat                               | structure | /habitat                                                   | habitat                                                    |
-| morphological\_type                   | select1   | /habitat/morphological\_type                               | habitat\_morphological\_type                               |
-| morphological\_type\_photo            | binary    | /habitat/morphological\_type\_photo                        | habitat\_morphological\_type\_photo                        |
-| vegetation\_stratum                   | repeat    | /vegetation\_stratum                                       | vegetation\_stratum                                        |
-| nvis\_level3\_broad\_floristic\_group | select1   | /vegetation\_stratum/nvis\_level3\_broad\_floristic\_group | vegetation\_stratum\_nvis\_level3\_broad\_floristic\_group |
-| max\_height\_m                        | decimal   | /vegetation\_stratum/max\_height\_m                        | vegetation\_stratum\_max\_height\_m                        |
[…]
