+++ data-raw/make_data.R
@@ -20,12 +20,15 @@ fq_raw_strata <- ruODK::odata_submission_get(table = fq_svc$name[2], parse = FAL
 fq_raw_taxa <- ruODK::odata_submission_get(table = fq_svc$name[3], parse = FALSE)
 
 fq_data <- ruODK::odata_submission_get(
-  table = fq_svc$name[1], wkt = TRUE, parse = T, odkc_version = get_test_odkc_version())
+  table = fq_svc$name[1], wkt = TRUE, parse = T, odkc_version = get_test_odkc_version()
+)
 fq_data_strata <- ruODK::odata_submission_get(
-  table = fq_svc$name[2], wkt = TRUE, odkc_version = get_test_odkc_version()) %>%
+  table = fq_svc$name[2], wkt = TRUE, odkc_version = get_test_odkc_version()
+) %>%
   dplyr::left_join(fq_data, by = c("submissions_id" = "id"))
 fq_data_taxa <- ruODK::odata_submission_get(
-  table = fq_svc$name[3], wkt = TRUE, odkc_version = get_test_odkc_version()) %>%
+  table = fq_svc$name[3], wkt = TRUE, odkc_version = get_test_odkc_version()
+) %>%
   dplyr::left_join(fq_data, by = c("submissions_id" = "id"))
 
 # Geofields
