### 2020-07-21 #335 @florianm
@maelle I've completed the TODO list and followed the "life after onboarding" chapter. GH actions are passing, codemeta and all docs are updated, and I've minted a DOI at Zenodo. Did I do this roughly in the correct order?

All, would you mind getting mentioned in the Zenodo record?
