+++ R/split_geoshape.R
@@ -68,7 +68,12 @@
 #'   is.numeric(wkt_first_gt$shape_location_shape_gps_altitude)
 #' )
 #' }
-split_geoshape <- function(data, colname, wkt = FALSE, odkc_version = odkc_version) {
+split_geoshape <- function(
+  data,
+  colname,
+  wkt = FALSE,
+  odkc_version = odkc_version
+) {
   if (nrow(data) == 0) {
     # Option 1: Early exit - nothing to do
     return(data)
