### 2020-07-21 40d1736
Lint package
* Add # nolint to cross-references between code and corresponding tests
* Add # nolint to opening curly braces on newlines as per styler auto-styling
* Enclose commented out code in # nolint start/end where the code adds context for further development or debugging
