+++ data-raw/make_release.R
@@ -19,6 +19,7 @@ source(here::here("data-raw/make_data.R"))
 devtools::test()
 #
 # Docs
+lintr:::addin_lint_package()
 styler::style_pkg()
 devtools::document(roclets = c("rd", "collate", "namespace"))
 spelling::spell_check_package()
