+++ tests/testthat/test-submission_detail.R
@@ -24,12 +24,11 @@ test_that("submission_detail works", {
   testthat::expect_equal(nrow(sub), 1)
 
   # The columns are metadata, plus the submission data in column 'xml`
-  # names(sub)
+  # names(sub) # nolint
   cn <- c(
     "instance_id", "submitter_id", "submitter", "created_at", "updated_at"
   )
   testthat::expect_equal(names(sub), cn)
 })
 
-# Tests code
-# usethis::edit_file("R/submission_detail.R")
+# usethis::edit_file("R/submission_detail.R") # nolint
