### 2020-07-21 de3c0b8
GHA testing: disable cache, enable segfaulting envs
