### 2020-07-21 967fe52
Update JOSS paper - not meant for immediate publication
