+++ inst/joss/paper.bib
@@ -1,7 +1,7 @@
 @Misc{ruodk,
   title = {ruODK: Client for the ODK Central API},
   author = {Florian W. Mayer},
-  note = {R package version 0.3.1},
+  note = {R package version 0.9.0},
   year = {2019},
   url = {https://github.com/ropensci/ruODK},
 }
@@ -16,17 +16,17 @@
 
 @online{odk,
   author = {OpenDataKit},
-  title = {opendatakit.org},
+  title = {getodk.org},
   year = 2019,
-  url = {https://opendatakit.org/},
+  url = {https://getodk.org/},
   urldate = {2019-07-25}
 }
 
 @online{odkdocs,
   author = {OpenDataKit Documentation},
-  title = {docs.opendatakit.org},
+  title = {docs.getodk.org},
   year = 2019,
-  url = {https://docs.opendatakit.org/},
+  url = {https://docs.getodk.org/},
   urldate = {2019-07-25}
 }
 
