+++ .github/workflows/main.yml
@@ -102,16 +102,16 @@ jobs:
       # - cache ccache weekly -> 'ccache' helps rebuilding the package cache faster
       #
       # Heisensegfault testing: disable cache
-      # - name: "[Cache] Prepare daily timestamp for cache"
-      #   if: runner.os != 'Windows'
-      #   id: date
-      #   run: echo "::set-output name=date::$(date '+%d-%m')"
-      #
-      # - name: "[Cache] Prepare weekly timestamp for cache"
-      #   if: runner.os != 'Windows'
-      #   id: datew
-      #   run: echo "::set-output name=datew::$(date '+%Y-%V')"
-      #
+      - name: "[Cache] Prepare daily timestamp for cache"
+        if: runner.os != 'Windows'
+        id: date
+        run: echo "::set-output name=date::$(date '+%d-%m')"
+
+      - name: "[Cache] Prepare weekly timestamp for cache"
+        if: runner.os != 'Windows'
+        id: datew
+        run: echo "::set-output name=datew::$(date '+%Y-%V')"
+
       # - name: "[Cache] Cache R packages"
       #   if: runner.os != 'Windows'
       #   uses: pat-s/always-upload-cache@v1.1.4
@@ -119,22 +119,21 @@ jobs:
       #     path: ${{ env.R_LIBS_USER }}
       #     key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
       #     restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-      #
-      # - name: "[Cache] Cache ccache"
-      #   if: runner.os != 'Windows'
-      #   uses: pat-s/always-upload-cache@v1.1.4
-      #   with:
-      #     path: ${{ env.CCACHE_DIR}}
-      #     key: ${{ runner.os }}-r-${{ matrix.config.r }}-ccache-${{steps.datew.outputs.datew}}
-      #     restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-ccache-${{steps.datew.outputs.datew}}
-      #
-      # # install ccache and write config file
-      # # install libicu-dev libxml2-dev for ubuntu-16.04
-      # - name: "[Linux] ccache"
-      #   if: runner.os == 'Linux'
-      #   run: |
-      #     sudo apt install ccache libcurl4-openssl-dev libxml2-dev
-      #     mkdir -p ~/.R && echo -e 'CC=ccache gcc -std=gnu99\nCXX=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
+
+      - name: "[Cache] Cache ccache"
+        if: runner.os != 'Windows'
+        uses: pat-s/always-upload-cache@v1.1.4
+        with:
+          path: ${{ env.CCACHE_DIR}}
+          key: ${{ runner.os }}-r-${{ matrix.config.r }}-ccache-${{steps.datew.outputs.datew}}
+          restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-ccache-${{steps.datew.outputs.datew}}
+
+      # install ccache and write config file
+      - name: "[Linux] ccache"
+        if: runner.os == 'Linux'
+        run: |
+          sudo apt install ccache libcurl4-openssl-dev libxml2-dev
+          mkdir -p ~/.R && echo -e 'CC=ccache gcc -std=gnu99\nCXX=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
 
       # install ccache and write config file
       # mirror the setup described in https://github.com/rmacoslib/r-macos-rtools
