### 2020-07-21 #335 @maelle
I think so. What will you use the DOI for in the near future btw? codemetar doesn't use those yet.

Don't list _me_ as an author, as [written in the dev guide](https://devguide.ropensci.org/building.html#authorship), _"Please do not list editors as contributors. Your participation in and contribution to rOpenSci is thanks enough!"_ :smile_cat:
